### 2019-07-04 #17 @whedon
[PRE REVIEW]: Differential geometric algebra using Leibniz, Grassmann
**Submitting author:** @chakravala (<a href="http://orcid.org/0000-0002-8006-7816">Michael Reed</a>)
**Repository:** <a href="https://github.com/chakravala/Grassmann.jl" target ="_blank">https://github.com/chakravala/Grassmann.jl</a>

Thanks for submitting your paper to JuliaCon @chakravala. The JuliaCon editor @matbesancon, will work with you on this issue to find a reviewer for your submission before creating the main review issue.

@chakravala if you have any suggestions for potential reviewers then please mention them here in this thread. In addition, [this list of people](http://bit.ly/2V9GSFq) have already agreed to review for JuliaCon and may be suitable for this submission.

The JuliaCon submission bot @whedon is here to help you find and assign reviewers and start the main review. To find out what @whedon can do for you type:
