### 2019-07-04 #17 @chakravala
@eschnett would you be interested in taking a look at this submission?

you might need to use `lualatex` to compile it and also some extra packages

I have already done some Travis CI compile testing using `lualatex` with this configuration:
https://github.com/chakravala/Math-Research-Notes/blob/master/.travis.yml
