### 2019-07-04 #17 @eschnett
@matbesancon I can review the paper. I'm using the `Grassmann.jl` package myself, but there is no conflict of interest.
