### 2019-07-04 #17 @snoeyink
Sure.  Tell me how to go about it, as I have not reviewed for JuliaCon before.
(Though I have for many conferences and journals.)

Jack Snoeyink@cs.unc.edu
Computer Science, UNC Chapel Hill

From: Mathieu Besançon
Sent: Thursday, July 4, 2019 10:52 AM
To: JuliaCon/proceedings-review
Cc: Jack Snoeyink; Mention
Subject: Re: [JuliaCon/proceedings-review] [PRE REVIEW]: Differentialgeometric algebra using Leibniz, Grassmann (#17)

@snoeyink would you be interested in reviewing this proceeding?
