### 2019-07-04 #17 @matbesancon
great thanks. This issue will be closed with a new one starting for the review. You can discuss and request modifications to the paper and package in the issue or post issues in Grassmann.jl directly.
