### 2019-07-04 #18 @whedon
[REVIEW]: Differential geometric algebra using Leibniz, Grassmann
**Submitting author:** @chakravala (<a href="http://orcid.org/0000-0002-8006-7816">Michael Reed</a>)
**Repository:** <a href="https://github.com/chakravala/Grassmann.jl" target ="_blank">https://github.com/chakravala/Grassmann.jl</a>
