### 2019-07-04 #18 @matbesancon
@chakravala there are still issues in the paper folder
