### 2019-07-04 #18 @eschnett
@chakravala I used `lualatex` and `bibtex` to format the paper. There are problems with the references:
- `bibtex` outputs an error message about commas in the author list
- Grassmann didn't publish in 2000
- several references are incomplete, e.g. urls might be missing
- the name "de Rham" appears in lower case; this is probably bibtex's fault that quite eagerly insists that titles should be all lower case
