### 2019-07-04 #18 @eschnett
@chakravala The manuscript itself is very interesting, although I haven't read it in detail yet. However, I notice that it is quite technical and doesn't fit with the target audience at JuliaCon. At least, an introduction to geometric algebra is missing, and a description why Julia is a good implementation language. It seems to me that mentioning the geometric aspects (rotations, reflections) and the connection to computer visualization in the introduction might go well with the Julia crowd.

I also notice that the paper is quite terse. For the record: neither the reviewer nor the author instructions contain a page limit.
