### 2019-07-04 #18 @chakravala
> there are still issues in the paper folder

> bibtex outputs an error message about commas in the author list

I will have a look at resolving that

> Grassmann didn't publish in 2000

The translation from the AMS was published in 2000

> neither the reviewer nor the author instructions contain a page limit.

I thought there was a limit for 6 pages, so I had to leave out details I would otherwise want to include, can you confirm that the paper can really be longer than 6 pages?

Is it possible to make the paper 1-column, or should I convert to 2-columns?
