### 2019-07-05 #18 @snoeyink
Summary of Comments on grassmann-juliacon-2019.pdf (I'll upload a commented pdf.)
Page: 1

This paper is a description of Grassmann.jl and its supporting packages, which are appear quite impressive and extensive. Unfortunately, it
assumes extensive knowledge of terminology and notation from the reader, without ever defining what that reader should know, whose
notations they should be familiar with, and what that reader can expect to gain from reading.

As someone who has written a lot of geometric and topological code in low dimensions, and been interested in Clifford algebra enough
to read the book of Dorst, Fontijne, & Mann and some of Hestenes, I was eager to review this. I was disappointed because it does not
clearly say what the package does, nor what decisions were made between alternative implementations. (There are many intriguing hints,
e.g., Due to the design of the VectorBundle type
system, these operations make code optimization easier at compile-time by evaluating the bit parameters.)

As Fred Brooks has said, "Since anyone can create something new [in a synthetic field], that alone does not establish a contribution.
Rather, one must show that the creation is better." (Quoted from https://cra.org/resources/best-practice-memos/evaluating-computerscientists-
and-engineers-for-promotion-and-tenure/) A paper should do more than quote from the documentation of a package; it
should provide context so the reader (from the author's desired audience) can understand the decisions made by the package author and
why they are better than alternatives.

I feel that I could learn from this package, but not the document, as the author is working entirely from his own context, and does not
attempt to reach a broader audience. Unfortunately, the GPL3 license also discourages me from exploring the package source because I can't
guarantee that I can GPL3 any results that I mey eventually derive.

There are many easy things that can be done to broaden the audience for the paper.

1. Ask yourself, "Who is the intended audience? What is their expected background knowledge and what can they expect to learn from reading
this and can expect to get from the package?" Then answer these questions in the abstract or introduction.

2. Focus on verbs.
Keep action in the present tense (e.g., defn 4 & after).
For package descriptions or section headings, a verb would help parse: (
DirectSum.jl: provides types and operations for abstract tangent bundle vector space (unions, intersections, sums, etc.)
1 Parametric type polymorphism supports DirectSum )
If half the verbs in a paragraph are passive or being verbs (is, are, was), consider rewriting.

3. Strengthen parallels: use consistent terms (factorize = partition?
Many places compare and contrast the Leibniz and Grassmann algebras,, the mathematics and its implementation, the type hierarchy and
executable code. Write about parallel concepts using the same wording and sentence structure so that the similarities and differences pop out.

4. Define before using. Defn 1 defines its notation (except for n) after using it, forcing the reader to move back and forth. (In fact, it is unclear if
defn 1 is defining a type or a data representation.) Define vector space V\in Vect_K with metric g:VxV-->K,
Defn 5 defines almost none of the notation it uses.

5. Context first. (See Gopen & Swan https://cseweb.ucsd.edu/~swanson/papers/science-of-writing.pdf, or Gopen's book "The Sense of Structure")
Gopen says that a reader of English expects context first, new information after, in a sentence, paragraph, or section. By providing the
information in this order you conserve the reader energy for understanding the idea

How should section heading 1 be parsed?

Page: 3

is denoted by
[Stay in present tense, and improve parallelism.]

This is defining two different notations for symmetric difference.

This paragraph on implementation of the mathematics defined above should more clearly parallel the mathematics and should indicate
what it means to "work with" a TensorAlgebra{V}. Rewrite context first, and focus on verbs. You also have the parallel between
TensorAlgebra{V} and Grassman.Algebra{V}: structure the discussions the same, so that the differences in the concepts are not masked by
differences in expression.

Here is what I might say to restructure this paragraph in implementation, starting with motivation:

The bases for Grassmann and Leibniz algebras are large, since they include all sets or multiset of basis elements. Dorst, Fontijne, & Mann
suggest that the efficient implementation of geometric algebra is a compiler problem, to be solved by caching the product computations
actually needed in an application. Julia's parametric types and multiple dispatch are well suited for this. TensorAlgebra{V} is designed so
a user can smoothly transition between very high dimensional and low dimensional algebras in a single session, with varying levels of caching
and optimization. Printing with full alpha-numeric labeling with lower-case and capital letters limits a Basis to $n\le 62$ indices. Memory limits a
full MultiVector to n ≤ 22, although sparse operations support higher dimensions. A Grassmann.Algebra{V} is a container for the TensorAlgebra
generated by V, a VectorBundle{n}. Grassmann.Algebra{V} supports full caching for n ≤ 8, Grassmann.SparseAlgebra supports dimensions n ≤ 22, and
Grassmann.ExtendedAlgebra supports higher dimensions.

Consider rewriting the first paragraph of section 3 "context first":
I think this paragraph is attempting to say that both Grassmann's exterior algebra and Leibnitz's algebra of symmetric differential functions are
both captured by the geometric product. (Some of it is a repeat of the concepts below Defn 4.)
----original, with comment----
While oriented sets of the Grassmann exterior algebra are simpler,
[simpler than what? The reader has no context yet. Starting with "while" (and, below, "when") misleads your reader into thinking that time is
important to the context: while --> although.]
the parity [1] of (−1)Π is factored into
[The reader has to parse [1] as a citation and (-1)Π as notation.]
transposition compositions when interchanging the ordering of the tensor product argument permutation.
[Context is at the end of this sentence, and this is really hard to parse -- interchanging applies to two things, argument is singular, permutation is
collective.
Might one say, "Transposing terms in a tensor product changes the parity in (-1)Π."]
Alternatively, symmetric algebra does not need to track the parity
[By using "track the parity" to describe the process of the previous sentence, you force the reader to reinterpret what they just read.]
but does have multiplicity in its indices. [First use of indices -- would be a useful to give context for transpositions for tensor product...]
Symmetric differential function algebra of Leibniz trivializes the orientation property of index multi-sets, ["trivializes the orientation property" =
"makes tracking orientation trivial?" ]
while Grassmann’s exterior algebra is partitioned
into two oriented equivalence classes by anti-symmetry. The full tensor algebra [is this the symmetric differential function algebra?]
can be [sub-partitioned --> partitioned -- is this the same as the "can be factored" above?] into equivalence classes in multiple ways based on the composite symmetry, grade, and metric signature properties.
Both of the symmetry classes [This is the first reference to symmetry classes. Does this refer to both equivalence classes? ]
can be characterized by
the same geometric product, which is written as multiplication but explicitly denoted by ⊖ for clarity here.

---Context first rewrite--- (see prev comment)
Grassmann's exterior algebra and Leibnitz's algebra of symmetric differential functions track orientation differently.
Exterior algebra partitions space into two classes equivalence classes by orientation; transposing terms in a tensor product changes the parity in (-1)Π.

Symmetric algebra includes terms with multiplicity. Rather than tracking parity, it has multiple options to partition into equivalence classes based
on composite symmetry, grade, and metric signature properties.
Both algebras can be expressed by the same geometric product, which is usually written implicitly as multiplication but here explicitly denoted by ⊖.

-- I realize that this is probably not what was intended, as it does not capture the anti-symmetric and symmetric index sets, but to capture that the reader needs more context to be introduced.

There is no clarity when you define \ominus in terms of undefined \omega_X and \eta_Y.

This expression is quite useless, since there are so many undefined terms: \PI and g are mentioned before, with \oplus and \otimes, but
\omega and \eta never appear on the right; w, i, \epsilon, \otimes \mu do, but their relationship to \omega and \eta is not defined. I find
it wrong to use \ominus for set symmetric difference in a definition of \ominus for geometric product.

Page 6
GPL 3 does inhibit extension and learning from this package in the minds of some.

The bibliographic references are incomplete. Include publisher and DOIs so that people can find what you are citing. If you, for example,
cite the Dover reprinting of Emil Artin's 1957 Geometric Algebra, include also the original publication date. http://www.cws.illinois.edu/workshop/writers/citation/apa/reprintedbook/

Put braces [ ] around to preserve capitalization of names in bibtex.
