### 2019-07-05 chakravala/Grassmann.jl#29 @snoeyink
Review comments for JuliaCon2019
Please see comments in the pdf of your JuliaCon submission.
[grassmann-juliacon-2019.pdf](https://github.com/chakravala/Grassmann.jl/files/3363482/grassmann-juliacon-2019.pdf)
[Summary of Comments on grassmann-juliacon-2019.pdf](https://github.com/chakravala/Grassmann.jl/files/3363483/Summary.of.Comments.on.grassmann-juliacon-2019.pdf)

This paper is a description of Grassmann.jl and its supporting packages, which are appear quite impressive and extensive. Unfortunately, it
assumes extensive knowledge of terminology and notation from the reader, without ever defining what that reader should know, whose
notations they should be familiar with, and what that reader can expect to gain from reading.

As someone who has written a lot of geometric and topological code in low dimensions, and been interested in Clifford algebra enough
to read the book of Dorst, Fontijne, & Mann and some of Hestenes, I was eager to review this. I was disappointed because it does not
clearly say what the package does, nor what decisions were made between alternative implementations. (There are many intriguing hints,
e.g., Due to the design of the VectorBundle type
system, these operations make code optimization easier at compile-time by evaluating the bit parameters.)

As Fred Brooks has said, "Since anyone can create something new [in a synthetic field], that alone does not establish a contribution.
Rather, one must show that the creation is better." (Quoted from https://cra.org/resources/best-practice-memos/evaluating-computerscientists-and-engineers-for-promotion-and-tenure/) A paper should do more than quote from the documentation of a package; it
should provide context so the reader (from the author's desired audience) can understand the decisions made by the package author and
why they are better than alternatives.
