### 2019-07-05 #18 @snoeyink
I've created an issue and uploaded an annotated pdf in: https://github.com/chakravala/Grassmann.jl/issues/29

This paper is a description of Grassmann.jl and its supporting packages, which are appear quite impressive and extensive. Unfortunately, it
assumes extensive knowledge of terminology and notation from the reader, without ever defining what that reader should know, whose
notations they should be familiar with, and what that reader can expect to gain from reading.
