### 2019-07-05 chakravala/Grassmann.jl#29 @chakravala
Thanks for your helpful comments, since I am not affiliated with a university, it is only through comments like yours that I will be able to improve and refine the presentation of the information.

From the submission website, I was under the impression that the paper has to be limited to 6 pages, so I had to leave out information that I would otherwise like to explain.. which is why I also want to write a much more detailed paper based on this one.

At the moment, I'm not sure whether the short version of the paper can really address the concerns you are having. To address those concerns, a much longer paper would have to be written, which would take more time and be more than 6 pages.

Certainly, I would like to address all those concerns; however, with the limited publication space for JuliaCon, it is difficult to cover. You will not be able to find some of the notation in any references, since some of it is original or modified to suit my own needs.. although it is all based in existing concepts and ideas (except for the underlying implementation details, which provide a great deal of new constructions and techniques, but are pointless to explain without the necessary background knowledge). This is why those concerns can't really be addressed in a short version.
