### 2019-07-05 #18 @vchuravy
> can you confirm that the paper can really be longer than 6 pages?

Yes a paper can be longer than 6 pages, we do not have an explicit page-limit but we expect most papers to be 6-8 pages, longer if necessary.
