### 2019-07-08 chakravala/Grassmann.jl#29 @snoeyink
Yes, I know that time is short, and you thought the number of pages was limited.  Since it is not, please do add some words to characterize the intended audience and tell them what they can expect to learn.

I noticed you were in North Carolina from your GitHub, and I'd be happy to talk to you about your experiences with Julia and coding for geometric algebra.  I'd been traveling in the last two weeks, but am around Chapel Hill for the rest of the summer.  If you want to give a practice run of your JuliaCon talk, I could arrange a small audience here in UNC Chapel Hill, Computer Science.
