### 2019-07-13 chakravala/Grassmann.jl#29 @chakravala
Alright then, I could come by next week sometime, perhaps the math department would be interested in it too? It's only a 10 minute presentation, but I can answer some questions.
