### 2019-07-14 106e4c4
implemented sums of SparseChain and MultiGrade types
