+++ src/algebra.jl
@@ -285,7 +285,7 @@ end
 
 export ⊙, ⊠
 
-⊙(x::TensorAlgebra...) = (K=length(x); sum([prod(x[k]) for k ∈ collect(permutations(1:K))])/factorial(K))
+⊙(x::TensorAlgebra...) = (K=length(x); sum([prod(x[k]) for k ∈ permutations(1:K)])/factorial(K))
 
 function ⊠(x::TensorAlgebra...)
     K,V,out = length(x),∪(vectorspace.(x)...),prod(x)
@@ -510,7 +510,7 @@ function generate_product_algebra(Field=Field,VEC=:mvec,MUL=:*,ADD=:+,SUB=:-,CON
                     for k ∈ 1:binomial(N,G)
                         @inbounds val = b.v[k]
                         if val≠0
-                            p = $p(V,ib[k])
+                            @inbounds p = $p(V,ib[k])
                             v = typeof(V)<:Signature ? (p ? $SUB(val) : val) : p*val
                             @inbounds setblade!(out,v,complement(N,ib[k],D),Dimension{N}())
                         end
@@ -828,6 +828,68 @@ function generate_product_algebra(Field=Field,VEC=:mvec,MUL=:*,ADD=:+,SUB=:-,CON
                 $(add_val(eop,:out,:(value(b,$VEC(N,t))),bop))
                 return MultiVector{t,V}(out)
             end
+            function $op(a::SparseChain{V,G},b::MultiVector{T,V}) where {T<:$Field,V,G}
+                $(insert_expr((:N,),VEC)...)
+                at = terms(a)
+                t = promote_type(T,valuetype.(at)...)
+                out = $(bcast(bop,:(copy(value(b,$VEC(N,t))),)))
+                for A ∈ at
+                    addmulti!(out,value(A,t),bits(A),Dimension{N}())
+                end
+                return MultiVector{t,V}(out)
+            end
+            function $op(a::MultiVector{T,V},b::SparseChain{V,G}) where {T<:$Field,V,G}
+                $(insert_expr((:N,),VEC)...)
+                bt = terms(b)
+                t = promote_type(T,valuetype.(bt)...)
+                out = copy(value(a,$VEC(N,t)))
+                for B ∈ bt
+                    addmulti!(out,$bop(value(B,t)),bits(B),Dimension{N}())
+                end
+                return MultiVector{t,V}(out)
+            end
+            function $op(a::MultiGrade{V,G},b::MultiVector{T,V}) where {T<:$Field,V,G}
+                $(insert_expr((:N,),VEC)...)
+                at = terms(a)
+                t = promote_type(T,valuetype.(at)...)
+                out = $(bcast(bop,:(copy(value(b,$VEC(N,t))),)))
+                for A ∈ at
+                    TA = typeof(A)
+                    if TA <: TensorTerm
+                        addmulti!(out,value(A,t),bits(A),Dimension{N}())
+                    elseif TA <: SparseChain
+                        for α ∈ terms(A)
+                            addmulti!(out,value(α,t),bits(α),Dimension{N}())
+                        end
+                    elseif TA <: TensorMixed
+                        g = grade(A)
+                        r = binomsum(N,g)
+                        @inbounds out[r+1:r+binomial(N,g)] += value(A,$VEC(N,g,t))
+                    end
+                end
+                return MultiVector{t,V}(out)
+            end
+            function $op(a::MultiVector{T,V},b::MultiGrade{V,G}) where {T<:$Field,V,G}
+                $(insert_expr((:N,),VEC)...)
+                bt = terms(b)
+                t = promote_type(T,valuetype.(bt)...)
+                out = copy(value(a,$VEC(N,t)))
+                for B ∈ bt
[…]
