### 2019-07-18 chakravala/Grassmann.jl#29 @snoeyink
I’m sorry, but I had the shingles vaccine knock me back  for a couple of days this week.  I hope you enjoy JuliaCon.

Jack Snoeyink@cs.unc.edu
Computer Science, UNC Chapel Hill

From: Dream Scatter
Sent: Friday, July 12, 2019 8:57 PM
To: chakravala/Grassmann.jl
Cc: Jack Snoeyink; Author
Subject: Re: [chakravala/Grassmann.jl] Review comments for JuliaCon2019 (#29)

Alright then, I could come by next week sometime, perhaps the math department would be interested in it too? It's only a 10 minute presentation, but I can answer some questions.
