### 2019-07-25 27a59f4
created JuliaCon 2019 and tidelift links
