+++ paper/paper.tex
@@ -39,6 +39,7 @@
 
 \author{Michael Reed (Crucial Flow Research)}
 \title{Differential geometric algebra using Leibniz, Grassmann}
+\date{July 25th, 2019}
 
 \begin{document}
 
@@ -202,7 +203,7 @@ These products satisfy various \verb`MultiVector` properties, including the asso
 	\begin{align*}
 		(\bigotimes_k \partial_{\sigma(k)})\omega &\asymp (\partial_p \circ \dots\circ  \partial_1)\omega \iff \bigominus_k\partial_k = \bigodot_k\partial_k, & \partial_k(\omega\eta) &= \partial_k(\omega)\eta + \omega\partial_k(\eta).
 	\end{align*}
-\end{definition}	
+\end{definition}
 
 %\subsubsection{Interoperability for TensorAlgebra}
 
@@ -369,7 +370,7 @@ Mixed-symmetry algebra with \textit{Leibniz.jl} and \textit{Grassmann.jl}, havin
 In conclusion, the Dirac-Clifford product yields generalized Hodge-Laplacian and $b_p=\dim\mc H^pM$ are the Betti numbers with Euler characteristic $\chi = \sum_p (-1)^pb_p$.
 There will be a more detailed follow-up paper.
 
-\textbf{Acknowledgements}: Many thanks for discussions with Steven De Keninck, Utensil Song, Alexander Arsenovic, Hugo Hadfield, Karl Wessel; the helpful Julia community (e.g. Jameson Nash, Stefan Karpinski, Jeff Bezanson, Kristoffer Carlsson, Tamas Papp, Nathan Smith, etc); and thanks for support from my family.
+\textbf{Acknowledgements}: Many thanks for discussions with Steven De Keninck, Utensil Song, Alexander Arsenovic, Hugo Hadfield, Karl Wessel; the helpful Julia community (e.g. Jameson Nash, Stefan Karpinski, Jeff Bezanson, Kristoffer Carlsson, Petr Krysl, Nathan Smith, etc); and thanks for support from my family.
 
 %\cite{bezanson2017julia}
 
