+++ paper/ref.bib
@@ -63,12 +63,12 @@
 @book{artin,
 	author = {Emil Artin},
 	title = {Geometric Algebra},
-	year = {2012},
+	year = {1957},
 }
 
 @book{shahshahani,
 	author = {Siavash Shahshahani},
-	title = {An Introductory Course on Differntiable Manifolds},
+	title = {An Introductory Course on Differentiable Manifolds},
 	year = {2016},
 }
 
