+++ .github/FUNDING.yml
@@ -1,9 +1,9 @@
 # These are supported funding model platforms
 
 github: [chakravala]
-patreon: # Replace with a single Patreon username
+patreon: dreamscatter
 open_collective: # Replace with a single Open Collective username
 ko_fi: # Replace with a single Ko-fi username
-tidelift: # Replace with a single Tidelift platform-name/package-name e.g., npm/babel
+tidelift: julia/Grassmann
 community_bridge: # Replace with a single Community Bridge project-name e.g., cloud-foundry
 custom: https://liberapay.com/chakravala
