### 2019-07-26 68d6398
added compatibility for QR decomposition
