### 2019-08-20 #18 @matbesancon
ping @chakravala, could you make the modifications following the reviewers' comments?
