### 2019-08-25 22b1330
initialized doc strings for most products
