+++ src/algebra.jl
@@ -121,6 +121,18 @@ for side ∈ (:left,:right)
     end
 end
 
+@doc """
+    complementright(ω::TensorAlgebra)
+
+Grassmann-Poincare-Hodge complement: ⋆ω = ω∗I
+""" complementright
+
+@doc """
+    complementleft(ω::TensorAlgebra)
+
+Grassmann-Poincare left complement: ⋆'ω = I∗'ω
+""" complementleft
+
 ## reverse
 
 import Base: reverse, conj, ~
@@ -134,14 +146,41 @@ for r ∈ (:reverse,:involute,:conj)
     end
 end
 
+"""
+    ~(ω::TensorAlgebra)
+
+Reverse of a `MultiVector` element: ~ω = (-1)^(grade(ω)*(grade(ω)-1)/2)*ω
+"""
 reverse(a::UniformScaling{Bool}) = UniformScaling(!a.λ)
 reverse(a::UniformScaling{T}) where T<:Field = UniformScaling(-a.λ)
 
+"""
+    reverse(ω::TensorAlgebra)
+
+Reverse of a `MultiVector` element: ~ω = (-1)^(grade(ω)*(grade(ω)-1)/2)*ω
+"""
 @inline ~(b::TensorAlgebra) = reverse(b)
 @inline ~(b::UniformScaling) = reverse(b)
 
+@doc """
+    involute(ω::TensorAlgebra)
+
+Involute of a `MultiVector` element: ~ω = (-1)^grade(ω)*ω
+""" involute
+
+@doc """
+    conj(ω::TensorAlgebra)
+
+Clifford conjugate of a `MultiVector` element: conj(ω) = involute(~ω)
+""" conj
+
 ## geometric product
 
+"""
+    *(a::TensorAlgebra,b::TensorAlgebra)
+
+Geometric algebraic product: a*b = (-1)ᵖdet(a∩b)⊗(Λ(a⊖b)∪L(a⊕b))
+"""
 @pure function *(a::Basis{V},b::Basis{V}) where V
     A,B = bits(a), bits(b)
     diffcheck(V,A,B) && (return g_zero(V))
@@ -177,6 +216,12 @@ end
 
 export ∗, ⊛, ⊖
 const ⊖ = *
+
+"""
+    ∗(a::TensorAlgebra,b::TensorAlgebra)
+
+Reversed geometric product: a∗b = (~a)*b
+"""
 @inline ∗(a::A,b::B) where {A<:TensorAlgebra{V},B<:TensorAlgebra{V}} where V = (~a)*b
 
 ## exterior product
@@ -201,6 +246,12 @@ end
 #∧(a::MultiGrade{V},b::Basis{V}) where V = MultiGrade{V}(a.v,basis(a)*b)
 #∧(a::Basis{V},b::MultiGrade{V}) where V = MultiGrade{V}(b.v,a*basis(b))
 #∧(a::MultiGrade{V},b::MultiGrade{V}) where V = MultiGrade{V}(a.v*b.v,basis(a)*basis(b))
+
+"""
+    ∧(ω::TensorAlgebra,η::TensorAlgebra)
+
+Exterior product as defined by the anti-symmetric quotient Λ≡⊗/~
+"""
 @inline ∧(a::X,b::Y) where {X<:TensorAlgebra,Y<:TensorAlgebra} = interop(∧,a,b)
 @inline ∧(a::TensorAlgebra{V},b::UniformScaling{T}) where {V,T<:Field} = a∧V(b)
[…]
