### 2019-09-02 #18 @chakravala
is it too late to finish submitting this paper?

the main problem is, with the 2 columns I can't fit my big equations into the page.. is there a way to split an equation across both columns?
