### 2019-09-03 #18 @matbesancon
no it's fine but it would be nice to finish up soon.

> the main problem is, with the 2 columns I can't fit my big equations into the page.. is there a way to split an equation across both columns?

Not sure it's possible, I'll try to find something, in the mean time it would be best to split your equations
