### 2019-09-06 #18 @chakravala
unfortunately, the `juliacon` document class does not seem to function properly. in the example document it is claimed that there are `definition, example, remark` environments, however when I try to compile my document they cause compilation errors. is this going to be fixed in the `juliacon` class?
