### 2019-09-07 chakravala/Grassmann.jl#31 @chakravala
Differential Manifold and SubManifold generalization
This issue is to track and discuss the new updates related to differential SubManifold algebra.

https://github.com/chakravala/DirectSum.jl/commit/e76bea707f6814f128c91ec64c03ea2c936b2fb5
https://github.com/chakravala/DirectSum.jl/commit/1009e8033a21be47c8bd88209303fa36e61dc937
https://github.com/chakravala/DirectSum.jl/commit/60c0e2707e576cf9cdd51e09512a616bf6498422
https://github.com/chakravala/AbstractTensors.jl/commit/b7d2ff6d37cc3d5d64b13a7702b55c3038c6ab86

This is related to the finalization of the JuliaCon submission from https://github.com/JuliaCon/proceedings-review/issues/18
