### 2019-09-07 chakravala/Grassmann.jl#31 @chakravala
To move forward with this repository, it is necessary for the following to get merged:

https://github.com/JuliaRegistries/General/pull/3295
https://github.com/JuliaRegistries/General/pull/3349
https://github.com/JuliaRegistries/General/pull/3290
https://github.com/JuliaRegistries/General/pull/3289

once that happens, `Grassmann` will be able to be tested with differential `SubManifold`
