### 2019-09-07 10d9966
generalized differential Manifold and SubManifold algebra
