+++ src/multivectors.jl
@@ -8,11 +8,15 @@ abstract type GradedAlgebra{V,G} <: TensorAlgebra{V} end
 abstract type TensorTerm{V,G} <: GradedAlgebra{V,G} end
 abstract type TensorMixed{T,V} <: TensorAlgebra{V} end
 
+# number fields
+
+Fields = (Real,Complex)
+
 # symbolic print types
 
-parany = (Expr,Any)
-parsym = (Expr,Symbol)
-parval = (Expr,)
+parany = (Expr,Complex,Any)
+parsym = (Expr,Complex,Symbol)
+parval = (Expr,Complex)
 
 ## pseudoscalar
 
@@ -28,8 +32,8 @@ end
 
 @pure bits(b::Basis{V,G,B} where {V,G}) where B = B
 Base.one(b::Type{Basis{V}}) where V = getbasis(V,bits(b))
-Base.zero(V::VectorBundle) = 0*one(V)
-Base.one(V::VectorBundle) = Basis{V}()
+Base.zero(V::Manifold) = 0*one(V)
+Base.one(V::Manifold) = Basis{V}()
 
 function getindex(b::Basis,i::Int)
     d = one(Bits) << (i-1)
@@ -69,8 +73,12 @@ end
 ==(a::Basis{V,G} where V,b::Basis{W,L} where W) where {G,L} = false
 ==(a::Basis{V,G},b::Basis{W,G}) where {V,W,G} = interop(==,a,b)
 
-==(a::Number,b::TensorTerm{V,G} where V) where G = G==0 ? a==value(b) : 0==a==value(b)
-==(a::TensorTerm{V,G} where V,b::Number) where G = G==0 ? value(a)==b : 0==value(a)==b
+for T ∈ Fields
+    @eval begin
+        ==(a::T,b::TensorTerm{V,G} where V) where {T<:$T,G} = G==0 ? a==value(b) : 0==a==value(b)
+        ==(a::TensorTerm{V,G} where V,b::T) where {T<:$T,G} = G==0 ? value(a)==b : 0==value(a)==b
+    end
+end
 
 @inline show(io::IO, e::Basis) = DirectSum.printindices(io,vectorspace(e),bits(e))
 
@@ -97,7 +105,10 @@ for Blade ∈ MSB
         $Blade{V,G}(v,b::MBlade{V,G}) where {V,G} = $Blade{V,G,basis(b)}(v*b.v)
         $Blade{V,G,B}(v::T) where {V,G,B,T} = $Blade{V,G,B,T}(v)
         $Blade{V}(v::T) where {V,T} = $Blade{V,0,Basis{V}(),T}(v)
-        show(io::IO,m::$Blade) = print(io,(valuetype(m)∉parany ? [m.v] : ['(',m.v,')'])...,basis(m))
+        function show(io::IO,m::$Blade)
+            T = valuetype(m)
+            print(io,((!(T<:TensorMixed))&&T∉parany ? [m.v] : ['(',m.v,')'])...,basis(m))
+        end
     end
 end
 
@@ -147,10 +158,12 @@ for (Chain,vector,Blade) ∈ ((MSC[1],:MVector,MSB[1]),(MSC[2],:SVector,MSB[2]))
             end
             @inbounds DirectSum.printindices(io,V,ib[1])
             for k ∈ 2:length(ib)
-                if T == Any && typeof(m.v[k]) ∈ parsym
-                    @inbounds typeof(m.v[k])∉parval ? print(io," + ",m.v[k]) : print(io," + (",m.v[k],")")
+                @inbounds tmv = typeof(m.v[k])
+                if T == Any && (tmv ∈ parsym || tmv <: TensorAlgebra)
[…]
