### 2019-09-09 #18 @matbesancon
@chakravala indeed the names of the environments are incorrect, you can check them out in the cls file:

```
\newtheorem{theorem}{Theorem}
\newtheorem{strategy}{Strategy}
\newtheorem{property}{Property}
\newtheorem{proposition}{Proposition}
\newtheorem{lemma}[theorem]{Lemma}
\newtheorem{exam}{Example}
\newenvironment{example}{\italicenvfalse\begin{exam}}{\end{exam}\italicenvtrue}
\newtheorem{defi}[theorem]{Definition}
```
