### 2019-09-09 225bb79
improved differential mixed symmetry dispatch #31
