+++ src/algebra.jl
@@ -5,7 +5,7 @@
 import Base: +, -, *, ^, /, //, inv, <, >, <<, >>, >>>
 import AbstractLattices: ∧, ∨, dist
 import AbstractTensors: ⊗, ⊛, ⊙, ⊠, ⨼, ⨽, ⋆, ∗, rem, div, contraction
-import DirectSum: diffcheck, tangent, hasinforigin, hasorigininf
+import DirectSum: diffcheck, diffmode, tangent, hasinforigin, hasorigininf
 export tangent
 
 const Field = Real
@@ -50,7 +50,7 @@ function declare_mutating_operations(M,F,set_val,SUB,MUL)
                     if v ≠ 0 && !diffcheck(V,a,b)
                         A,B,Q,Z = symmetricmask(V,a,b)
                         val = (typeof(V)<:Signature || count_ones(A&B)==0) ? (parity(A,B,V) ? $SUB(v) : v) : $MUL(parityinner(A,B,V),v)
-                        diffvars(V)≠0 && !iszero(Z) && (val *= Basis{V}(Z))
+                        diffvars(V)≠0 && !iszero(Z) && (val *= Basis{V}(Z); count_ones(Q)+order(val)>diffmode(V) && (return m))
                         $s(m,val,(A⊻B)|Q,Dimension{N}())
                     end
                     return m
@@ -60,7 +60,7 @@ function declare_mutating_operations(M,F,set_val,SUB,MUL)
                         A,B,Q,Z = symmetricmask(V,a,b)
                         pcc,bas,cc = (hasinf(V) && hasorigin(V)) ? conformal(A,B,V) : (false,A⊻B,false)
                         val = (typeof(V)<:Signature || count_ones(A&B)==0) ? (parity(A,B,V)⊻pcc ? $SUB(v) : v) : $MUL(parityinner(A,B,V),pcc ? $SUB(v) : v)
-                        diffvars(V)≠0 && !iszero(Z) && (val *= Basis{V}(Z))
+                        diffvars(V)≠0 && !iszero(Z) && (val *= Basis{V}(Z); count_ones(Q)+order(val)>diffmode(V) && (return m))
                         $s(m,val,bas|Q,Dimension{N}())
                         cc && $s(m,hasinforigin(V,A,B) ? $SUB(val) : val,(conformalmask(V)⊻bas)|Q,Dimension{N}())
                     end
@@ -78,7 +78,11 @@ function declare_mutating_operations(M,F,set_val,SUB,MUL)
                         if val ≠ 0
                             g,C,t,Z = $uct(A,B,V)
                             v = val
-                            diffvars(V)≠0 && !iszero(Z) && (v *= Basis{V}(Z))
+                            if diffvars(V)≠0 && !iszero(Z)
+                                _,_,Q,_ = symmetricmask(V,A,B)
+                                v *= Basis{V}(Z)
+                                count_ones(Q)+order(v)>diffmode(V) && (return m)
+                            end
                             t && $s(m,typeof(V) <: Signature ? g ? $SUB(v) : v : $MUL(g,v),C,Dimension{N}())
                         end
                         return m
@@ -200,8 +204,14 @@ end
 
 for Blade ∈ MSB
     @eval begin
[…]
