### 2019-09-09 ccad321
further improved dispatch of differentials
