+++ src/algebra.jl
@@ -50,21 +50,29 @@ function declare_mutating_operations(M,F,set_val,SUB,MUL)
                     if v ≠ 0 && !diffcheck(V,a,b)
                         A,B,Q,Z = symmetricmask(V,a,b)
                         val = (typeof(V)<:Signature || count_ones(A&B)==0) ? (parity(A,B,V) ? $SUB(v) : v) : $MUL(parityinner(A,B,V),v)
-                        diffvars(V)≠0 && !iszero(Z) && (val *= Basis{V}(Z); count_ones(Q)+order(val)>diffmode(V) && (return m))
+                        if diffvars(V)≠0 && !iszero(Z)
+                            #T≠Any && (return true)
+                            val *= getbasis(V,Z)
+                            count_ones(Q)+order(val)>diffmode(V) && (return false)
+                        end
                         $s(m,val,(A⊻B)|Q,Dimension{N}())
                     end
-                    return m
+                    return false
                 end
                 @inline function $(Symbol(:geom,s))(V::W,m::$M,a::Bits,b::Bits,v::S) where W<:Manifold{N} where {N,T<:$F,S<:$F,M}
                     if v ≠ 0 && !diffcheck(V,a,b)
                         A,B,Q,Z = symmetricmask(V,a,b)
                         pcc,bas,cc = (hasinf(V) && hasorigin(V)) ? conformal(A,B,V) : (false,A⊻B,false)
                         val = (typeof(V)<:Signature || count_ones(A&B)==0) ? (parity(A,B,V)⊻pcc ? $SUB(v) : v) : $MUL(parityinner(A,B,V),pcc ? $SUB(v) : v)
-                        diffvars(V)≠0 && !iszero(Z) && (val *= Basis{V}(Z); count_ones(Q)+order(val)>diffmode(V) && (return m))
+                        if diffvars(V)≠0 && !iszero(Z)
+                            #T≠Any && (return true)
+                            val *= getbasis(V,Z)
+                            count_ones(Q)+order(val)>diffmode(V) && (return false)
+                        end
                         $s(m,val,bas|Q,Dimension{N}())
                         cc && $s(m,hasinforigin(V,A,B) ? $SUB(val) : val,(conformalmask(V)⊻bas)|Q,Dimension{N}())
                     end
-                    return m
+                    return false
                 end
             end
             for j ∈ (:join,:geom)
@@ -79,13 +87,14 @@ function declare_mutating_operations(M,F,set_val,SUB,MUL)
                             g,C,t,Z = $uct(A,B,V)
                             v = val
                             if diffvars(V)≠0 && !iszero(Z)
+                                #T≠Any && (return true)
                                 _,_,Q,_ = symmetricmask(V,A,B)
-                                v *= Basis{V}(Z)
-                                count_ones(Q)+order(v)>diffmode(V) && (return m)
+                                v *= getbasis(V,Z)
+                                count_ones(Q)+order(v)>diffmode(V) && (return false)
                             end
[…]
