+++ src/algebra.jl
@@ -51,7 +51,7 @@ function declare_mutating_operations(M,F,set_val,SUB,MUL)
                         A,B,Q,Z = symmetricmask(V,a,b)
                         val = (typeof(V)<:Signature || count_ones(A&B)==0) ? (parity(A,B,V) ? $SUB(v) : v) : $MUL(parityinner(A,B,V),v)
                         if diffvars(V)≠0 && !iszero(Z)
-                            #T≠Any && (return true)
+                            T≠Any && (return true)
                             val *= getbasis(V,Z)
                             count_ones(Q)+order(val)>diffmode(V) && (return false)
                         end
@@ -65,7 +65,7 @@ function declare_mutating_operations(M,F,set_val,SUB,MUL)
                         pcc,bas,cc = (hasinf(V) && hasorigin(V)) ? conformal(A,B,V) : (false,A⊻B,false)
                         val = (typeof(V)<:Signature || count_ones(A&B)==0) ? (parity(A,B,V)⊻pcc ? $SUB(v) : v) : $MUL(parityinner(A,B,V),pcc ? $SUB(v) : v)
                         if diffvars(V)≠0 && !iszero(Z)
-                            #T≠Any && (return true)
+                            T≠Any && (return true)
                             val *= getbasis(V,Z)
                             count_ones(Q)+order(val)>diffmode(V) && (return false)
                         end
@@ -87,7 +87,7 @@ function declare_mutating_operations(M,F,set_val,SUB,MUL)
                             g,C,t,Z = $uct(A,B,V)
                             v = val
                             if diffvars(V)≠0 && !iszero(Z)
-                                #T≠Any && (return true)
+                                T≠Any && (return true)
                                 _,_,Q,_ = symmetricmask(V,A,B)
                                 v *= getbasis(V,Z)
                                 count_ones(Q)+order(v)>diffmode(V) && (return false)
@@ -481,7 +481,7 @@ function generate_product_algebra(Field=Field,VEC=:mvec,MUL=:*,ADD=:+,SUB=:-,CON
         @inline function inneraddvalue!(mv::MBlade{V,0,B,T} where {W,B},α,β,γ::T) where {V,T<:$Field}
             if γ≠0
                 g,C,f,Z = interior(α,β,V)
-                #!iszero(C) && T≠Any && (return true)
+                !iszero(C) && T≠Any && (return true)
                 v = iszero(C) ? γ : γ*getbasis(V,C)
                 if diffvars(V)≠0 && !iszero(Z)
                     _,_,Q,_ = symmetricmask(V,α,β)
@@ -556,7 +556,7 @@ function generate_product_algebra(Field=Field,VEC=:mvec,MUL=:*,ADD=:+,SUB=:-,CON
                 $(insert_expr((:N,:t,:mv,:ib,:μ),VEC)...)
                 for i ∈ 1:binomial(N,G)
                     if @inbounds inneraddvalue!(mv,ib[i],bits(b),a[i])&μ
-                        $(insert_expr((:mv,);mv=:(value(mv))))
+                        $(insert_expr((:mv,);mv=:(value(mv)))...)
                         @inbounds inneraddvalue!(mv,ib[i],bits(b),a[i])
                     end
                 end
@@ -566,7 +566,7 @@ function generate_product_algebra(Field=Field,VEC=:mvec,MUL=:*,ADD=:+,SUB=:-,CON
                 $(insert_expr((:N,:t,:mv,:ib,:μ),VEC)...)
                 for i ∈ 1:binomial(N,G)
                     if @inbounds inneraddvalue!(mv,bits(a),ib[i],b[i])&μ
-                        $(insert_expr((:mv,);mv=:(value(mv))))
+                        $(insert_expr((:mv,);mv=:(value(mv)))...)
[…]
