### 2019-09-10 e0138ce
finalized mixed-symmetry algebra dispatch
