### 2019-09-11 2ab467b
refined MultiVector reverse and printing
