### 2019-09-11 a516cbb
extended inv for differential Chain elements
