+++ src/algebra.jl
@@ -157,7 +157,10 @@ Grassmann-Poincare left complement: ⋆'ω = I∗'ω
 import Base: reverse, conj, ~
 export involute
 
-@pure grade(V,B) = count_ones(B&(one(Bits)<<DirectSum.grade(V)-1))
+@pure grade_basis(V,B) = B&(one(Bits)<<DirectSum.grade(V)-1)
+@pure grade_basis(v,::Basis{V,G,B} where G) where {V,B} = grade_basis(V,B)
+@pure grade(V,B) = count_ones(grade_basis(V,B))
+@pure grade(v,::Basis{V,G,B} where G) where {V,B} = grade(V,B)
 
 for r ∈ (:reverse,:involute,:conj)
     p = Symbol(:parity,r)
@@ -274,7 +277,10 @@ function ∧(a::X,b::Y) where {X<:TensorTerm{V},Y<:TensorTerm{V}} where V
     A,B,Q,Z = symmetricmask(V,ba,bb)
     ((count_ones(A&B)>0) || diffcheck(V,ba,bb)) && (return g_zero(V))
     v = value(a)*value(b)
-    diffvars(V)≠0 && !iszero(Z) && (v = SBlade{V}(getbasis(V,Z),v); count_ones(Q)+order(v)>diffmode(V) && (return zero(V)))
+    if diffvars(V)≠0 && !iszero(Z)
+        v=typeof(v)<:TensorMixed ? SBlade{V}(getbasis(V,Z),v) : SBlade{V}(v,getbasis(V,Z))
+        count_ones(Q)+order(v)>diffmode(V) && (return zero(V))
+    end
     return SBlade{V}(parity(x,y) ? -v : v,getbasis(V,(A⊻B)|Q))
 end
 
@@ -309,7 +315,7 @@ function ∨(a::X,b::Y) where {X<:TensorTerm{V},Y<:TensorTerm{V}} where V
     v = value(a)*value(b)
     if diffvars(V)≠0 && !iszero(Z)
         _,_,Q,_ = symmetricmask(V,bits(basis(a)),bits(basis(b)))
-        v = SBlade{V}(getbasis(V,Z),v)
+        v=typeof(v)<:TensorMixed ? SBlade{V}(getbasis(V,Z),v) : SBlade{V}(v,getbasis(V,Z))
         count_ones(Q)+order(v)>diffmode(V) && (return zero(V))
     end
     return SBlade{V}(p ? -v : v,getbasis(V,C))
@@ -355,7 +361,7 @@ function contraction(a::X,b::Y) where {X<:TensorTerm{V},Y<:TensorTerm{V}} where 
     v = value(a)*value(b)
     if diffvars(V)≠0 && !iszero(Z)
         _,_,Q,_ = symmetricmask(V,bits(basis(a)),bits(basis(b)))
-        v = SBlade{V}(getbasis(V,Z),v)
+        v=typeof(v)<:TensorMixed ? SBlade{V}(getbasis(V,Z),v) : SBlade{V}(v,getbasis(V,Z))
         count_ones(Q)+order(v)>diffmode(V) && (return zero(V))
     end
     return SBlade{V}(typeof(V) <: Signature ? (g ? -v : v) : g*v,getbasis(V,C))
@@ -404,7 +410,7 @@ function cross(a::X,b::Y) where {X<:TensorTerm{V},Y<:TensorTerm{V}} where V
     v = value(a)*value(b)
     if diffvars(V)≠0 && !iszero(Z)
         _,_,Q,_ = symmetricmask(V,bits(basis(a)),bits(basis(b)))
-        v = SBlade{V}(getbasis(V,Z),v)
[…]
