### 2019-09-19 347fb90
transitioned JuliaCon paper and revised dispatch
