+++ paper/paper.tex
@@ -6,7 +6,7 @@
 %\usepackage[utf8]{inputenc}
 
 \usepackage{graphicx}
-\usepackage[left=1.00in, right=1.00in, top=1.00in, bottom=1.00in]{geometry}
+%\usepackage[left=1.00in, right=1.00in, top=1.00in, bottom=1.00in]{geometry}
 
 \input{mathsym}
 
@@ -22,7 +22,7 @@
 %\newtheorem*{remark}{Remark}
 %\newtheorem{example}{Example}
 %\newtheorem*{recall}{Recall}
-%\newtheorem*{note}{Note}
+%\newtheorem*{remark}{Note}
 %\newtheorem*{observe}{Observe}
 %\newtheorem*{question}{\underline{Question}}
 %\newtheorem*{fact}{Fact}
@@ -38,7 +38,7 @@
 \DeclareMathOperator*{\obackslash}{\text{\reflectbox{$\oslash$}}}
 
 \author{Michael Reed (Crucial Flow Research)}
-\title{Differential geometric algebra using Leibniz, Grassmann}
+\title{Differential geometric algebra with Leibniz and Grassmann}
 \date{July 25th, 2019}
 
 \begin{document}
@@ -52,157 +52,170 @@
 	package provides tools for computations based on multi-linear algebra and spin groups using the extended geometric algebra known as Grassmann-Clifford-Hestenes-Leibniz algebra.
 	The primary operations are
 	outer, regressive, inner, geometric, and cross products along with the Hodge star, adjoint, and multivector reversal operations. 
-	The kernelized operations are built on composite sparse tensor products and Hodge duality, with high dimensional support for up to 62 indices and staged caching / precompilation. Code generation enables making concise and extensible definitions.
+	The kernelized operations are built up from composite sparse tensor products and Hodge duality, with high dimensional support for up to 62 indices using staged caching and precompilation. Code generation enables concise and also extensible definitions.
 	The \textit{DirectSum.jl}
-	multivector parametric type polymorphism is based on various tangent vector spaces and conformal projective geometry.
-	Additionally, interoperability between different sub-algebras is enabled by \textit{AbstractTensors.jl},
+	multivector parametric type polymorphism is based on tangent vector spaces and conformal projective geometry.
+	Additionally, universal interoperability between different sub-algebras is enabled by \textit{AbstractTensors.jl},
 	on which the type system is built.
 	%\headingtable
 \end{abstract}
 
-In this paper, the mathematical foundations and some of the nuances in the definitions specific to the \textit{Grassmann.jl} implementation are concisely described.
-The \textit{Grassmann.jl} package and its accompanying support packages provide an extensible platform for computing with geometric algebra at high dimensions. 
-The design is based on the \verb`TensorAlgebra` abstract type system interoperability from \textit{AbstractTensors.jl} with a \verb`VectorBundle` type parameter from \textit{DirectSum.jl}. Abstract vector space type operations happen at compile-time, resulting in a differential conformal geometric algebra of hyper-dual multivector forms.
+Mathematical foundations and some of the nuances in the definitions specific to the \textit{Grassmann.jl} implementation are concisely described, along with the accompanying support packages that provide an extensible platform for computing with geometric algebra at high dimensions.
[…]
