### 2019-09-19 #18 @chakravala
@matbesancon this is what I needed to change to make it work https://github.com/chakravala/Grassmann.jl/commit/347fb907c0dd289d57c0c1ceb78a31a133be1402

```
\newtheorem{theorem}{Theorem}
\newtheorem{strategy}{Strategy}
\newtheorem{property}{Property}
\newtheorem{proposition}{Proposition}
\newtheorem{lemma}[theorem]{Lemma}
\newtheorem{exam}{Example}
\newenvironment{example}{\begin{exam}}{\end{exam}}
\newtheorem{defi}[theorem]{Definition}
\newenvironment{definition}{\begin{defi}}{\end{defi}}
\newtheorem{rem}{Remark}
\newenvironment{remark}{\begin{rem}}{\end{rem}}
```
the `italicenv` stuff didn't work at all

draft is updated, all equations now fit into 4 pages, will post another update soon
