### 2019-09-20 97eac43
improved arbitrary precision dispatch and SymEngine compat
