### 2019-09-21 eb06d08
implemented support for AbstractAlgebra and GaloisFields
