### 2019-09-22 284be3a
revised ∇ Derivation promotion and bug fix
