### 2019-09-23 f2113fb
implemented SimpleDiGraph constructor for LightGraphs
