### 2019-09-29 #18 @chakravala
Hi folks, sorry about being a bit late with my paper.

I have now implemented the finalizing feature for the differential geometric algebra, which will allow me to add more explicit and concrete examples with visualization into the paper.

Please be a little more patient, it should only take about 1 more week to wrap up this review, as it is only a matter of sorting out the some details for the examples.
