### 2019-09-29 #18 @matbesancon
> this is what I needed to change to make it work chakravala/Grassmann.jl@347fb90

I think you can introduce that in the paper and not in the cls file
