### 2019-09-29 #18 @chakravala
No, the cls file had to be edited because otherwise it causes a compilation error, specifically `italicenv` causes the compilation error and must be removed from the cls. I can move the extra definitions to the main paper, but the cls filed needs to have the `italicenv` part removed.
