### 2019-09-30 29bf98b
created derivation detection and application scheme
