+++ src/algebra.jl
@@ -5,7 +5,7 @@
 import Base: +, -, *, ^, /, //, inv, <, >, <<, >>, >>>
 import AbstractLattices: ∧, ∨, dist
 import AbstractTensors: ⊗, ⊛, ⊙, ⊠, ⨼, ⨽, ⋆, ∗, rem, div, contraction
-import DirectSum: diffcheck, diffmode, tangent, hasinforigin, hasorigininf
+import DirectSum: diffcheck, diffmode, tangent, hasinforigin, hasorigininf, symmetricsplit
 export tangent
 
 const Field = Real
@@ -26,7 +26,50 @@ const SymField = Any
 
 set_val(set,expr,val) = Expr(:(=),expr,set≠:(=) ? Expr(:call,:($Sym.:∑),expr,val) : val)
 
-function declare_mutating_operations(M,F,set_val,SUB,MUL)
+derive(n::N,b) where N<:Number = zero(typeof(n))
+derive(n,b,a,t) = t ? (a,derive(n,b)) : (derive(n,b),a)
+#derive(n,b,a::T,t) where T<:TensorAlgebra = t ? (a,derive(n,b)) : (derive(n,b),a)
+
+symmetricsplit(V::M,b::Basis) where M<:Manifold = symmetricsplit(V,bits(b))
+
+@inline function derive_mul(V,A,B,v,x::Bool)
+    if !(diffvars(V)≠0 && mixedmode(V)<0)
+        return v
+    else
+        sa,sb = symmetricsplit(V,A),symmetricsplit(V,B)
+        ca,cb = count_ones(sa[2]),count_ones(sb[2])
+        return if ca == cb == 0
+            v
+        elseif ca == 0
+            derive(x ? one(typeof(v)) : v,getbasis(V,sa[1]))
+        elseif cb == 0
+            derive(x ? v : one(typeof(v)),getbasis(V,sb[1]))
+        else
+            v
+        end
+    end
+end
+
+@inline function derive_mul(V,A,B,a,b,*)
+    if !(diffvars(V)≠0 && mixedmode(V)<0)
+        return a*b
+    else
+        sa,sb = symmetricsplit(V,A),symmetricsplit(V,B)
+        ca,cb = count_ones(sa[2]),count_ones(sb[2])
+        α,β = if ca == cb == 0
+            a,b
+        elseif ca == 0
+            derive(b,getbasis(V,sa[1]),a,true)
+        elseif cb == 0
+            derive(a,getbasis(V,sb[1]),b,false)
+        else
+            a,b
+        end
+        return α*β
+    end
+end
+
+function generate_mutators(M,F,set_val,SUB,MUL)
     for (op,set) ∈ ((:add,:(+=)),(:set,:(=)))
         sm = Symbol(op,:multi!)
         sb = Symbol(op,:blade!)
@@ -108,16 +151,6 @@ end
 @inline exteraddmulti!(V::W,out,α,β,γ) where W<:Manifold = (count_ones(α&β)==0) && joinaddmulti!(V,out,α,β,γ)
 
 @inline outeraddblade!(V::W,out,α,β,γ) where W<:Manifold = (count_ones(α&β)==0) && joinaddblade!(V,out,α,β,γ)
-
-@inline function add!(out::MultiVector{T,V},val::T,a::Int,b::Int) where {T,V}
-    A,B = Bits(a), Bits(b)
-    add!(out,val,Basis{V,count_ones(A),A},Basis{V,count_ones(B),B})
-end
-@inline function add!(m::MultiVector{T,V},v::T,a::Basis{V},b::Basis{V}) where {T<:Field,V}
-    A,B = bits(a), bits(b)
-    !diffcheck(V,A,B) && addmulti!(m.v,parity(A,B) ? -(v) : v,A.⊻B)
-    return out
-end
 
[…]
