### 2019-10-01 #18 @matbesancon
You should not change the cls overall, if the italicenv doesn't work there is an incompatible part in the paper
