+++ src/algebra.jl
@@ -164,7 +164,7 @@ for side ∈ (:left,:right)
     c = Symbol(:complement,side)
     p = Symbol(:parity,side)
     @eval @pure function $c(b::Basis{V,G,B}) where {V,G,B}
-        d = getbasis(V,complement(ndims(V),B,diffvars(V)))
+        d = getbasis(V,complement(ndims(V),B,diffvars(V),hasinf(V)+hasorigin(V)))
         mixedmode(V)<0 && throw(error("Complement for mixed tensors is undefined"))
         typeof(V)<:Signature ? ($p(b) ? SBlade{V}(-value(d),d) : d) : SBlade{V}($p(b)*value(d),d)
     end
@@ -759,7 +759,7 @@ function generate_products(Field=Field,VEC=:mvec,MUL=:*,ADD=:+,SUB=:-,CONJ=:conj
             @eval begin
                 function $c(b::$Chain{T,V,G}) where {T<:$Field,V,G}
                     mixedmode(V)<0 && throw(error("Complement for mixed tensors is undefined"))
-                    $(insert_expr((:N,:ib,:D),VEC)...)
+                    $(insert_expr((:N,:ib,:D,:P),VEC)...)
                     out = zeros($VEC(N,G,T))
                     D = diffvars(V)
                     for k ∈ 1:binomial(N,G)
@@ -767,7 +767,7 @@ function generate_products(Field=Field,VEC=:mvec,MUL=:*,ADD=:+,SUB=:-,CONJ=:conj
                         if val≠0
                             @inbounds p = $p(V,ib[k])
                             v = typeof(V)<:Signature ? (p ? $SUB(val) : val) : p*val
-                            @inbounds setblade!(out,v,complement(N,ib[k],D),Dimension{N}())
+                            @inbounds setblade!(out,v,complement(N,ib[k],D,P),Dimension{N}())
                         end
                     end
                     return $Chain{T,V,N-G}(out)
@@ -777,7 +777,7 @@ function generate_products(Field=Field,VEC=:mvec,MUL=:*,ADD=:+,SUB=:-,CONJ=:conj
         @eval begin
             function $c(m::MultiVector{T,V}) where {T<:$Field,V}
                 mixedmode(V)<0 && throw(error("Complement for mixed tensors is undefined"))
-                $(insert_expr((:N,:bs,:bn),VEC)...)
+                $(insert_expr((:N,:bs,:bn,:P),VEC)...)
                 out = zeros($VEC(N,T))
                 D = diffvars(V)
                 for g ∈ 1:N+1
@@ -786,7 +786,7 @@ function generate_products(Field=Field,VEC=:mvec,MUL=:*,ADD=:+,SUB=:-,CONJ=:conj
                         @inbounds val = m.v[bs[g]+i]
                         if val≠0
                             v = typeof(V)<:Signature ? ($p(V,ib[i]) ? $SUB(val) : val) : $p(V,ib[i])*val
-                            @inbounds setmulti!(out,v,complement(N,ib[i],D),Dimension{N}())
+                            @inbounds setmulti!(out,v,complement(N,ib[i],D,P),Dimension{N}())
                         end
                     end
                 end
