### 2019-10-04 bcde7d6
fixed #20 geometric product of negative null-basis
