### 2019-10-05 5a6e9e5
improved SizedArray and Number dispatch
