+++ src/algebra.jl
@@ -1102,7 +1102,7 @@ function generate_products(Field=Field,VEC=:mvec,MUL=:*,ADD=:+,SUB=:-,CONJ=:conj
                 end
                 function $op(a::$Blade{V,G,A,S} where A,b::MultiVector{T,V}) where {T<:$Field,V,G,S<:$Field}
                     $(insert_expr((:N,:t),VEC)...)
-                    out = $(bcast(bop,:(copy(value(b,$VEC(N,t))),)))
+                    out = convert($VEC(N,t),$(bcast(bop,:(copy(value(b,$VEC(N,t))),))))
                     addmulti!(out,value(a,t),bits(basis(a)),Dimension{N}())
                     return MultiVector{t,V}(out)
                 end
@@ -1118,7 +1118,7 @@ function generate_products(Field=Field,VEC=:mvec,MUL=:*,ADD=:+,SUB=:-,CONJ=:conj
             $op(a::MultiVector{T,V}) where {T<:$Field,V} = MultiVector{$TF,V}($(bcast(bop,:(value(a),))))
             function $op(a::Basis{V,G},b::MultiVector{T,V}) where {T<:$Field,V,G}
                 $(insert_expr((:N,:t),VEC)...)
-                out = $(bcast(bop,:(copy(value(b,$VEC(N,t))),)))
+                out = convert($VEC(N,t),$(bcast(bop,:(copy(value(b,$VEC(N,t))),))))
                 addmulti!(out,value(a,t),bits(basis(a)),Dimension{N}())
                 return MultiVector{t,V}(out)
             end
@@ -1138,7 +1138,7 @@ function generate_products(Field=Field,VEC=:mvec,MUL=:*,ADD=:+,SUB=:-,CONJ=:conj
                 $(insert_expr((:N,),VEC)...)
                 at = terms(a)
                 t = promote_type(T,valuetype.(at)...)
-                out = $(bcast(bop,:(copy(value(b,$VEC(N,t))),)))
+                out = convert($VEC(N,t),$(bcast(bop,:(copy(value(b,$VEC(N,t))),))))
                 for A ∈ at
                     addmulti!(out,value(A,t),bits(A),Dimension{N}())
                 end
@@ -1158,7 +1158,7 @@ function generate_products(Field=Field,VEC=:mvec,MUL=:*,ADD=:+,SUB=:-,CONJ=:conj
                 $(insert_expr((:N,),VEC)...)
                 at = terms(a)
                 t = promote_type(T,valuetype.(at)...)
-                out = $(bcast(bop,:(copy(value(b,$VEC(N,t))),)))
+                out = convert($VEC(N,t),$(bcast(bop,:(copy(value(b,$VEC(N,t))),))))
                 for A ∈ at
                     TA = typeof(A)
                     if TA <: TensorTerm
@@ -1224,7 +1224,7 @@ function generate_products(Field=Field,VEC=:mvec,MUL=:*,ADD=:+,SUB=:-,CONJ=:conj
                     end
                     function $op(a::$Blade{V,G,A,S} where A,b::$Chain{T,V,G}) where {T<:$Field,V,G,S<:$Field}
                         $(insert_expr((:N,:t),VEC)...)
-                        out = $(bcast(bop,:(copy(value(b,$VEC(N,G,t))),)))
+                        out = convert($VEC(N,G,t),$(bcast(bop,:(copy(value(b,$VEC(N,G,t))),))))
                         addblade!(out,value(a,t),basis(a),Dimension{N}())
                         return MChain{t,V,G}(out)
                     end
[…]
