### 2019-10-06 2826467
extended LinearAlgebra.UniformScaling for DiagonalForm
