### 2019-10-06 9968946
created generate_inverses methods for symbols
