+++ src/Grassmann.jl
@@ -433,7 +433,7 @@ function __init__()
     @require Reduce="93e0c654-6965-5f22-aba9-9c1ae6b3c259" begin
         *(a::Reduce.RExpr,b::Basis{V}) where V = SBlade{V}(a,b)
         *(a::Basis{V},b::Reduce.RExpr) where V = SBlade{V}(b,a)
-        *(a::Reduce.RExpr,b::MultiVector{T,V}) where {T,V} = MultiVector{promote_type(T,F),V}(broadcast(Reduce.Algebra.:*,a,b.v))
+        *(a::Reduce.RExpr,b::MultiVector{T,V}) where {T,V} = MultiVector{promote_type(T,F),V}(broadcast(Reduce.Algebra.:*,Refa,b.v))
         *(a::MultiVector{T,V},b::Reduce.RExpr) where {T,V} = MultiVector{promote_type(T,F),V}(broadcast(Reduce.Algebra.:*,a.v,b))
         *(a::Reduce.RExpr,b::MultiGrade{V}) where V = MultiGrade{V}(broadcast(Reduce.Algebra.:*,a,b.v))
         *(a::MultiGrade{V},b::Reduce.RExpr) where V = MultiGrade{V}(broadcast(Reduce.Algebra.:*,a.v,b))
@@ -443,9 +443,10 @@ function __init__()
         parany = (parany...,Reduce.RExpr)
         parval = (parval...,Reduce.RExpr)
         parsym = (parsym...,Reduce.RExpr)
-        generate_derivation(:(Reduce.Algebra),:RExpr,:df,:RExpr)
-        generate_derivation(:(Reduce.Algebra),:Symbol,:df,:RExpr)
-        generate_derivation(:(Reduce.Algebra),:Expr,:df,:RExpr)
+        for T ∈ (:RExpr,:Symbol,:Expr)
+            generate_inverses(:(Reduce.Algebra),T)
+            generate_derivation(:(Reduce.Algebra),T,:df,:RExpr)
+        end
     end
     @require SymPy="24249f21-da20-56a4-8eb1-6a02cf4ae2e6" generate_algebra(:SymPy,:Sym,:diff,:symbols)
     @require SymEngine="123dc426-2d89-5057-bbad-38513e3affd8" generate_algebra(:SymEngine,:Basic,:diff,:symbols)
