+++ src/algebra.jl
@@ -80,6 +80,11 @@ end
     end
 end
 
+for M ∈ (:Signature,:DiagonalForm)
+    @eval @pure loworder(V::$M{N,M,S,D,O}) where {N,M,S,D,O} = O≠0 ? $M{N,M,S,D,O-1}() : V
+end
+@pure loworder(::SubManifold{N,M,S}) where {N,M,S} = SubManifold{N,loworder(M),S}()
+
 function generate_mutators(M,F,set_val,SUB,MUL)
     for (op,set) ∈ ((:add,:(+=)),(:set,:(=)))
         sm = Symbol(op,:multi!)
@@ -105,7 +110,7 @@ function generate_mutators(M,F,set_val,SUB,MUL)
                         A,B,Q,Z = symmetricmask(V,a,b)
                         val = (typeof(V)<:Signature || count_ones(A&B)==0) ? (parity(A,B,V) ? $SUB(v) : v) : $MUL(parityinner(A,B,V),v)
                         if diffvars(V)≠0
-                            !iszero(Z) && (T≠Any ? (return true) : (val *= getbasis(V,Z)))
+                            !iszero(Z) && (T≠Any ? (return true) : (val *= getbasis(loworder(V),Z)))
                             count_ones(Q)+order(val)>diffmode(V) && (return false)
                         end
                         $s(m,val,(A⊻B)|Q,Dimension{N}())
@@ -118,7 +123,7 @@ function generate_mutators(M,F,set_val,SUB,MUL)
                         pcc,bas,cc = (hasinf(V) && hasorigin(V)) ? conformal(A,B,V) : (false,A⊻B,false)
                         val = (typeof(V)<:Signature || count_ones(A&B)==0) ? (parity(A,B,V)⊻pcc ? $SUB(v) : v) : $MUL(parityinner(A,B,V),pcc ? $SUB(v) : v)
                         if diffvars(V)≠0
-                            !iszero(Z) && (T≠Any ? (return true) : (val *= getbasis(V,Z)))
+                            !iszero(Z) && (T≠Any ? (return true) : (val *= getbasis(loworder(V),Z)))
                             count_ones(Q)+order(val)>diffmode(V) && (return false)
                         end
                         $s(m,val,bas|Q,Dimension{N}())
@@ -142,7 +147,7 @@ function generate_mutators(M,F,set_val,SUB,MUL)
                                 if !iszero(Z)
                                     T≠Any && (return true)
                                     _,_,Q,_ = symmetricmask(V,A,B)
-                                    v *= getbasis(V,Z)
+                                    v *= getbasis(loworder(V),Z)
                                     count_ones(Q)+order(v)>diffmode(V) && (return false)
                                 end
                             end
@@ -260,7 +265,7 @@ function mul(a::Basis{V},b::Basis{V},der=derive_mul(V,bits(a),bits(b),1,true)) w
     pcc,bas,cc = (hasinf(V) && hasorigin(V)) ? conformal(A,B,V) : (false,A⊻B,false)
     d = getbasis(V,bas|Q)
[…]
