### 2019-10-09 49134d8
separated Poincare duality from Hodge complement
