### 2019-10-09 74b9366
implemented factor of 2 for Poincare complement of null-basis
