### 2019-10-10 chakravala/Grassmann.jl#29 @goballooning
Personally I feel that this work can be divided into two papers:
- one targeting researchers in GA applications,
- and the other for the audience cares more about the fundamental concepts behind Grassmann.jl and how these concepts are properly handled to fit into Julia.
