### 2019-10-10 57ae200
implemented conformal ↑,↓ and various bug fixes
