### 2019-10-11 5ff3a2c
fixed #34 bug with symbolic += dispatch
