+++ src/Grassmann.jl
@@ -431,8 +431,28 @@ d(ω::T) where T<:TensorAlgebra{V} where V = V(∇)∧ω
 
 @pure ℙ(V) = ((i,o)=(hasinf(V),hasorigin(V));i+o==2 ? V : (i+o==0 ? S"∞∅"⊕V : V))
 
-↑(ω::T) where T<:TensorAlgebra{V} where V = (G=Λ(ℙ(V));ω+ω^2*G.v∞/2 + G.v∅)
-↓(ω::T) where T<:TensorAlgebra{V} where V = (G=Λ(ℙ(V));inv(G.v∞∅)*(G.v∞∅∧ω)/(-ω⋅G.v∞))
+function ↑(ω::T) where T<:TensorAlgebra{V} where V
+    PV = ℙ(V)
+    G = Λ(PV)
+    return if hasinf(PV) && hasorigin(PV)
+        ((G.v∞/2)*ω^2+G.v∅)+ω
+    else
+        ω2 = ω^2
+        iω2 = inv(ω2+1)
+        (i ? G.v∞ : G.v∅)*(ω2-1)*iω2 + 2*iω2*ω
+    end
+end
+
+function ↓(ω::T) where T<:TensorAlgebra{V} where V
+    PV = ℙ(V)
+    G = Λ(PV)
+    return if hasinf(PV) && hasorigin(PV)
+        inv(G.v∞∅)*(G.v∞∅∧ω)/(-ω⋅G.v∞)
+    else
+        b = i ? G.v∞ : G.v∅
+        ((ω∧b)*b)/(1-b⋅ω)
+    end
+end
 
 function __init__()
     @require Reduce="93e0c654-6965-5f22-aba9-9c1ae6b3c259" begin
