### 2019-10-12 cc12645
improved SubManifold and GeometryTypes.Point interoperability
