### 2019-10-12 0737421
revised SubManifold interoperability
