### 2019-10-12 9e7f457
improved SubManifold and GeometryTypes.Point interoperability for fast performance
