### 2019-10-13 d784b4d
resolved some minor interoperability issues
