### 2019-10-15 315d3b4
implemented Betti numbers and Euler characteristic
