+++ src/multivectors.jl
@@ -109,15 +109,19 @@ for Blade ∈ MSB
         @pure $Blade(b::Basis{V,G}) where {V,G} = $Blade{V}(b)
         @pure $Blade{V}(b::Basis{V,G}) where {V,G} = $Blade{V,G,b,Int}(1)
         $Blade{V}(v::T) where {V,T} = $Blade{V,0,Basis{V}(),T}(v)
-        $Blade{V}(v::TensorTerm{V}) where V = v
+        $Blade{V}(v::S) where S<:TensorTerm where V = v
         $Blade{V,G,B}(v::T) where {V,G,B,T} = $Blade{V,G,B,T}(v)
-        $Blade(v,b::TensorTerm{V}) where V = $Blade{V}(v,b)
-        $Blade{V}(v,b::S) where S<:TensorMixed where V = v*b
-        $Blade{V}(v::T,b::Basis{V,G}) where {V,G,T} = $Blade{V,G}(v,b)
+        $Blade(v,b::S) where S<:TensorTerm{V} where V = $Blade{V}(v,b)
+        $Blade{V}(v,b::S) where S<:TensorAlgebra where V = v*b
+        $Blade{V}(v,b::Basis{V,G}) where {V,G} = $Blade{V,G}(v,b)
+        $Blade{V}(v,b::Basis{W,G}) where {V,W,G} = $Blade{V,G}(v,b)
         function $Blade{V,G}(v::T,b::Basis{V,G}) where {V,G,T}
             order(v)+order(b)>diffmode(V) ? zero(V) : $Blade{V,G,b,T}(v)
         end
-        function $Blade{V,G}(v::T,b::Basis{V,G}) where T<:TensorTerm{V} where {V,G}
+        function $Blade{V,G}(v::T,b::Basis{W,G}) where {V,W,G,T}
+            order(v)+order(b)>diffmode(V) ? zero(V) : $Blade{V,G,V(b),T}(v)
+        end
+        function $Blade{V,G}(v::T,b::Basis{V,G}) where T<:TensorTerm where {V,G}
             order(v)+order(b)>diffmode(V) ? zero(V) : $Blade{V,G,b,Any}(v)
         end
         function $Blade{V,G,B}(b::T) where T<:TensorTerm{V} where {V,G,B}
@@ -485,7 +489,7 @@ MultiVector(v::MultiGrade{V}) where V = MultiVector{promote_type(typeval.(v.v)..
 import Base: isinf, isapprox
 import DirectSum: grade
 import AbstractTensors: scalar, involute, unit, even, odd
-export basis, grade, hasinf, hasorigin, isorigin, scalar, norm
+export basis, grade, hasinf, hasorigin, isorigin, scalar, norm, betti, χ
 
 const VBV = Union{MBlade,SBlade,MChain,SChain,MultiVector}
 
@@ -515,6 +519,37 @@ valuetype(t::MultiGrade) = promote_type(valuetype.(terms(t))...)
 @pure hasorigin(e::Basis{V}) where V = hasorigin(V) && (hasinf(V) ? e[2] : isodd(bits(e)))
 @pure hasorigin(t::Union{MBlade,SBlade}) = hasorigin(basis(t))
 @pure hasorigin(m::TensorAlgebra) = hasorigin(vectorspace(m))
+
+@inline χ(V,b::UInt,t) = iszero(t) ? 0 : isodd(count_ones(symmetricmask(V,b,b)[1])) ? 1 : -1
+χ(t::T) where T<:TensorTerm{V,G} where {V,G} = χ(V,bits(basis(t)),t)
[…]
