+++ src/Grassmann.jl
@@ -410,20 +410,21 @@ end
 # ParaAlgebra
 
 using Leibniz
-import Leibniz: ∂, d, ∇
-export ∇, ∂, d, ↑, ↓
+import Leibniz: ∂, d, ∇, Δ
+export ∇, Δ, ∂, d, ↑, ↓
 
 generate_products(:(Leibniz.Operator),:svec)
 
 @pure function (W::Signature{N})(d::Leibniz.Derivation{T,O}) where {N,T,O}
+    O < 1 && (return SChain{Int,W,1}(ones(Int,grade(W))))
     C = mixedmode(W)<0
-    V = diffvars(W)≠0 ? W : tangent(W,O<2 ? 2 : O,C ? Int(ndims(W)/2) : ndims(W))
+    V = diffvars(W)≠0 ? W : tangent(W,O,C ? Int(ndims(W)/2) : ndims(W))
     G,D = grade(V),diffvars(V)==1
     G2 = (C ? Int(G/2) : G)-1
     ∇ = sum([getbasis(V,1<<(D ? G : k+G))*getbasis(V,1<<k) for k ∈ 0:G2])
     isone(O) && (return ∇)
     x = (∇⋅∇)^div(isodd(O) ? O-1 : O,2)
-    isodd(O) ? sum([x*(getbasis(V,1<<(k+G))*getbasis(V,1<<k)) for k ∈ 0:G2]) : x
+    isodd(O) ? sum([(x*getbasis(V,1<<(k+G)))*getbasis(V,1<<k) for k ∈ 0:G2]) : x
 end
 
 ∂(ω::T) where T<:TensorAlgebra{V} where V = ω⋅V(∇)
@@ -454,6 +455,47 @@ function ↓(ω::T) where T<:TensorAlgebra{V} where V
     end
 end
 
+## skeleton
+
+export skeleton
+
+absym(t) = abs(t)
+absym(t::Basis) = t
+absym(t::T) where T<:TensorTerm{V,G} where {V,G} = SBlade{V,G}(absym(value(t)),basis(t))
+absym(t::SChain{T,V,G}) where {T,V,G} = SChain{T,V,G}(absym.(value(t)))
+absym(t::MChain{T,V,G}) where {T,V,G} = MChain{T,V,G}(absym.(value(t)))
+absym(t::MultiVector{T,V}) where {T,V} = MultiVector{T,V}(absym.(value(t)))
+
+function skeleton(x::T) where T<:TensorTerm{V} where V
+    X,B = absym(x),bits(basis(x))
+    count_ones(symmetricmask(V,B,B)[1])>1 ? X + subcomplex(absym(∂(X))) : X
+end
+skeleton(x::S) where {S<:TensorMixed{T,V} where T} where V = absym(x) + subcomplex(absym(∂(x)))
+function subcomplex(x::S) where {S<:TensorMixed{T,V} where T} where V
+    N,G,g = ndims(V),grade(x),0
+    ib = indexbasis(N,G)
+    for k ∈ 1:binomial(N,G)
+        if !iszero(x.v[k]) && count_ones(symmetricmask(V,ib[k],ib[k])[1]) > 0
+            g += skeleton(SBlade{V,G}(x.v[k],getbasis(V,ib[k])))
+        end
+    end
+    return g
+end
+function subcomplex(x::MultiVector{T,V} where T) where V
+   N,g = ndims(V),0
+   for i ∈ 2:N
+        R = binomsum(N,i)
+        ib = indexbasis(N,i)
+        for k ∈ 1:binomial(N,i)
+            if !iszero(x.v[k+R])
+                G = count_ones(symmetricmask(V,ib[k],ib[k])[1])
[…]
