### 2019-10-16 5a2f3c3
revised 𝒫, skeleton, subcomplex, and chain
