### 2019-10-18 9f14a5a
improved ∇, ∂, subcomplex graphs, and symbolic compatibility
