### 2019-10-19 ed96ced
revised JuliaCon paper with some new examples
