+++ paper/paper.tex
@@ -6,6 +6,7 @@
 %\usepackage[utf8]{inputenc}
 
 \usepackage{graphicx}
+\usepackage{subcaption}
 %\usepackage[left=1.00in, right=1.00in, top=1.00in, bottom=1.00in]{geometry}
 
 \input{mathsym}
@@ -46,30 +47,31 @@
 %\input{header}
 
 \maketitle
-
-\begin{abstract}
-	The \textit{Grassmann.jl}
-	package provides tools for computations based on multi-linear algebra and spin groups using the extended geometric algebra known as Grassmann-Clifford-Hestenes-Leibniz algebra.
-	The primary operations are
-	outer, regressive, inner, geometric, and cross products along with the Hodge star, adjoint, and multivector reversal operations. 
-	The kernelized operations are built up from composite sparse tensor products and Hodge duality, with high dimensional support for up to 62 indices using staged caching and precompilation. Code generation enables concise and also extensible definitions.
-	The \textit{DirectSum.jl}
-	multivector parametric type polymorphism is based on tangent vector spaces and conformal projective geometry.
-	Additionally, universal interoperability between different sub-algebras is enabled by \textit{AbstractTensors.jl},
-	on which the type system is built.
-	%\headingtable
-\end{abstract}
-
-Mathematical foundations and some of the nuances in the definitions specific to the \textit{Grassmann.jl} implementation are concisely described, along with the accompanying support packages that provide an extensible platform for computing with geometric algebra at high dimensions.
-The design is based on the \verb`TensorAlgebra` abstract type interoperability from \textit{AbstractTensors.jl} with a \verb`VectorBundle` type parameter from \textit{DirectSum.jl}. Abstract vector space type operations happen at compile-time, resulting in a differential conformal geometric algebra of hyper-dual multivector forms.
-
+\begin{figure}[ht]
+\centerline{\includegraphics[width=4cm]{grassmann.png}}
+\caption*{github.com/chakravala/Grassmann.jl}
+\end{figure}
 \begin{itemize}
 	\item \textbf{DirectSum.jl}: Abstract tangent bundle vector space types (unions, intersections, sums, etc.)
 	\item \textbf{AbstractTensors.jl}: Tensor algebra abstract type interoperability with vector bundle parameter	
-	\item \textbf{Grassmann.jl}: $\langle$Leibniz+Grassmann-Clifford-Hestenes$\rangle$ differential conformal geometric algebra
+	\item \textbf{Grassmann.jl}: $\langle$Leibniz+Grassmann-Clifford-Hestenes$\rangle$ differential geometric algebra of multivector forms
 	\item \textbf{Leibniz.jl}: Derivation operator algebras for tensor fields
 	\item \textbf{Reduce.jl}: Symbolic parser generator for Julia expressions using REDUCE algebra term rewriter
 \end{itemize}
+\begin{abstract}
+	The \textit{Grassmann.jl}
+	package provides tools for computations based on multi-linear algebra and spin groups using the extended geometric algebra known as Leibniz-Grassmann-Clifford-Hestenes algebra.
+	Combinatorial products include
+	exterior, regressive, inner, and geometric; along with the Hodge star, adjoint, reversal, and boundary operators.
+	The kernelized operations are built up from composite sparse tensor products and Hodge duality, with high dimensional support for up to 62 indices using staged caching and precompilation. Code generation enables concise with highly extensible definitions.
[…]
