### 2019-10-27 e9f723e
improved Betti numbers, paper revisions, changed Blade to Simplex
