+++ paper/paper.tex
@@ -48,16 +48,10 @@
 
 \maketitle
 \begin{figure}[ht]
-\centerline{\includegraphics[width=4cm]{grassmann.png}}
+%\centerline{\includegraphics[width=4cm]{grassmann.png}}
+\centerline{\includegraphics[width=4cm]{../docs/src/assets/logo.png}}
 \caption*{github.com/chakravala/Grassmann.jl}
 \end{figure}
-\begin{itemize}
-	\item \textbf{DirectSum.jl}: Abstract tangent bundle vector space types (unions, intersections, sums, etc.)
-	\item \textbf{AbstractTensors.jl}: Tensor algebra abstract type interoperability with vector bundle parameter	
-	\item \textbf{Grassmann.jl}: $\langle$Leibniz+Grassmann-Clifford-Hestenes$\rangle$ differential geometric algebra of multivector forms
-	\item \textbf{Leibniz.jl}: Derivation operator algebras for tensor fields
-	\item \textbf{Reduce.jl}: Symbolic parser generator for Julia expressions using REDUCE algebra term rewriter
-\end{itemize}
 \begin{abstract}
 	The \textit{Grassmann.jl}
 	package provides tools for computations based on multi-linear algebra and spin groups using the extended geometric algebra known as Leibniz-Grassmann-Clifford-Hestenes algebra.
@@ -70,6 +64,13 @@
 	on which the type system is built.
 	%\headingtable
 \end{abstract}
+\begin{itemize}
+	\item \textbf{DirectSum.jl}: Abstract tangent bundle vector space types (unions, intersections, sums, etc.)
+	\item \textbf{AbstractTensors.jl}: Tensor algebra abstract type interoperability with vector bundle parameter
+	\item \textbf{Grassmann.jl}: $\langle$Leibniz+Grassmann-Clifford-Hestenes$\rangle$ differential geometric algebra of multivector forms
+	\item \textbf{Leibniz.jl}: Derivation operator algebras for tensor fields
+	\item \textbf{Reduce.jl}: Symbolic parser generator for Julia expressions using REDUCE algebra term rewriter
+\end{itemize}
 Mathematical foundations and some of the nuances in the definitions specific to the \textit{Grassmann.jl} implementation are concisely described, along with the accompanying support packages that provide an extensible platform for computing with geometric algebra at high dimensions.
 The design is based on the \verb`TensorAlgebra` abstract type interoperability from \textit{AbstractTensors.jl} with a \verb`VectorBundle` type parameter from \textit{DirectSum.jl}. Abstract vector space type operations happen at compile-time, resulting in a differential conformal geometric algebra of hyper-dual multivector forms.
 
@@ -173,7 +174,7 @@ In order to work with a \verb`TensorAlgebra{V}`, it is necessary for some comput
 Staging of precompilation and caching is designed so that a user can smoothly transition between very high dimensional and low dimensional algebras in a single session, with varying levels of extra caching and optimizations.
[…]
