+++ src/algebra.jl
@@ -186,7 +186,7 @@ for side ∈ (:left,:right)
             d = getbasis(V,complement(ndims(V),B,diffvars(V),$(c≠h ? 0 : :(hasinf(V)+hasorigin(V)))))
             mixedmode(V)<0 && throw(error("Complement for mixed tensors is undefined"))
             v = $(c≠h ? :($pn(V,B,value(d))) : :(value(d)))
-            typeof(V)<:Signature ? ($p(b) ? SBlade{V}(-v,d) : isone(v) ? d : SBlade{V}(v,d)) : SBlade{V}($p(b)*v,d)
+            typeof(V)<:Signature ? ($p(b) ? Simplex{V}(-v,d) : isone(v) ? d : Simplex{V}(v,d)) : Simplex{V}($p(b)*v,d)
         end
         for B ∈ MSB
             @eval $c(b::$B) = value(b)≠0 ? value(b)*$c(basis(b)) : g_zero(vectorspace(b))
@@ -219,10 +219,10 @@ export involute
 for r ∈ (:reverse,:involute,:conj)
     p = Symbol(:parity,r)
     @eval @pure function $r(b::Basis{V,G,B}) where {V,G,B}
-        $p(grade(V,B)) ? SBlade{V}(-value(b),b) : b
+        $p(grade(V,B)) ? Simplex{V}(-value(b),b) : b
     end
-    for Blade ∈ MSB
-        @eval $r(b::$Blade) = value(b) ≠ 0 ? value(b) * $r(basis(b)) : g_zero(vectorspace(b))
+    for implex ∈ MSB
+        @eval $r(b::$implex) = value(b) ≠ 0 ? value(b) * $r(basis(b)) : g_zero(vectorspace(b))
     end
 end
 
@@ -269,22 +269,22 @@ function mul(a::Basis{V},b::Basis{V},der=derive_mul(V,bits(a),bits(b),1,true)) w
     A,B,Q,Z = symmetricmask(V,bits(a),bits(b))
     pcc,bas,cc = (hasinf(V) && hasorigin(V)) ? conformal(A,B,V) : (false,A⊻B,false)
     d = getbasis(V,bas|Q)
-    out = (typeof(V)<:Signature || count_ones(A&B)==0) ? (parity(a,b)⊻pcc ? SBlade{V}(-1,d) : d) : SBlade{V}((pcc ? -1 : 1)*parityinner(A,B,V),d)
-    diffvars(V)≠0 && !iszero(Z) && (out = SBlade{V}(getbasis(loworder(V),Z),out))
-    return cc ? (v=value(out);out+SBlade{V}(hasinforigin(V,A,B) ? -(v) : v,getbasis(V,conformalmask(V)⊻bits(d)))) : out
+    out = (typeof(V)<:Signature || count_ones(A&B)==0) ? (parity(a,b)⊻pcc ? Simplex{V}(-1,d) : d) : Simplex{V}((pcc ? -1 : 1)*parityinner(A,B,V),d)
+    diffvars(V)≠0 && !iszero(Z) && (out = Simplex{V}(getbasis(loworder(V),Z),out))
+    return cc ? (v=value(out);out+Simplex{V}(hasinforigin(V,A,B) ? -(v) : v,getbasis(V,conformalmask(V)⊻bits(d)))) : out
 end
 
-for Blade ∈ MSB
+for implex ∈ MSB
     @eval begin
-        function *(a::$Blade{V},b::Basis{V}) where V
+        function *(a::$implex{V},b::Basis{V}) where V
             v = derive_mul(V,bits(basis(a)),bits(b),a.v,true)
             bas = mul(basis(a),b,v)
-            order(a.v)+order(bas)>diffmode(V) ? zero(V) : SBlade{V}(v,bas)
[…]
