+++ src/utilities.jl
@@ -160,7 +160,7 @@ Base.@pure promote_type(t...) = Base.promote_type(t...)
     assign_expr!(e,x,:M,:(Int(N/2)))
     assign_expr!(e,x,:t,vec≠:mvec ? :Any : :(promote_type($T,$S)))
     assign_expr!(e,x,:out,mv≠0 ? :(t=Any;convert(svec(N,Any),out)) : :(zeros($vec(N,t))))
-    assign_expr!(e,x,:mv,:(MBlade{V,0,getbasis(V,0),$(mv≠0 ? Any : :t)}(convert($(mv≠0 ? Any : :t),$mv))))
+    assign_expr!(e,x,:mv,:(MSimplex{V,0,getbasis(V,0),$(mv≠0 ? Any : :t)}(convert($(mv≠0 ? Any : :t),$mv))))
     assign_expr!(e,x,:r,:(binomsum(N,G)))
     assign_expr!(e,x,:bng,:(binomial(N,G)))
     assign_expr!(e,x,:bnl,:(binomial(N,L)))
