### 2019-10-29 23501ab
introduced vectorfield SubManifold interoperability for GeometryTypes.Point visualization
