### 2019-10-31 5559506
improved vectorfield and finalized JuliaCon paper, fixed #37
