+++ paper/paper.tex
@@ -74,7 +74,7 @@
 Mathematical foundations and some of the nuances in the definitions specific to the \textit{Grassmann.jl} implementation are concisely described, along with the accompanying support packages that provide an extensible platform for computing with geometric algebra at high dimensions.
 The design is based on the \verb`TensorAlgebra` abstract type interoperability from \textit{AbstractTensors.jl} with a \verb`VectorBundle` type parameter from \textit{DirectSum.jl}. Abstract vector space type operations happen at compile-time, resulting in a differential conformal geometric algebra of hyper-dual multivector forms.
 
-The nature of the geometric algebra code generation enables one to easily extend the abstract product operations to any specific number field type (including Leibniz differentials with \textit{Leibniz.jl} or symbolic coefficients with \textit{Reduce.jl}), by making use of Julia's type system. Mixed tensor products with their coefficients are constructed from these operations to work with bivector elements of Lie groups \cite{hestenes}\cite{shahshahani}.
+The nature of the geometric algebra code generation enables one to easily extend the abstract product operations to any specific number field type (including differential operators with \textit{Leibniz.jl} or symbolic coefficients with \textit{Reduce.jl}), by making use of Julia's type system. Mixed tensor products with their coefficients are constructed from these operations to work with bivector elements of Lie groups \cite{hestenes}\cite{shahshahani}.
 
 \section{Direct sum parametric type polymorphism}
 
@@ -115,7 +115,7 @@ such that computing unions of submanifolds is done by inspecting the parameter $
 
 The metric signature of the \verb+Basis{V,1}+ elements of a vector space \verb+V+ can be specified with the \verb+V"..."+ constructor by using \verb-+- and \verb+-+ to specify whether the \verb+Basis{V,1}+ element of the corresponding index squares to \verb`+1` or \verb`-1`.
 For example, \verb`S"+++"` constructs a positive definite 3-dimensional \verb`VectorBundle`.
-It is also possible to specify an arbitrary \verb`DiagonalForm` having numerical values for the basis with \verb`D"0,0,0"` or \verb`D"-1,1,1,1"`, although the $\pm$ format provides better performance.
+It is also possible to specify an arbitrary \verb`DiagonalForm` having numerical values for the basis with degenracy \verb`D"1,1,1,0"`, although the $\pm$ format has a more compact representation.
 Further development will result in more metric types.
 
 Declaring an additional plane at infinity is done by specifying it in the string constructor with \verb`∞` at the first index (i.e. Riemann sphere \verb`S"∞+++"`). The hyperbolic geometry can be declared by \verb`∅` subsequently (i.e. Minkowski spacetime \verb`S"∅+++"`). 
[…]
