+++ src/Grassmann.jl
@@ -428,17 +428,15 @@ end
 ∂(ω::T) where T<:TensorAlgebra{V} where V = ω⋅V(∇)
 d(ω::T) where T<:TensorAlgebra{V} where V = V(∇)∧ω
 
-@pure ℙ(V) = ((i,o)=(hasinf(V),hasorigin(V));i+o==2 ? V : (i+o==0 ? S"∞∅"⊕V : V))
-
 function ↑(ω::T) where T<:TensorAlgebra{V} where V
-    PV = ℙ(V)
-    G = Λ(PV)
-    return if hasinf(PV) && hasorigin(PV)
+    !(hasinf(V)||hasorigin(V)) && (return ω)
+    G = Λ(V)
+    return if hasinf(V) && hasorigin(V)
         ((G.v∞/2)*ω^2+G.v∅)+ω
     else
         ω2 = ω^2
         iω2 = inv(ω2+1)
-        (hasinf(PV) ? G.v∞ : G.v∅)*(ω2-1)*iω2 + 2*iω2*ω
+        (hasinf(V) ? G.v∞ : G.v∅)*(ω2-1)*iω2 + 2*iω2*ω
     end
 end
 function ↑(ω,b)
@@ -453,12 +451,12 @@ function ↑(ω,p,m)
 end
 
 function ↓(ω::T) where T<:TensorAlgebra{V} where V
-    PV = ℙ(V)
-    G = Λ(PV)
-    return if hasinf(PV) && hasorigin(PV)
+    !(hasinf(V)||hasorigin(V)) && (return ω)
+    G = Λ(V)
+    return if hasinf(V) && hasorigin(V)
         inv(G.v∞∅)*(G.v∞∅∧ω)/(-ω⋅G.v∞)
     else
-        b = hasinf(PV) ? G.v∞ : G.v∅
+        b = hasinf(V) ? G.v∞ : G.v∅
         ((ω∧b)*b)/(1-b⋅ω)
     end
 end
@@ -494,8 +492,9 @@ function chain(t::S,::Val{T}=Val{true}()) where S<:TensorTerm{V} where {V,T}
 end
 path(t) = chain(t,Val{false}())
 
-𝒫(t::T) where T<:TensorAlgebra = skeleton(t,Val{false}())
-subcomplex(x::S,v=Val{true}()) where S<:TensorAlgebra = skeleton(absym(∂(x)),v)
+@inline (::Leibniz.Derivation)(x::T,v=Val{true}()) where T<:TensorAlgebra = skeleton(x,v)
+𝒫(t::T) where T<:TensorAlgebra = Δ(t,Val{false}())
+subcomplex(x::S,v=Val{true}()) where S<:TensorAlgebra = Δ(absym(∂(x)),v)
 function skeleton(x::S,v::Val{T}=Val{true}()) where S<:TensorTerm{V} where {V,T}
     B = bits(basis(x))
     count_ones(symmetricmask(V,B,B)[1])>0 ? absym(x)+skeleton(absym(∂(x)),v) : (T ? g_zero(V) : absym(x))
@@ -601,10 +600,10 @@ function __init__()
         Base.convert(::Type{GeometryTypes.Point},t::MChain{T,V,G}) where {T,V,G} = G == 1 ? GeometryTypes.Point(value(vector(t))) : GeometryTypes.Point(zeros(T,ndims(V))...)
         Base.convert(::Type{GeometryTypes.Point},t::SChain{T,V,G}) where {T,V,G} = G == 1 ? GeometryTypes.Point(value(vector(t))) : GeometryTypes.Point(zeros(T,ndims(V))...)
         GeometryTypes.Point(t::T) where T<:TensorAlgebra = convert(GeometryTypes.Point,t)
[…]
