### 2019-10-31 #18 @chakravala
I have committed the final draft for my JuliaCon paper, here is a PDF download of it
