### 2019-11-01 6838f5f
revised Project.toml, Makie.jl links, and revised paper grammar
