+++ Project.toml
@@ -20,6 +20,12 @@ julia = "1"
 Reduce = "1.2"
 DirectSum = "0.4"
 AbstractTensors = "0.3"
+AbstractLattices = ">=0"
+Combinatorics = ">=0"
+ComputedFieldTypes = ">=0"
+Leibniz = ">=0"
+Requires = ">=0"
+StaticArrays = ">=0"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
