### 2019-11-04 2b4d96d
created projective unsplitter norm and tidelift enterprise referral
