+++ .github/FUNDING.yml
@@ -4,6 +4,6 @@ github: [chakravala]
 patreon: dreamscatter
 open_collective: # Replace with a single Open Collective username
 ko_fi: # Replace with a single Ko-fi username
-tidelift: julia/Grassmann
+tidelift: "julia/Grassmann"
 community_bridge: # Replace with a single Community Bridge project-name e.g., cloud-foundry
 custom: https://liberapay.com/chakravala
