+++ Project.toml
@@ -14,18 +14,13 @@ LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Reduce = "93e0c654-6965-5f22-aba9-9c1ae6b3c259"
 Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
+SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 
 [compat]
 julia = "1"
 Reduce = "1.2"
 DirectSum = "0.4"
 AbstractTensors = "0.3"
-AbstractLattices = ">=0"
-Combinatorics = ">=0"
-ComputedFieldTypes = ">=0"
-Leibniz = ">=0"
-Requires = ">=0"
-StaticArrays = ">=0"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"

[…]
