### 2019-11-06 #18 @matbesancon
Hi, the paper really need to be built with the same environment to be accepted for the proceedings of the JuliaCon, it basically ensures the layout and the tooling will work for all papers we publish
