### 2019-11-14 fb83d62
created hybrid generated function for MultiVector arithmetic
