+++ paper/paper.tex
@@ -48,8 +48,8 @@
 
 \maketitle
 \begin{figure}[ht]
-\centerline{\includegraphics[width=4cm]{img/grassmann.png}}
-%\centerline{\includegraphics[width=4cm]{../docs/src/assets/logo.png}}
+%\centerline{\includegraphics[width=4cm]{img/grassmann.png}}
+\centerline{\includegraphics[width=4cm]{../docs/src/assets/logo.png}}
 \caption*{github.com/chakravala/Grassmann.jl}
 \end{figure}
 \begin{abstract}
