+++ src/algebra.jl
@@ -19,7 +19,6 @@ const ExprField = Union{Expr,Symbol}
 
 ## mutating operations
 
-
 const Sym = :DirectSum
 const SymField = Any
 
@@ -77,6 +76,66 @@ end
     end
 end
 
+function derive_pre(V,A,B,a,b,p)
+    if !(diffvars(V)≠0 && mixedmode(V)<0)
+        return Expr(:call,p,a,b)
+    else
+        return :(derive_post($V,$(Val{A}()),$(Val{B}()),a,b,$p))
+    end
+end
+
+function derive_post(V,::Val{A},::Val{B},v,x::Bool) where {A,B}
+    if !(diffvars(V)≠0 && mixedmode(V)<0)
+        return v
+    else
+        sa,sb = symmetricsplit(V,A),symmetricsplit(V,B)
+        ca,cb = count_ones(sa[2]),count_ones(sb[2])
+        return if (ca == cb == 0) || ((ca ≠ 0) && (cb ≠ 0))
+            v
+        else
+            prev = ca == 0 ? (x ? one(typeof(v)) : v) : (x ? v : one(typeof(v)))
+            for k ∈ DirectSum.indexsplit((ca==0 ? sa : sb)[1],ndims(V))
+                prev = derive(prev,getbasis(V,k))
+            end
+            prev
+        end
+    end
+end
+
+function derive_post(V,::Val{A},::Val{B},a,b,*) where {A,B}
+    if !(diffvars(V)≠0 && mixedmode(V)<0)
+        return a*b
+    else
+        sa,sb = symmetricsplit(V,A),symmetricsplit(V,B)
+        ca,cb = count_ones(sa[2]),count_ones(sb[2])
+        α,β = if (ca == cb == 0) || ((ca ≠ 0) && (cb ≠ 0))
+            a,b
+        else
+            prev = ca == 0 ? (a,b) : (b,a)
+            for k ∈ DirectSum.indexsplit((ca==0 ? sa : sb)[1],ndims(V))
+                prev = derive(prev[2],getbasis(V,k),prev[1],true)
+            end
+            #base = getbasis(V,0)
+            while typeof(prev[1]) <: TensorTerm
+                basi = basis(prev[1])
+                #base *= basi
+                inds = DirectSum.indexsplit(bits(basi),ndims(V))
+                prev = (value(prev[1]),prev[2])
+                for k ∈ inds
+                    prev = derive(prev[2],getbasis(V,k),prev[1],true)
+                end
+            end
+            #base ≠ getbasis(V,0) && (prev = (base*prev[1],prev[2]))
+            ca == 0 ? prev : (prev[2],prev[1])
+        end
+        return α*β
+    end
+end
+
+@pure isnull(::Expr) = false
+@pure isnull(::Symbol) = false
+isnull(n) = iszero(n)
+
 for M ∈ (:Signature,:DiagonalForm)
     @eval @pure loworder(V::$M{N,M,S,D,O}) where {N,M,S,D,O} = O≠0 ? $M{N,M,S,D,O-1}() : V
 end
@@ -86,11 +145,14 @@ add_val(set,expr,val,OP) = Expr(OP∉(:-,:+) ? :.= : set,expr,OP∉(:-,:+) ? Exp
 
 set_val(set,expr,val) = Expr(:(=),expr,set≠:(=) ? Expr(:call,:($Sym.:∑),expr,val) : val)
 
+pre_val(set,expr,val) = set≠:(=) ? :(isnull($expr) ? ($expr=Expr(:call,:($Sym.:∑),$val)) : push!($expr.args,$val)) : Expr(:(=),expr,val)
+
 function generate_mutators(M,F,set_val,SUB,MUL)
[…]
