### 2019-11-14 cea3871
improved StaticArrays performance of MultiVector inbounds
