### 2019-11-14 1ef43b4
significant performance improvements for low-dimensional subalgebras
