+++ src/algebra.jl
@@ -76,60 +76,60 @@ end
     end
 end
 
+function derive_pre(V,A,B,v,x)
+    if !(diffvars(V)≠0 && mixedmode(V)<0)
+        return v
+    else
+        return :(derive_post($V,$(Val{A}()),$(Val{B}()),$v,$x))
+    end
+end
+
 function derive_pre(V,A,B,a,b,p)
     if !(diffvars(V)≠0 && mixedmode(V)<0)
         return Expr(:call,p,a,b)
     else
-        return :(derive_post($V,$(Val{A}()),$(Val{B}()),a,b,$p))
+        return :(derive_post($V,$(Val{A}()),$(Val{B}()),$a,$b,$p))
     end
 end
 
 function derive_post(V,::Val{A},::Val{B},v,x::Bool) where {A,B}
-    if !(diffvars(V)≠0 && mixedmode(V)<0)
-        return v
+    sa,sb = symmetricsplit(V,A),symmetricsplit(V,B)
+    ca,cb = count_ones(sa[2]),count_ones(sb[2])
+    return if (ca == cb == 0) || ((ca ≠ 0) && (cb ≠ 0))
+        v
     else
-        sa,sb = symmetricsplit(V,A),symmetricsplit(V,B)
-        ca,cb = count_ones(sa[2]),count_ones(sb[2])
-        return if (ca == cb == 0) || ((ca ≠ 0) && (cb ≠ 0))
-            v
-        else
-            prev = ca == 0 ? (x ? one(typeof(v)) : v) : (x ? v : one(typeof(v)))
-            for k ∈ DirectSum.indexsplit((ca==0 ? sa : sb)[1],ndims(V))
-                prev = derive(prev,getbasis(V,k))
-            end
-            prev
+        prev = ca == 0 ? (x ? one(typeof(v)) : v) : (x ? v : one(typeof(v)))
+        for k ∈ DirectSum.indexsplit((ca==0 ? sa : sb)[1],ndims(V))
+            prev = derive(prev,getbasis(V,k))
         end
+        prev
     end
 end
 
 function derive_post(V,::Val{A},::Val{B},a,b,*) where {A,B}
-    if !(diffvars(V)≠0 && mixedmode(V)<0)
-        return a*b
+    sa,sb = symmetricsplit(V,A),symmetricsplit(V,B)
+    ca,cb = count_ones(sa[2]),count_ones(sb[2])
+    α,β = if (ca == cb == 0) || ((ca ≠ 0) && (cb ≠ 0))
+        a,b
     else
-        sa,sb = symmetricsplit(V,A),symmetricsplit(V,B)
-        ca,cb = count_ones(sa[2]),count_ones(sb[2])
-        α,β = if (ca == cb == 0) || ((ca ≠ 0) && (cb ≠ 0))
-            a,b
-        else
-            prev = ca == 0 ? (a,b) : (b,a)
-            for k ∈ DirectSum.indexsplit((ca==0 ? sa : sb)[1],ndims(V))
+        prev = ca == 0 ? (a,b) : (b,a)
+        for k ∈ DirectSum.indexsplit((ca==0 ? sa : sb)[1],ndims(V))
+            prev = derive(prev[2],getbasis(V,k),prev[1],true)
+        end
+        #base = getbasis(V,0)
+        while typeof(prev[1]) <: TensorTerm
+            basi = basis(prev[1])
+            #base *= basi
+            inds = DirectSum.indexsplit(bits(basi),ndims(V))
+            prev = (value(prev[1]),prev[2])
+            for k ∈ inds
                 prev = derive(prev[2],getbasis(V,k),prev[1],true)
             end
-            #base = getbasis(V,0)
-            while typeof(prev[1]) <: TensorTerm
[…]
