+++ src/algebra.jl
@@ -582,7 +582,7 @@ export ×
 
 Cross product: ω×η = ⋆(ω∧η)
 """
-cross(a::TensorAlgebra{V},b::TensorAlgebra{V}) where V = ⋆(a∧b)
+cross(t::T...) where T<:TensorAlgebra = ⋆(∧(t...))
 
 @pure function cross(a::Basis{V},b::Basis{V}) where V
     p,C,t,Z = crossprod(a,b)
@@ -717,20 +717,36 @@ function generate_products(Field=Field,VEC=:mvec,MUL=:*,ADD=:+,SUB=:-,CONJ=:conj
             end
             return false
         end
-        function adjoint(m::MultiVector{T,V}) where {T<:$Field,V}
-            if mixedmode(V)<0
-                $(insert_expr((:N,:M,:bs,:bn),VEC)...)
-                out = zeros($VEC(N,$TF))
-                for g ∈ 1:N+1
-                    ib = indexbasis(N,g-1)
-                    @inbounds for i ∈ 1:bn[g]
-                        @inbounds setmulti!(out,$CONJ(m.v[bs[g]+i]),dual(V,ib[i],M))
+        @generated function adjoint(m::MultiVector{T,V}) where {T<:$Field,V}
+            if ndims(V)<cache_limit
+                if mixedmode(V)<0
+                    $(insert_expr((:N,:M,:bs,:bn),:vec)...)
+                    out = zeros(svec(N,Any))
+                    for g ∈ 1:N+1
+                        ib = indexbasis(N,g-1)
+                        @inbounds for i ∈ 1:bn[g]
+                            @inbounds setmulti!_pre(out,:($$CONJ(m.v[$(bs[g]+i)])),dual(V,ib[i],M))
+                        end
                     end
+                else
+                    out = :(MultiVector{$(dual(V))}(SVector($$CONJ.(value(m)))))
                 end
-            else
-                out = $CONJ.(value(m))
-            end
-            MultiVector{$TF,dual(V)}(out)
+                return :(MultiVector{$(dual(V))}($(Expr(:call,:SVector,out...))))
+            else return quote
+                if mixedmode(V)<0
+                    $(insert_expr((:N,:M,:bs,:bn),$(QuoteNode(VEC)))...)
+                    out = zeros($$VEC(N,$$TF))
+                    for g ∈ 1:N+1
+                        ib = indexbasis(N,g-1)
+                        @inbounds for i ∈ 1:bn[g]
+                            @inbounds setmulti!(out,$$CONJ(m.v[bs[g]+i]),dual(V,ib[i],M))
+                        end
+                    end
+                else
+                    out = $$CONJ.(value(m))
+                end
+                MultiVector{$$TF,dual(V)}(out)
+            end end
         end
         *(a::F,b::Basis{V}) where {F<:$EF,V} = Simplex{V}(a,b)
         *(a::Basis{V},b::F) where {F<:$EF,V} = Simplex{V}(b,a)
@@ -762,17 +778,30 @@ function generate_products(Field=Field,VEC=:mvec,MUL=:*,ADD=:+,SUB=:-,CONJ=:conj
     end
     for Chain ∈ MSC
         @eval begin
-            function adjoint(m::$Chain{T,V,G}) where {T<:$Field,V,G}
-                if mixedmode(V)<0
-                    $(insert_expr((:N,:M,:ib),VEC)...)
-                    out = zeros($VEC(N,G,$TF))
-                    for i ∈ 1:binomial(N,G)
-                        @inbounds setblade!(out,$CONJ(m.v[i]),dual(V,ib[i],M),Dimension{N}())
+            @generated function adjoint(m::$Chain{T,V,G}) where {T<:$Field,V,G}
[…]
