+++ .github/FUNDING.yml
@@ -2,8 +2,9 @@
 
 github: [chakravala]
 patreon: dreamscatter
+liberapay: chakravala
 open_collective: # Replace with a single Open Collective username
 ko_fi: # Replace with a single Ko-fi username
 tidelift: "julia/Grassmann"
-community_bridge: # Replace with a single Community Bridge project-name e.g., cloud-foundry
-custom: https://liberapay.com/chakravala
+otechie: # Replace with a single Otechie username
+custom: https://music.crucialflow.com
