+++ src/composite.jl
@@ -24,41 +24,32 @@ function Base.expm1(t::T) where T<:TensorAlgebra{V} where V
     return S
 end
 
-@eval function Base.expm1(b::MultiVector{T,V}) where {T,V}
-    B = value(b)
-    sb,nb = scalar(b),norm(B)
-    sb ≈ nb && (return Simplex{V}(expm1(sb)))
-    $(insert_expr((:N,:t,:out,:bs,:bn),:mvec,:T,Float64)...)
-    S = zeros(mvec(N,t))
-    term = zeros(mvec(N,t))
-    S .= B
-    out .= value(b^2)/2
-    norms = MVector(nb,norm(out),norm(term))
-    k::Int = 3
-    @inbounds while (norms[2]<norms[1] || norms[2]>1) && k ≤ 10000
-        S += out
-        ns = norm(S)
-        @inbounds ns ≈ norms[3] && break
-        term .= out
-        out .= 0
-        # term *= b/k
-        for g ∈ 1:N+1
-            Y = indexbasis(N,g-1)
-            @inbounds for i ∈ 1:bn[g]
-                @inbounds val = B[bs[g]+i]/k
-                val≠0 && for G ∈ 1:N+1
-                    @inbounds R = bs[G]
-                    X = indexbasis(N,G-1)
-                    @inbounds for j ∈ 1:bn[G]
-                        @inbounds geomaddmulti!(V,out,X[j],Y[i],*(term[R+j],val))
-                    end
-                end
-            end
+@eval @generated function Base.expm1(b::MultiVector{T,V}) where {T,V}
+    loop = generate_loop_multivector(V,:term,:B,:*,:geomaddmulti!,geomaddmulti!_pre,:k)
+    return quote
+        B = value(b)
+        sb,nb = scalar(b),norm(B)
+        sb ≈ nb && (return Simplex{V}(expm1(sb)))
+        $(insert_expr(loop[1],:mvec,:T,Float64)...)
+        S = zeros(mvec(N,t))
+        term = zeros(mvec(N,t))
+        S .= B
+        out .= value(b^2)/2
+        norms = MVector(nb,norm(out),norm(term))
+        k::Int = 3
+        @inbounds while (norms[2]<norms[1] || norms[2]>1) && k ≤ 10000
+            S += out
+            ns = norm(S)
+            @inbounds ns ≈ norms[3] && break
+            term .= out
+            out .= 0
+            # term *= b/k 
+            $(loop[2])
+            @inbounds norms .= (norms[2],norm(out),ns)
+            k += 1
         end
-        @inbounds norms .= (norms[2],norm(out),ns)
-        k += 1
+        return MultiVector{t,V}(S)
     end
-    return MultiVector{t,V}(S)
 end
 
 function qlog(w::T,x::Int=10000) where T<:TensorAlgebra{V} where V
@@ -79,45 +70,36 @@ function qlog(w::T,x::Int=10000) where T<:TensorAlgebra{V} where V
     return 2S
 end # http://www.netlib.org/cephes/qlibdoc.html#qlog
 
-@eval function qlog_fast(b::MultiVector{T,V,E},x::Int=10000) where {T,V,E}
-    $(insert_expr((:N,:t,:out,:bs,:bn),:mvec,:T,Float64)...)
-    f = norm(b)
-    w2::MultiVector{T,V,E} = b^2
-    B = value(w2)
-    S = zeros(mvec(N,t))
-    prod = zeros(mvec(N,t))
-    term = zeros(mvec(N,t))
-    S .= value(b)
-    out .= value(b*w2)
-    term .= out/3
-    norms = MVector(f,norm(term),f)
-    k::Int = 5
[…]
