### 2019-11-17 e7fd9e0
improved dispatch performance and deprecated MSimplex, MChain
