+++ src/algebra.jl
@@ -310,14 +310,14 @@ for side ∈ (:left,:right)
     c,p = Symbol(:complement,side),Symbol(:parity,side)
     h,pg,pn = Symbol(c,:hodge),Symbol(p,:hodge),Symbol(p,:null)
     for (c,p) ∈ ((c,p),(h,pg))
-        @eval @pure function $c(b::Basis{V,G,B}) where {V,G,B}
-            d = getbasis(V,complement(ndims(V),B,diffvars(V),$(c≠h ? 0 : :(hasinf(V)+hasorigin(V)))))
-            mixedmode(V)<0 && throw(error("Complement for mixed tensors is undefined"))
-            v = $(c≠h ? :($pn(V,B,value(d))) : :(value(d)))
-            typeof(V)<:Signature ? ($p(b) ? Simplex{V}(-v,d) : isone(v) ? d : Simplex{V}(v,d)) : Simplex{V}($p(b)*v,d)
-        end
-        for B ∈ MSB
-            @eval $c(b::$B) = value(b)≠0 ? value(b)*$c(basis(b)) : g_zero(vectorspace(b))
+        @eval begin
+            @pure function $c(b::Basis{V,G,B}) where {V,G,B}
+                d = getbasis(V,complement(ndims(V),B,diffvars(V),$(c≠h ? 0 : :(hasinf(V)+hasorigin(V)))))
+                mixedmode(V)<0 && throw(error("Complement for mixed tensors is undefined"))
+                v = $(c≠h ? :($pn(V,B,value(d))) : :(value(d)))
+                typeof(V)<:Signature ? ($p(b) ? Simplex{V}(-v,d) : isone(v) ? d : Simplex{V}(v,d)) : Simplex{V}($p(b)*v,d)
+            end
+            $c(b::Simplex) = value(b)≠0 ? value(b)*$c(basis(b)) : g_zero(vectorspace(b))
         end
     end
 end
@@ -346,11 +346,11 @@ export involute
 
 for r ∈ (:reverse,:involute,:conj)
     p = Symbol(:parity,r)
-    @eval @pure function $r(b::Basis{V,G,B}) where {V,G,B}
-        $p(grade(V,B)) ? Simplex{V}(-value(b),b) : b
-    end
-    for implex ∈ MSB
-        @eval $r(b::$implex) = value(b) ≠ 0 ? value(b) * $r(basis(b)) : g_zero(vectorspace(b))
+    @eval begin
+        @pure function $r(b::Basis{V,G,B}) where {V,G,B}
+            $p(grade(V,B)) ? Simplex{V}(-value(b),b) : b
+        end
+        $r(b::Simplex) = value(b) ≠ 0 ? value(b) * $r(basis(b)) : g_zero(vectorspace(b))
     end
 end
 
@@ -402,37 +402,25 @@ function mul(a::Basis{V},b::Basis{V},der=derive_mul(V,bits(a),bits(b),1,true)) w
     return cc ? (v=value(out);out+Simplex{V}(hasinforigin(V,A,B) ? -(v) : v,getbasis(V,conformalmask(V)⊻bits(d)))) : out
 end
 
-for implex ∈ MSB
-    @eval begin
-        function *(a::$implex{V},b::Basis{V}) where V
-            v = derive_mul(V,bits(basis(a)),bits(b),a.v,true)
-            bas = mul(basis(a),b,v)
-            order(a.v)+order(bas)>diffmode(V) ? zero(V) : Simplex{V}(v,bas)
-        end
-        function *(a::Basis{V},b::$implex{V}) where V
-            v = derive_mul(V,bits(a),bits(basis(b)),b.v,false)
[…]
