### 2019-11-28 c829d3f
generalized dispatch for symbolic computation
