### 2019-11-29 2b5e836
vastly improved performance of exp, including symbolic computation
