### 2019-12-03 32c5f7f
revised SparseChain and MultiGrade types
