+++ src/multivectors.jl
@@ -2,14 +2,14 @@
 #   This file is part of Grassmann.jl. It is licensed under the GPL license
 #   Grassmann Copyright (C) 2019 Michael Reed
 
-export TensorTerm, TensorMixed, Basis, MultiVector, SparseChain, MultiGrade
+export TensorTerm, TensorGraded, TensorMixed, Basis, MultiVector, SparseChain, MultiGrade
 
-abstract type GradedAlgebra{V,G} <: TensorAlgebra{V} end
-abstract type TensorTerm{V,G} <: GradedAlgebra{V,G} end
-abstract type TensorMixed{T,V} <: TensorAlgebra{V} end
+abstract type TensorGraded{V,G} <: TensorAlgebra{V} end
+abstract type TensorTerm{V,G} <: TensorGraded{V,G} end
+abstract type TensorMixed{V} <: TensorAlgebra{V} end
 
 function Base.show_unquoted(io::IO, z::T, ::Int, prec::Int) where T<:TensorAlgebra
-    if T<:TensorMixed && Base.operator_precedence(:+) <= prec
+    if !(T<:TensorTerm) && Base.operator_precedence(:+) <= prec
         print(io, "(")
         show(io, z)
         print(io, ")")
@@ -24,9 +24,8 @@ Fields = (Real,Complex)
 
 # symbolic print types
 
-parany = (Expr,Complex,Rational,TensorMixed)
-parsym = (Expr,Complex,Rational,TensorAlgebra,Symbol)
-parval = (Expr,Complex,Rational,TensorMixed)
+parval = (Expr,Complex,Rational,TensorAlgebra)
+parsym = (Symbol,parval...)
 
 ## pseudoscalar
 
@@ -124,7 +123,7 @@ function Simplex{V,G,B}(b::T) where T<:TensorTerm{V} where {V,G,B}
 end
 function show(io::IO,m::Simplex)
     T = typeof(value(m))
-    par = !(T <: TensorTerm) && |(broadcast(<:,T,parany)...)
+    par = !(T <: TensorTerm) && |(broadcast(<:,T,parval)...)
     print(io,(par ? ['(',m.v,')'] : [m.v])...,basis(m))
 end
 for VG ∈ ((:V,),(:V,:G))
@@ -138,59 +137,58 @@ end
 
 ## Chain{T,N}
 
-@eval begin
-    @computed struct Chain{T,V,G} <: TensorMixed{T,V}
-        v::SVector{binomial(ndims(V),G),T}
-    end
-
-    export Chain
-    getindex(m::Chain,i::Int) = m.v[i]
-    getindex(m::Chain,i::UnitRange{Int}) = m.v[i]
-    setindex!(m::Chain{T},k::T,i::Int) where T = (m.v[i] = k)
-    Base.firstindex(m::Chain) = 1
-    @pure Base.lastindex(m::Chain{T,V,G}) where {T,V,G} = binomial(ndims(V),G)
-    @pure Base.length(m::Chain{T,V,G}) where {T,V,G} = binomial(ndims(V),G)
+@computed struct Chain{T,V,G} <: TensorGraded{V,G}
+    v::SVector{binomial(ndims(V),G),T}
 end
-@eval begin
-    function (m::Chain{T,V,G})(i::Integer) where {T,V,G}
-        Simplex{V,G,Basis{V}(indexbasis(ndims(V),G)[i]),T}(m[i])
+
+export Chain
+getindex(m::Chain,i::Int) = m.v[i]
+getindex(m::Chain,i::UnitRange{Int}) = m.v[i]
[…]
