+++ src/algebra.jl
@@ -153,13 +153,17 @@ function generate_mutators(M,F,set_val,SUB,MUL)
         sb = Symbol(op,:blade!)
         for (s,index) ∈ ((sm,:basisindex),(sb,:bladeindex))
             spre = Symbol(s,:_pre)
+            @eval @inline function $s(out::$M,val::S,i) where {M,T,S}
+                @inbounds $(set_val(set,:(out[i]),:val))
+                return out
+            end
             for (i,B) ∈ ((:i,Bits),(:(bits(i)),Basis))
                 @eval begin
                     @inline function $s(out::$M,val::S,i::$B) where {M,T<:$F,S<:$F}
                         @inbounds $(set_val(set,:(out[$index(intlog(M),$i)]),:val))
                         return out
                     end
-                    @inline function $s(out::Q,val::S,i::$B,::Dimension{N}) where Q<:$M where {M,T<:$F,S<:$F,N}
+                    @inline function $s(out::Q,val::S,i::$B,::Val{N}) where Q<:$M where {M,T<:$F,S<:$F,N}
                         @inbounds $(set_val(set,:(out[$index(N,$i)]),:val))
                         return out
                     end
@@ -172,7 +176,7 @@ function generate_mutators(M,F,set_val,SUB,MUL)
                         @inbounds $(pre_val(set,:(out[ind]),:val))
                         return out
                     end
-                    @inline function $spre(out::Q,val::S,i::$B,::Dimension{N}) where Q<:$M where {M,T<:$F,S<:$F,N}
+                    @inline function $spre(out::Q,val::S,i::$B,::Val{N}) where Q<:$M where {M,T<:$F,S<:$F,N}
                         ind = $index(N,$i)
                         @inbounds $(pre_val(set,:(out[ind]),:val))
                         return out
@@ -191,7 +195,7 @@ function generate_mutators(M,F,set_val,SUB,MUL)
                             !iszero(Z) && (T≠Any ? (return true) : (val *= getbasis(loworder(V),Z)))
                             count_ones(Q)+order(val)>diffmode(V) && (return false)
                         end
-                        $s(m,val,(A⊻B)|Q,Dimension{N}())
+                        $s(m,val,(A⊻B)|Q,Val{N}())
                     end
                     return false
                 end
@@ -203,7 +207,7 @@ function generate_mutators(M,F,set_val,SUB,MUL)
                             !iszero(Z) && (val = Expr(:call,:*,getbasis(loworder(V),Z)))
                             val = :(h=$val;$(count_ones(Q))+order(h)>$(diffmode(V)) ? 0 : h)
                         end
-                        $spre(m,val,(A⊻B)|Q,Dimension{N}())
+                        $spre(m,val,(A⊻B)|Q,Val{N}())
                     end
                     return false
                 end
@@ -216,8 +220,8 @@ function generate_mutators(M,F,set_val,SUB,MUL)
                             !iszero(Z) && (T≠Any ? (return true) : (val *= getbasis(loworder(V),Z)))
                             count_ones(Q)+order(val)>diffmode(V) && (return false)
                         end
-                        $s(m,val,bas|Q,Dimension{N}())
[…]
