### 2019-12-04 91d6d0d
revised MultiVector and Chain type parametrization
