+++ src/multivectors.jl
@@ -135,32 +135,32 @@ end
 ==(a::TensorTerm{V,G},b::TensorTerm{V,G}) where {V,G} = basis(a) == basis(b) ? value(a) == value(b) : 0 == value(a) == value(b)
 ==(a::TensorTerm,b::TensorTerm) = 0 == value(a) == value(b)
 
-## Chain{T,N}
+## Chain{V,G,T}
 
-@computed struct Chain{T,V,G} <: TensorGraded{V,G}
+@computed struct Chain{V,G,T} <: TensorGraded{V,G}
     v::SVector{binomial(ndims(V),G),T}
 end
 
 export Chain
 getindex(m::Chain,i::Int) = m.v[i]
 getindex(m::Chain,i::UnitRange{Int}) = m.v[i]
-setindex!(m::Chain{T},k::T,i::Int) where T = (m.v[i] = k)
+setindex!(m::Chain{V,G,T} where {V,G},k::T,i::Int) where T = (m.v[i] = k)
 Base.firstindex(m::Chain) = 1
-@pure Base.lastindex(m::Chain{T,V,G}) where {T,V,G} = binomial(ndims(V),G)
-@pure Base.length(m::Chain{T,V,G}) where {T,V,G} = binomial(ndims(V),G)
+@pure Base.lastindex(m::Chain{V,G}) where {V,G} = binomial(ndims(V),G)
+@pure Base.length(m::Chain{V,G}) where {V,G} = binomial(ndims(V),G)
 
-function (m::Chain{T,V,G})(i::Integer) where {T,V,G}
+function (m::Chain{V,G,T})(i::Integer) where {V,G,T}
     Simplex{V,G,Basis{V}(indexbasis(ndims(V),G)[i]),T}(m[i])
 end
 
-function Chain{T,V,G}(val::T,v::Basis{V,G}) where {T,V,G}
+function Chain{V,G,T}(val::T,v::Basis{V,G}) where {V,G,T}
     N = ndims(V)
-    Chain{T,V,G}(setblade!(zeros(mvec(N,G,T)),val,bits(v),Val{N}()))
+    Chain{V,G,T}(setblade!(zeros(mvec(N,G,T)),val,bits(v),Val{N}()))
 end
 
-Chain(v::Basis{V,G}) where {V,G} = Chain{Int,V,G}(one(Int),v)
+Chain(v::Basis{V,G}) where {V,G} = Chain{V,G,Int}(one(Int),v)
 
-function show(io::IO, m::Chain{T,V,G}) where {T,V,G}
+function show(io::IO, m::Chain{V,G,T}) where {V,G,T}
     ib = indexbasis(ndims(V),G)
     @inbounds if T == Any && typeof(m.v[1]) ∈ parsym
         @inbounds tmv = typeof(m.v[1])
@@ -183,97 +183,98 @@ function show(io::IO, m::Chain{T,V,G}) where {T,V,G}
         @inbounds DirectSum.printindices(io,V,ib[k])
     end
 end
-function ==(a::Chain{S,V,G} where S,b::T) where T<:TensorTerm{V,G} where {V,G}
+
+function ==(a::Chain{V,G},b::T) where T<:TensorTerm{V,G} where {V,G}
     i = bladeindex(ndims(V),bits(basis(b)))
     @inbounds a[i] == value(b) && (isempty(a[1:i-1]) ? true : (prod(a[1:i-1].==0) && prod(a[i+1:end].==0)))
 end
-==(a::T,b::Chain{S,V} where S) where T<:TensorTerm{V} where V = b==a
-==(a::Chain{S,V} where S,b::T) where T<:TensorTerm{V} where V = prod(0==value(b).==value(a))
+==(a::T,b::Chain{V}) where T<:TensorTerm{V} where V = b==a
[…]
