+++ src/algebra.jl
@@ -423,8 +423,8 @@ end
 
 *(a::UniformScaling,b::Simplex{V}) where V = V(a)*b
 *(a::Simplex{V},b::UniformScaling) where V = a*V(b)
-*(a::UniformScaling,b::Chain{T,V} where T) where V = V(a)*b
-*(a::Chain{T,V} where T,b::UniformScaling) where V = a*V(b)
+*(a::UniformScaling,b::Chain{V}) where V = V(a)*b
+*(a::Chain{V},b::UniformScaling) where V = a*V(b)
 
 export ∗, ⊛, ⊖
 const ⊖ = *
@@ -711,7 +711,7 @@ function generate_products(Field=Field,VEC=:mvec,MUL=:*,ADD=:+,SUB=:-,CONJ=:conj
     TF = Field ∉ Fields ? :Any : :T
     EF = Field ≠ Any ? Field : ExprField
     @eval begin
-        @generated function adjoint(m::MultiVector{T,V}) where {T<:$Field,V}
+        @generated function adjoint(m::MultiVector{V,T}) where {V,T<:$Field}
             if ndims(V)<cache_limit
                 if mixedmode(V)<0
                     $(insert_expr((:N,:M,:bs,:bn),:vec)...)
@@ -739,20 +739,20 @@ function generate_products(Field=Field,VEC=:mvec,MUL=:*,ADD=:+,SUB=:-,CONJ=:conj
                 else
                     out = $$CONJ.(value(m))
                 end
-                MultiVector{$$TF,dual(V)}(out)
+                MultiVector{dual(V),$$TF}(out)
             end end
         end
         *(a::F,b::Basis{V}) where {F<:$EF,V} = Simplex{V}(a,b)
         *(a::Basis{V},b::F) where {F<:$EF,V} = Simplex{V}(b,a)
-        *(a::F,b::MultiVector{T,V}) where {F<:$Field,T,V} = MultiVector{promote_type(T,F),V}(broadcast($Sym.∏,Ref(a),b.v))
-        *(a::MultiVector{T,V},b::F) where {F<:$Field,T,V} = MultiVector{promote_type(T,F),V}(broadcast($Sym.∏,a.v,Ref(b)))
+        *(a::F,b::MultiVector{V,T}) where {F<:$Field,T,V} = MultiVector{V,promote_type(T,F)}(broadcast($Sym.∏,Ref(a),b.v))
+        *(a::MultiVector{V,T},b::F) where {F<:$Field,T,V} = MultiVector{V,promote_type(T,F)}(broadcast($Sym.∏,a.v,Ref(b)))
         *(a::F,b::MultiGrade{V,G}) where {F<:$EF,V,G} = MultiGrade{V,G}(broadcast($MUL,Ref(a),b.v))
         *(a::MultiGrade{V,G},b::F) where {F<:$EF,V,G} = MultiGrade{V,G}(broadcast($MUL,a.v,Ref(b)))
         ∧(a::$Field,b::$Field) = $MUL(a,b)
         ∧(a::F,b::B) where B<:TensorTerm{V,G} where {F<:$EF,V,G} = Simplex{V,G}(a,b)
         ∧(a::A,b::F) where A<:TensorTerm{V,G} where {F<:$EF,V,G} = Simplex{V,G}(b,a)
-        #=∧(a::$Field,b::MultiVector{T,V}) where {T<:$Field,V} = MultiVector{T,V}(a.*b.v)
-        ∧(a::MultiVector{T,V},b::$Field) where {T<:$Field,V} = MultiVector{T,V}(a.v.*b)
[…]
