### 2019-12-06 c2c6bfa
improved exp methods and dispatch
