+++ docs/src/library.md
@@ -1,0 +1,39 @@
+# Grassmann.jl Library
+
+```@contents
+Pages = ["index.md","library.md"]
+```
+
+## Index
+
+```@index
+Pages = ["library.md"]
+```
+
+## Grassmann Interface
+
+```@docs
+Grassmann.complementrighthodge
+```
+
+```@docs
+Grassmann.complementlefthodge
+```
+
+```@docs
+Grassmann.reverse
+```
+
+```@docs
+Grassmann.involute
+```
+
+```@docs
+Grassmann.:∧
+```
+
+```@docs
+Grassmann.:∨
+```
+
+etc, ...
