### 2019-12-07 22ed06f
fixed deploydocs initialization script
