+++ dev/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"library/#Grassmann.jl-Library-1","page":"Library","title":"Grassmann.jl Library","text":"","category":"section"},{"location":"library/#","page":"Library","title":"Library","text":"Pages = [\"index.md\",\"library.md\"]","category":"page"},{"location":"library/#Index-1","page":"Library","title":"Index","text":"","category":"section"},{"location":"library/#","page":"Library","title":"Library","text":"Pages = [\"library.md\"]","category":"page"},{"location":"library/#Grassmann-Interface-1","page":"Library","title":"Grassmann Interface","text":"","category":"section"},{"location":"library/#","page":"Library","title":"Library","text":"Grassmann.complementrighthodge","category":"page"},{"location":"library/#Grassmann.complementrighthodge","page":"Library","title":"Grassmann.complementrighthodge","text":"complementrighthodge(ω::TensorAlgebra)\n\nGrassmann-Poincare-Hodge complement: ⋆ω = ω∗I\n\n\n\n\n\n","category":"function"},{"location":"library/#","page":"Library","title":"Library","text":"Grassmann.complementlefthodge","category":"page"},{"location":"library/#Grassmann.complementlefthodge","page":"Library","title":"Grassmann.complementlefthodge","text":"complementlefthodge(ω::TensorAlgebra)\n\nGrassmann-Poincare left complement: ⋆'ω = I∗'ω\n\n\n\n\n\n","category":"function"},{"location":"library/#","page":"Library","title":"Library","text":"Grassmann.reverse","category":"page"},{"location":"library/#Base.reverse","page":"Library","title":"Base.reverse","text":"~(ω::TensorAlgebra)\n\nReverse of a MultiVector element: ~ω = (-1)^(grade(ω)(grade(ω)-1)/2)ω\n\n\n\n\n\n","category":"function"},{"location":"library/#","page":"Library","title":"Library","text":"Grassmann.involute","category":"page"},{"location":"library/#AbstractTensors.involute","page":"Library","title":"AbstractTensors.involute","text":"involute(ω::TensorAlgebra)\n\nInvolute of a MultiVector element: ~ω = (-1)^grade(ω)*ω\n\n\n\n\n\n","category":"function"},{"location":"library/#","page":"Library","title":"Library","text":"Grassmann.:∧","category":"page"},{"location":"library/#AbstractLattices.:∧","page":"Library","title":"AbstractLattices.:∧","text":"∧(ω::TensorAlgebra,η::TensorAlgebra)\n\nExterior product as defined by the anti-symmetric quotient Λ≡⊗/~\n\n\n\n\n\n","category":"function"},{"location":"library/#","page":"Library","title":"Library","text":"Grassmann.:∨","category":"page"},{"location":"library/#AbstractLattices.:∨","page":"Library","title":"AbstractLattices.:∨","text":"∨(ω::TensorAlgebra,η::TensorAlgebra)\n\nRegressive product as defined by the DeMorgan's law: ∨(ω...) = ⋆⁻¹(∧(⋆.(ω)...))\n\n\n\n\n\n","category":"function"},{"location":"library/#","page":"Library","title":"Library","text":"etc, ...","category":"page"},{"location":"#Grassmann.jl-1","page":"Home","title":"Grassmann.jl","text":"","category":"section"},{"location":"#","page":"Home","title":"Home","text":"⟨Leibniz-Grassmann-Clifford-Hestenes⟩ differential geometric algebra / multivector simplicial complex","category":"page"},{"location":"#","page":"Home","title":"Home","text":"(Image: Build Status) (Image: Build status) (Image: Coverage Status) (Image: codecov.io)
[…]
