### 2019-12-07 50fbe9a
added tutorials section and revised README links for docs
