### 2019-12-07 4d98245
revised homepage links
