### 2019-12-07 f5666d8
refined presentation of homepage
