### 2019-12-07 18a942d
fixed homepage README configuation
