### 2019-12-07 a3257a2
revised homepage grammar
