### 2019-12-08 955a28c
revised markdown badges for homepage
