### 2019-12-08 6a932af
revised links for markdown badges
