+++ README.md
@@ -1,11 +1,11 @@
+[![GitHub stars](https://img.shields.io/github/stars/chakravala/Grassmann.jl?style=social)](https://github.com/chakravala/Grassmann.jl/stargazers)
 *⟨Leibniz-Grassmann-Clifford-Hestenes⟩ differential geometric algebra / multivector simplicial complex*
-![GitHub stars](https://img.shields.io/github/stars/chakravala/Grassmann.jl?style=social)
 
 <p align="center">
   <img src="./dev/assets/logo.png" alt="Grassmann.jl"/>
 </p>
 
-![GitHub release (latest SemVer)](https://img.shields.io/github/v/release/chakravala/Grassmann.jl)
+[![GitHub release (latest SemVer)](https://img.shields.io/github/v/release/chakravala/Grassmann.jl)](https://github.com/chakravala/Grassmann.jl/releases)
 [![YouTube](https://img.shields.io/badge/JuliaCon%202019-YouTube-red)](https://www.youtube.com/watch?v=eQjDN0JQ6-s)
 [![DropBox](https://img.shields.io/badge/download_PDF-DropBox-blue.svg)](https://www.dropbox.com/sh/tphh6anw0qwija4/AAACiaXig5djrLVAKLPFmGV-a/Geometric-Algebra?preview=grassmann-juliacon-2019.pdf)
 [![Docs Stable](https://img.shields.io/badge/docs-stable-blue.svg)](https://grassmann.crucialflow.com/stable)
@@ -19,7 +19,7 @@ The kernelized operations are built up from composite sparse tensor products and
 The [DirectSum.jl](https://github.com/chakravala/DirectSum.jl) multivector parametric type polymorphism is based on tangent bundle vector spaces and conformal projective geometry to make the dispatch highly extensible for many applications.
 Additionally, the universal interoperability between different sub-algebras is enabled by [AbstractTensors.jl](https://github.com/chakravala/AbstractTensors.jl), on which the type system is built.
 
-![GitHub commits since latest release](https://img.shields.io/github/commits-since/chakravala/Grassmann.jl/latest?label=new%20commits)
+[![GitHub commits since latest release](https://img.shields.io/github/commits-since/chakravala/Grassmann.jl/latest?label=new%20commits)](https://github.com/chakravala/Grassmann.jl/commits)
 [![Build Status](https://travis-ci.org/chakravala/Grassmann.jl.svg?branch=master)](https://travis-ci.org/chakravala/Grassmann.jl)
 [![Build status](https://ci.appveyor.com/api/projects/status/c36u0rgtm2rjcquk?svg=true)](https://ci.appveyor.com/project/chakravala/grassmann-jl)
 [![Coverage Status](https://coveralls.io/repos/chakravala/Grassmann.jl/badge.svg?branch=master&service=github)](https://coveralls.io/github/chakravala/Grassmann.jl?branch=master)
