### 2019-12-08 0a4d3b5
fixed heading and title of tutorial section
