+++ docs/src/tutorials/quick-start.md
@@ -1,3 +1,3 @@
-## Quick start
+# Quick start
 
 Add some quick start tutorials here.
