+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"tutorials/install/#Installation-1","page":"-","title":"Installation","text":"","category":"section"},{"location":"tutorials/install/#","page":"-","title":"-","text":"Availability of this package and its subpackages can be automatically handled with the Julia package manager with using Pkg and Pkg.add(\"Grassmann\") or by entering:","category":"page"},{"location":"tutorials/install/#","page":"-","title":"-","text":"pkg> add Grassmann","category":"page"},{"location":"tutorials/install/#Requirements-1","page":"-","title":"Requirements","text":"","category":"section"},{"location":"tutorials/install/#","page":"-","title":"-","text":"When the master branch is used it is possible that some of the dependencies also require a development branch before the release. This may include (but is not limited to) the following packages:","category":"page"},{"location":"tutorials/install/#","page":"-","title":"-","text":"This requires a merged version of ComputedFieldTypes at https://github.com/vtjnash/ComputedFieldTypes.jl","category":"page"},{"location":"tutorials/install/#","page":"-","title":"-","text":"Interoperability of TensorAlgebra with other packages is automatically enabled by DirectSum.jl and AbstractTensors.jl.","category":"page"},{"location":"tutorials/install/#","page":"-","title":"-","text":"The package is compatible via Requires.jl with  Reduce.jl, SymPy.jl, SymEngine.jl, AbstractAlgebra.jl, Nemo.jl, GaloisFields.jl, LightGraphs,jl, Compose.jl, GeometryTypes,jl, Makie.jl.","category":"page"},{"location":"library/#Grassmann.jl-Library-1","page":"Library","title":"Grassmann.jl Library","text":"","category":"section"},{"location":"library/#","page":"Library","title":"Library","text":"Pages = [\"index.md\",\"library.md\"]","category":"page"},{"location":"library/#Index-1","page":"Library","title":"Index","text":"","category":"section"},{"location":"library/#","page":"Library","title":"Library","text":"Pages = [\"library.md\"]","category":"page"},{"location":"library/#Grassmann-Interface-1","page":"Library","title":"Grassmann Interface","text":"","category":"section"},{"location":"library/#","page":"Library","title":"Library","text":"Grassmann.complementrighthodge","category":"page"},{"location":"library/#Grassmann.complementrighthodge","page":"Library","title":"Grassmann.complementrighthodge","text":"complementrighthodge(ω::TensorAlgebra)\n\nGrassmann-Poincare-Hodge complement: ⋆ω = ω∗I\n\n\n\n\n\n","category":"function"},{"location":"library/#","page":"Library","title":"Library","text":"Grassmann.complementlefthodge","category":"page"},{"location":"library/#Grassmann.complementlefthodge","page":"Library","title":"Grassmann.complementlefthodge","text":"complementlefthodge(ω::TensorAlgebra)\n\nGrassmann-Poincare left complement: ⋆'ω = I∗'ω\n\n\n\n\n\n","category":"function"},{"location":"library/#","page":"Library","title":"Library","text":"Grassmann.reverse","category":"page"},{"location":"library/#Base.reverse","page":"Library","title":"Base.reverse","text":"~(ω::TensorAlgebra)\n\nReverse of a MultiVector element: ~ω = (-1)^(grade(ω)(grade(ω)-1)/2)ω\n\n\n\n\n\n","category":"function"},{"location":"library/#","page":"Library","title":"Library","text":"Grassmann.involute","category":"page"},{"location":"library/#AbstractTensors.involute","page":"Library","title":"AbstractTensors.involute","text":"involute(ω::TensorAlgebra)\n\nInvol
[…]
