### 2019-12-13 4becffe
transitioned documentation material into tutorials
