### 2019-12-13 613e097
revised homepage documentation page
