### 2019-12-14 673566b
initialized algebra-of-space tutorial
