+++ docs/src/tutorials/algebra-of-space.md
@@ -1,0 +1,214 @@
+# The Algebra of Space (G3)
+
+This notebook is an adaptation from the [clifford](https://clifford.readthedocs.io/en/latest/TheAlgebraOfSpaceG3.html) python documentation.
+
+Import `Grassmann` and instantiate a three dimensional geometric algebra
+
+```Julia
+julia> using Grassmann
+
+julia> basis"3"
+(⟨+++⟩, v, v₁, v₂, v₃, v₁₂, v₁₃, v₂₃, v₁₂₃)
+```
+
+Given a three dimensional GA with the orthonormal basis ``v_i\cdot v_j = \delta_{ij}``, the basis consists of scalars, three vectors, three bivectors, and a trivector.
+```math
+\{\underbrace{v}_{\mbox{scalar}},\qquad\underbrace{v_{1},v_{2},v_{3}}_{\mbox{vectors}},\qquad\underbrace{v_{12},v_{23},v_{13}}_{\mbox{bivectors}},\qquad\underbrace{v_{123}}_{\mbox{trivector}}\}
+```
+The `@basis` macro declares the algebra and assigns the `Basis` elements to local variables. The `Grassmann.Algebra` can also be assigned to `G3` as
+```Julia
+julia> G3 = Λ(3)
+Grassmann.Algebra{⟨+++⟩,8}(v, v₁, v₂, v₃, v₁₂, v₁₃, v₂₃, v₁₂₃)
+```
+You may wish to explicitly assign the blades to variables like so,
+```Julia
+e1 = G3.v1
+e2 = G3.v2
+# etc ...
+```
+Or, if you're lazy you can use the macro with different local names
+```Julia
+julia> @basis ℝ^3 E e
+(⟨+++⟩, v, v₁, v₂, v₃, v₁₂, v₁₃, v₂₃, v₁₂₃)
+
+julia> e3, e123
+(v₃, v₁₂₃)
+```
+
+## Basics
+
+The basic products are available
+
+```Julia
+julia> v1*v2 # geometric product
+v₁₂
+
+julia> v1|v2 # inner product
+0v
+
+julia> v1 ∧ v2 # exterior product
+v₁₂
+
+julia> v1 ∧ v2 ∧ v3 # even more exterior products
+v₁₂₃
+```
+
+Multivectors can be defined in terms of the basis blades. For example, you can construct a rotor as a sum of a scalar and a bivector, like so
+```Julia
+julia> θ = π/4
+0.7853981633974483
+
+julia> R = cos(θ) - sin(θ)*v23
+0.7071067811865476 - 0.7071067811865475v₂₃
+
+julia> R = ~exp(θ*v23)
+0.7071067811865476 - 0.7071067811865475v₂₃
+```
+You can also mix grades without any reason
+```Julia
+julia> A = 1 + 2v1 + 3v12 + 4v123
+1 + 2v₁ + 3v₁₂ + 4v₁₂₃
+```
+The reversion operator is accomplished with the tilde `~` in front of the `MultiVector` on which it acts
+```Julia
+julia> ~A
+1 + 2v₁ - 3v₁₂ - 4v₁₂₃
+```
+Taking a projection into a specific `grade` of a `MultiVector` is usually written ``\langle A\rangle_n`` and can be done using the soft brackets, like so
+```Julia
+julia> A(0)
+1v
+
+julia> A(1)
+2v₁ + 0v₂ + 0v₃
+
+julia> A(2)
+3v₁₂ + 0v₁₃ + 0v₂₃
+```
+Using the reversion and grade projection operators, we can define the magnitude of `A` as ``|A|^2 = \langle{\tilde A A\rangle``
+```Julia
+julia> ~A*A
[…]
