+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"references/#References-1","page":"References","title":"References","text":"","category":"section"},{"location":"references/#","page":"References","title":"References","text":"(Image: YouTube) (Image: DropBox) (Image: Docs Stable) (Image: Docs Dev) (Image: Gitter) (Image: BiVector)","category":"page"},{"location":"references/#","page":"References","title":"References","text":"Emil Artin, Geometric Algebra. 1957.\nJohn Browne, Grassmann Algebra, Volume 1: Foundations. 2011.\nC. Doran, D. Hestenes, F. Sommen, and N. Van Acker, Lie groups as spin groups, J. Math Phys. (1993)\nDavid Hestenes, Universal Geometric Algebra, Pure and Applied (1988)\nDavid Hestenes, Tutorial on geometric calculus. Advances in Applied Clifford Algebra, 2013.\nLachlan Gunn, Derek Abbott, James Chappell, Ashar Iqbal, Functions of multivector variables. 2011.\nAaron D. Schutte, A nilpotent algebra approach to Lagrangian mechanics and constrained motion. 2016.\nVladimir and Tijana Ivancevic, Undergraduate lecture notes in DeRahm-Hodge theory. arXiv, 2011.\nPeter Woit, Clifford algebras and spin groups, Lecture Notes (2012)","category":"page"},{"location":"tutorials/install/#Installation-1","page":"Installation","title":"Installation","text":"","category":"section"},{"location":"tutorials/install/#","page":"Installation","title":"Installation","text":"Availability of this package and its subpackages can be automatically handled with the Julia package manager with using Pkg and Pkg.add(\"Grassmann\") or by entering:","category":"page"},{"location":"tutorials/install/#","page":"Installation","title":"Installation","text":"pkg> add Grassmann","category":"page"},{"location":"tutorials/install/#Requirements-1","page":"Installation","title":"Requirements","text":"","category":"section"},{"location":"tutorials/install/#","page":"Installation","title":"Installation","text":"When the master branch is used it is possible that some of the dependencies also require a development branch before the release. This may include (but is not limited to) the following packages:","category":"page"},{"location":"tutorials/install/#","page":"Installation","title":"Installation","text":"This requires a merged version of ComputedFieldTypes at ComputedFieldTypes.jl.","category":"page"},{"location":"tutorials/install/#","page":"Installation","title":"Installation","text":"Interoperability of TensorAlgebra with other packages is automatically enabled by DirectSum.jl and AbstractTensors.jl.","category":"page"},{"location":"tutorials/install/#","page":"Installation","title":"Installation","text":"The package is compatible via Requires.jl with  Reduce.jl, SymPy.jl, SymEngine.jl, AbstractAlgebra.jl, Nemo.jl, GaloisFields.jl, LightGraphs,jl, Compose.jl, GeometryTypes,jl, Makie.jl.","category":"page"},{"location":"design/#Grassmann.jl-design-and-code-generation-1","page":"Design","title":"Grassmann.jl design and code generation","text":"","category":"section"},{"location":"design/#","page":"Design","title":"Design","text":"Mathematical foundations and definitions specific to the Grassmann.jl implementation provide an extensible platform for computing with geometric algebra at high dimensions, along with the accompanying support packages.  The design is based on the TensorAlgebra abstract type interoperability from AbstractTensors.jl with a VectorBundle parameter from DirectSum.jl. Abstract tangent vector space type operations happen at compile-time, resulting in a differential conformal geometric algebra of hyper-dual multivector forms.","category":"page"},{"location":"design/#","page":"Design","title":"Design","text":"The nature of the geometric algebra code generation enables
[…]
