### 2019-12-14 0b4b83c
reorganized documentation library
