+++ docs/make.jl
@@ -1,11 +1,11 @@
 #   This file is part of Grassmann.jl. It is licensed under the GPL license
 #   Grassmann Copyright (C) 2019 Michael Reed
 
-using Documenter, Grassmann
+using Documenter, DirectSum, AbstractTensors, Grassmann
 
 makedocs(
     # options
-    modules = [Grassmann],
+    modules = [DirectSum,AbstractTensors,Grassmann],
     doctest = false,
     format = Documenter.HTML(prettyurls = get(ENV, "CI", nothing) == "true"),
     sitename = "Grassmann.jl",
@@ -19,7 +19,7 @@ makedocs(
             "tutorials/install.md",
             "tutorials/quick-start.md",
             "tutorials/algebra-of-space.md",
-            "tutorials/forms.md"
+            "tutorials/mixed-tensors.md"
             ],
         "References" => "references.md"
         ]
