### 2019-12-14 4e78d72
revised homepage text styling
