### 2019-12-15 39bedbf
fixed chain contraction product of tangent algebra
