+++ src/algebra.jl
@@ -292,13 +292,15 @@ function generate_mutators(M,F,set_val,SUB,MUL)
     end
 end
 
-@inline exteraddmulti!(V::W,out,α,β,γ) where W<:Manifold = (count_ones(α&β)==0) && joinaddmulti!(V,out,α,β,γ)
+@inline exterbits(V,α,β) = diffvars(V)≠0 ? ((a,b)=symmetricmask(V,α,β);count_ones(a&b)==0) : count_ones(α&β)==0
 
-@inline outeraddblade!(V::W,out,α,β,γ) where W<:Manifold = (count_ones(α&β)==0) && joinaddblade!(V,out,α,β,γ)
+@inline exteraddmulti!(V::W,out,α,β,γ) where W<:Manifold = exterbits(V,α,β) && joinaddmulti!(V,out,α,β,γ)
 
-@inline exteraddmulti!_pre(V::W,out,α,β,γ) where W<:Manifold = (count_ones(α&β)==0) && joinaddmulti!_pre(V,out,α,β,γ)
+@inline outeraddblade!(V::W,out,α,β,γ) where W<:Manifold = exterbits(V,α,β) && joinaddblade!(V,out,α,β,γ)
 
-@inline outeraddblade!_pre(V::W,out,α,β,γ) where W<:Manifold = (count_ones(α&β)==0) && joinaddblade!_pre(V,out,α,β,γ)
+@inline exteraddmulti!_pre(V::W,out,α,β,γ) where W<:Manifold = exterbits(V,α,β) && joinaddmulti!_pre(V,out,α,β,γ)
+
+@inline outeraddblade!_pre(V::W,out,α,β,γ) where W<:Manifold = exterbits(V,α,β) && joinaddblade!_pre(V,out,α,β,γ)
 
 
 # Hodge star ★
@@ -793,273 +795,388 @@ function generate_products(Field=Field,VEC=:mvec,MUL=:*,ADD=:+,SUB=:-,CONJ=:conj
         #∧(a::$Field,b::Chain{V,G,T}) where {V,G,T<:$Field} = Chain{V,G,T}(a.*b.v)
         #∧(a::Chain{V,G,T},b::$Field) where {V,G,T<:$Field} = Chain{V,G,T}(a.v.*b)
         @generated function contraction(a::Chain{V,G,T},b::Basis{V,L}) where {V,G,T<:$Field,L}
-            G<L && (return g_zero(V))
+            G<L && diffvars(V)==0 && (return g_zero(V))
             if binomial(ndims(V),G)<(1<<cache_limit)
-                $(insert_expr((:N,:t,:ib,:bng),:svec)...)
-                out = zeros(svec(N,G-L,Any))
+                $(insert_expr((:N,:t,:ib,:bng,:μ),:svec)...)
+                out = zeros(μ ? svec(N,Any) : svec(N,G-L,Any))
                 for i ∈ 1:bng
-                    @inbounds skewaddblade!_pre(V,out,ib[i],bits(b),derive_pre(V,ib[i],bits(b),:(a[$i]),true))
+                    if μ
+                        @inbounds skewaddmulti!_pre(V,out,ib[i],bits(b),derive_pre(V,ib[i],bits(b),:(a[$i]),true))
+                    else
+                        @inbounds skewaddblade!_pre(V,out,ib[i],bits(b),derive_pre(V,ib[i],bits(b),:(a[$i]),true))
+                    end
                 end
                 #return :(value_diff(Simplex{V,0,$(getbasis(V,0))}($(value(mv)))))
-                return :(Chain{$V,G-L,promote_type(T,valuetype(b))}($(Expr(:call,:SVector,out...))))
+                return if μ
+                    :(MultiVector{$V}($(Expr(:call,:SVector,out...))))
[…]
