+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"references/#References-1","page":"References","title":"References","text":"","category":"section"},{"location":"references/#","page":"References","title":"References","text":"(Image: YouTube) (Image: DropBox) (Image: Docs Stable) (Image: Docs Dev) (Image: Gitter) (Image: BiVector)","category":"page"},{"location":"references/#","page":"References","title":"References","text":"Michael Reed, Differential geometric algebra with Leibniz and Grassmann. 2019.\nEmil Artin, Geometric Algebra. 1957.\nJohn Browne, Grassmann Algebra, Volume 1: Foundations. 2011.\nC. Doran, D. Hestenes, F. Sommen, and N. Van Acker, Lie groups as spin groups, J. Math Phys. (1993)\nDavid Hestenes, Universal Geometric Algebra, Pure and Applied (1988)\nDavid Hestenes, Tutorial on geometric calculus. Advances in Applied Clifford Algebra, 2013.\nLachlan Gunn, Derek Abbott, James Chappell, Ashar Iqbal, Functions of multivector variables. 2011.\nAaron D. Schutte, A nilpotent algebra approach to Lagrangian mechanics and constrained motion. 2016.\nVladimir and Tijana Ivancevic, Undergraduate lecture notes in DeRahm-Hodge theory. arXiv, 2011.\nPeter Woit, Clifford algebras and spin groups, Lecture Notes (2012)","category":"page"},{"location":"tutorials/install/#Installation-1","page":"Installation","title":"Installation","text":"","category":"section"},{"location":"tutorials/install/#","page":"Installation","title":"Installation","text":"Grassmann.jl is a package for the Julia language, which should be installed from website or the recommended method for your operating system (GNU/Linux/Mac/Windows).","category":"page"},{"location":"tutorials/install/#","page":"Installation","title":"Installation","text":"Availability of this package and its subpackages can be automatically handled with the Julia package manager with using Pkg and Pkg.add(\"Grassmann\") or by entering:","category":"page"},{"location":"tutorials/install/#","page":"Installation","title":"Installation","text":"pkg> add Grassmann","category":"page"},{"location":"tutorials/install/#","page":"Installation","title":"Installation","text":"If you would like to keep up to date with the latest commits, instead use","category":"page"},{"location":"tutorials/install/#","page":"Installation","title":"Installation","text":"pkg> add Grassmann#master","category":"page"},{"location":"tutorials/install/#","page":"Installation","title":"Installation","text":"which is not recommended if you want to use a stable release.","category":"page"},{"location":"tutorials/install/#Requirements-1","page":"Installation","title":"Requirements","text":"","category":"section"},{"location":"tutorials/install/#","page":"Installation","title":"Installation","text":"When the master branch is used it is possible that some of the dependencies also require a development branch before the release.
[…]
