### 2019-12-15 84d70c9
revised homepage installation instruction requirements
