### 2019-12-16 b8f87c0
revised installation requirement paragraph
