### 2019-12-17 91c947d
greatly extended the Algebra documentation
