+++ docs/src/tutorials/install.md
@@ -2,7 +2,7 @@
 
 *Grassmann.jl* is a package for the [Julia language](https://julialang.org), which can be obtained from their website or the recommended method for your operating system (GNU/Linux/Mac/Windows). Go to [docs.julialang.org](https://docs.julialang.org) for documentation.
 
-Availability of this package and its subpackages can be automatically handled with the Julia package manager `using Pkg` and `Pkg.add("Grassmann")` or by entering:
+Availability of this package and its subpackages is automatically handled with Julia's package manager `using Pkg` and `Pkg.add("Grassmann")` or by entering into `]` mode:
 ```Julia
 pkg> add Grassmann
 ```
@@ -31,3 +31,9 @@ The package is compatible via [Requires.jl](https://github.com/MikeInnes/Require
 [Compose.jl](https://github.com/GiovineItalia/Compose.jl),
 [GeometryTypes,jl](https://github.com/JuliaGeometry/GeometryTypes.jl),
 [Makie.jl](https://github.com/JuliaPlots/Makie.jl).
+
+## Grassmann for enterprise
+
+Sponsor this at [liberapay](https://liberapay.com/chakravala), [GitHub Sponsors](https://github.com/sponsors/chakravala), [Patreon](https://patreon.com/dreamscatter), or [Bandcamp](https://music.crucialflow.com); also available as part of the [Tidelift](https://tidelift.com/funding/github/julia/Grassmann) Subscription:
+
+The maintainers of Grassmann and thousands of other packages are working with Tidelift to deliver commercial support and maintenance for the open source dependencies you use to build your applications. Save time, reduce risk, and improve code health, while paying the maintainers of the exact dependencies you use. [Learn more.](https://tidelift.com/subscription/pkg/julia-grassmann?utm_source=julia-grassmann&utm_medium=referral&utm_campaign=enterprise&utm_term=repo)
