### 2019-12-19 b8f6d88
fixed documentation/issue with exterior differential
