+++ docs/src/algebra.md
@@ -1,6 +1,6 @@
 # Grassmann elements and geometric algebra Λ(V)
 
-The Grassmann `Basis` elements ``v_k\in\Lambda_1V`` and ``w^k\in\Lambda^1V`` are linearly independent vector and covector elements of ``V``, while the Leibniz `Operator` elements ``\partial_k\in L_1V`` are partial tangent derivations and ``\epsilon_k\in L^1V`` are dependent functions of the `tangent` manifold.
+The Grassmann `Basis` elements ``v_k\in\Lambda^1V`` and ``w^k\in\Lambda^1V`` are linearly independent vector and covector elements of ``V``, while the Leibniz `Operator` elements ``\partial_k\in L^1V`` are partial tangent derivations and ``\epsilon_k\in L^1V`` are dependent functions of the `tangent` manifold.
 Let ``V\in\text{Vect}_{\mathbb k}`` be a `VectorBundle` with dual space ``V'`` and the basis elements ``w_k:V\rightarrow\mathbb K``, then for all ``x\in V,c\in\mathbb K`` it holds: ``(w^i+w^j)(x) = w^i(x)+w^j(x)`` and ``(cw^k)(x) = cw^k(x)`` hold.
 An element of a mixed-symmetry `TensorAlgebra{V}` is a multilinear mapping that is formally constructed by taking the tensor products of linear and multilinear maps,
 ``(\bigotimes_k \omega_k)(v_1,\dots,v_{\sum_k p_k}) = \prod_k \omega_k(v_1,\dots,v_{p_k})``.
@@ -36,7 +36,7 @@ Combined, the mixed-symmetry algebra yield a multi-linear propositional lattice.
 The formal sum of equal `grade` elements is an oriented `Chain` and with mixed `grade` it is a `MultiVector` simplicial complex.
 Thus, various standard operations on the oriented multi-sets are possible including ``\cup,\cap,\oplus`` and the index operation ``\ominus``, which is symmetric difference operation.
 
-By virtue of Julia's multiple dispatch on the field type ``T``, methods can specialize on the dimension ``N`` and grade ``G`` with a `VectorBundle{N}` via the `TensorAlgebra{V}` subtypes, such as `Basis{V,G}`, `Simplex{V,G,B,T}`, `Chain{V,G,T}`, `SparseChain{V,G,T}`, `MultiVector{V,T}`, and `MultiGrade{V,G}` types.
+By virtue of Julia's multiple dispatch on the field type ``\mathbb K``, methods can specialize on the dimension ``n`` and grade ``G`` with a `VectorBundle{n}` via the `TensorAlgebra{V}` subtypes, such as `Basis{V,G}`, `Simplex{V,G,B,𝕂}`, `Chain{V,G,𝕂}`, `SparseChain{V,G,𝕂}`, `MultiVector{V,𝕂}`, and `MultiGrade{V,G}` types.
 
 The elements of the `Algebra` can be generated in many ways using the `Basis` elements created by the `@basis` macro,
 ```@repl ga
@@ -204,7 +204,9 @@ In topology there is *boundary* operator ``\partial`` defined by ``\partial\epsi
 
 Vorticity curl of vector-field:
 ``\star d(dx_1+dx_2+dx_3) = (∂_2 -∂_3)dx_1 + (∂_3 -∂_1)dx_2 + (∂_1 -∂_2)dx_3``.
-
+```@repl ga
+@basis tangent(ℝ^3,2,3); ⋆d(v1+v2+v3)
+```
[…]
