### 2019-12-19 3829229
refined homepage documentation
