### 2019-12-20 4c9e741
extended Grassmann.Algebra documentation
