+++ src/multivectors.jl
@@ -33,8 +33,13 @@ import LinearAlgebra
 import LinearAlgebra: I, UniformScaling
 export UniformScaling, I
 
-## MultiBasis{N}
+## Basis{V,G,B}
 
+"""
+    Basis{V,G,B} <: TensorTerm{V,G} <: TensorGraded{V,G} <: TensorAlgebra{V}
+
+Basis type with pseudoscalar `V::Manifold`, grade/rank `G::Int`, bits `B::UInt64`.
+"""
 struct Basis{V,G,B} <: TensorTerm{V,G}
     @pure Basis{V,G,B}() where {V,G,B} = new{V,G,B}()
 end
@@ -91,8 +96,13 @@ end
 
 @inline show(io::IO, e::Basis) = DirectSum.printindices(io,vectorspace(e),bits(e))
 
-## Simplex{N}
+## Simplex{V,G,B,𝕂}
 
+"""
+    Simplex{V,G,B,𝕂} <: TensorTerm{V,G} <: TensorGraded{V,G} <: TensorAlgebra{V}
+
+Simplex type with pseudoscalar `V::Manifold`, grade/rank `G::Int`, basis `B::Basis{V,G}`, and the scalar field `𝕂::Type`.
+"""
 struct Simplex{V,G,B,T} <: TensorTerm{V,G}
     v::T
     Simplex{A,B,C,D}(t::E) where E<:D where {A,B,C,D} = new{A,B,C,D}(t)
@@ -135,11 +145,17 @@ end
 ==(a::TensorTerm{V,G},b::TensorTerm{V,G}) where {V,G} = basis(a) == basis(b) ? value(a) == value(b) : 0 == value(a) == value(b)
 ==(a::TensorTerm,b::TensorTerm) = 0 == value(a) == value(b)
 
-## Chain{V,G,T}
+## Chain{V,G,𝕂}
 
 @computed struct Chain{V,G,T} <: TensorGraded{V,G}
     v::SVector{binomial(ndims(V),G),T}
 end
+
+@doc """
+    Chain{V,G,𝕂} <: TensorGraded{V,G} <: TensorAlgebra{V}
+
+Chain type with pseudoscalar `V::Manifold`, grade/rank `G::Int`, scalar field `𝕂::Type`.
+""" Chain
 
 export Chain
 getindex(m::Chain,i::Int) = m.v[i]
@@ -208,8 +224,13 @@ end
 ==(a::Chain{V,G,T},b::Chain{V,G,S}) where {V,G,T,S} = prod(a.v .== b.v)
 ==(a::Chain{V},b::Chain{V}) where V = prod(0 .==value(a)) && prod(0 .== value(b))
 
-## MultiVector{V,T}
+## MultiVector{V,𝕂}
 
+"""
+    MultiVector{V,𝕂} <: TensorMixed{V} <: TensorAlgebra{V}
+
+Chain type with pseudoscalar `V::Manifold` and scalar field `𝕂::Type`.
+"""
 struct MultiVector{V,T,E} <: TensorMixed{V}
     v::SArray{Tuple{E},T,1,E}
 end
@@ -322,6 +343,11 @@ end
 
 ## SparseChain{V,G}
 
+"""
+    SparseChain{V,G} <: TensorGraded{V,G} <: TensorAlgebra{V}
+
+Sparse chain type with pseudoscalar `V::Manifold` and grade/rank `G::Int`.
+"""
 struct SparseChain{V,G,T} <: TensorGraded{V,G}
     v::SparseVector{T,Int}
 end
@@ -389,6 +415,12 @@ end
     v::SVector{count_ones(G),TensorGraded{V}}
 end
 
+@doc """
+    MultiGrade{V,G} <: TensorMixed{V,G} <: TensorAlgebra{V}
+
[…]
