+++ src/Grassmann.jl
@@ -49,6 +49,12 @@ end
 
 export @basis, @basis_str, @dualbasis, @dualbasis_str, @mixedbasis, @mixedbasis_str
 
+"""
+    basis(V::Manifold,:V,"v","w","∂","ϵ")
+
+Generates `Basis` declaration of the `Algebra{V}` having `Manifold` specified by `V`.
+The first argument provides pseudoscalar specifications, the second argument is the variable name for the `Manifold`, and the third and fourth argument are variable prefixes of the `Basis` vector names (and covector basis names).
+"""
 function basis(V::Manifold,sig=vsn[1],vec=pre[1],cov=pre[2],duo=pre[3],dif=pre[4])
     N = ndims(V)
     if N > algebra_limit
@@ -69,6 +75,14 @@ function basis(V::Manifold,sig=vsn[1],vec=pre[1],cov=pre[2],duo=pre[3],dif=pre[4
     return Expr(:block,exp...,Expr(:tuple,esc(sig),esc.(sym)...))
 end
 
+"""
+    @basis
+
+Generates `Basis` elements of the `Algebra{V}` having `Manifold` specified by `V`.
+As a result of this macro, all of the `Basis{V,G}` elements generated by that `VectorBundle` become available in the local workspace with the specified naming.
+The first argument provides pseudoscalar specifications, the second argument is the variable name for the `Manifold`, and the third and fourth argument are variable prefixes of the `Basis` vector names (and covector basis names).
+Default for `@basis M` is `@basis M V v w ∂ ϵ`.
+"""
 macro basis(q,sig=vsn[1],vec=pre[1],cov=pre[2],duo=pre[3],dif=pre[4])
     basis(typeof(q)∈(Symbol,Expr) ? (@eval(__module__,$q)) : vectorspace(q),sig,string.([vec,cov,duo,dif])...)
 end
@@ -77,14 +91,30 @@ macro basis_str(str)
     basis(vectorspace(str))
 end
 
-macro dualbasis(q,sig=vsn[2],vec=pre[1],cov=pre[2],duo=pre[3],dif=pre[4])
-    basis((typeof(q)∈(Symbol,Expr) ? (@eval(__module__,$q)) : vectorspace(q))',sig,string.([vec,cov,duo,dif])...)
+"""
+    @dualbasis
+
+Generates `Basis` elements of the `Algebra{V'}` having `Manifold` specified by `V'`.
+As a result of this macro, all of the `Basis{V',G}` elements generated by that `VectorBundle` become available in the local workspace with the specified naming.
+The first argument provides pseudoscalar specifications, the second argument is the variable name for the dual `Manifold`, and the third and fourth argument are variable prefixes of the `Basis` covector names (and tensor field basis names).
+Default for `@dualbasis M` is `@dualbasis M VV w ϵ`.
+"""
+macro dualbasis(q,sig=vsn[2],cov=pre[2],dif=pre[4])
+    basis((typeof(q)∈(Symbol,Expr) ? (@eval(__module__,$q)) : vectorspace(q))',sig,string.([pre[1],cov,pre[3],dif])...)
 end
 
 macro dualbasis_str(str)
     basis(vectorspace(str)',vsn[2])
 end
 
+"""
+    @mixedbasis
+
+Generates `Basis` elements of the `Algebra{V⊕V'}` having `Manifold` specified by `V⊕V'`.
[…]
