+++ Project.toml
@@ -19,8 +19,11 @@ SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 [compat]
 julia = "1"
 Reduce = "1.2"
+Leibniz = "0.0.3"
 DirectSum = "0.4"
 AbstractTensors = "0.3"
+AbstractLattices = "0.1.2"
+ComputedFieldTypes = "0.1"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"

[…]
