### 2019-12-20 21f9191
revised references documentation
