### 2019-12-22 ffa3761
edited documentation styling
