### 2020-01-04 175872d
transitioned to Affero GPL license
