### 2020-01-04 chakravala/Grassmann.jl#29 @chakravala
This issue is being closed for now, although you are welcome to re-open or discuss further!

@snoeyink, I would be interested to give an extended presentation at your department, if desired.

More research papers will be written, but I can't currently provide a schedule for when they finish.
