### 2020-01-05 chakravala/Grassmann.jl#29 @waldyrious
> although you are welcome to re-open

FYI, only the repo owner or manually added collaborators can reopen issues that have been closed. Not even the issue author can do so. 🤷‍
