### 2020-01-14 6e97f6f
upgraded to AbstractTensors v0.4.2 and DirectSum v0.5
