+++ src/algebra.jl
@@ -1,32 +1,22 @@
 
-#   This file is part of Grassmann.jl. It is licensed under the GPL license
+#   This file is part of Grassmann.jl. It is licensed under the AGPL license
 #   Grassmann Copyright (C) 2019 Michael Reed
 
 import Base: +, -, *, ^, /, //, inv, <, >, <<, >>, >>>
 import AbstractLattices: ∧, ∨, dist
-import AbstractTensors: ⊗, ⊛, ⊙, ⊠, ⨼, ⨽, ⋆, ∗, rem, div, contraction
+import AbstractTensors: ⊗, ⊛, ⊙, ⊠, ⨼, ⨽, ⋆, ∗, rem, div, contraction, TAG
 import DirectSum: diffcheck, diffmode, tangent, hasinforigin, hasorigininf, symmetricsplit
 export tangent
 
-const Field = Real
-const ExprField = Union{Expr,Symbol}
-
-@pure g_one(b::Type{Basis{V}}) where V = getbasis(V,bits(b))
-@pure g_zero(V::Manifold) = 0*one(V)
-@pure g_one(V::Manifold) = Basis{V}()
-@pure g_one(::Type{T}) where T = one(T)
-@pure g_zero(::Type{T}) where T = zero(T)
-
 ## mutating operations
 
-const Sym = :DirectSum
+import DirectSum: g_one, g_zero, Field, ExprField
+const Sym = :AbstractTensors
 const SymField = Any
 
 @pure derive(n::N,b) where N<:Number = zero(typeof(n))
 derive(n,b,a,t) = t ? (a,derive(n,b)) : (derive(n,b),a)
 #derive(n,b,a::T,t) where T<:TensorAlgebra = t ? (a,derive(n,b)) : (derive(n,b),a)
-
-symmetricsplit(V::M,b::Basis) where M<:Manifold = symmetricsplit(V,bits(b))
 
 @inline function derive_mul(V,A,B,v,x::Bool)
     if !(diffvars(V)≠0 && mixedmode(V)<0)
@@ -132,20 +122,15 @@ function derive_post(V,::Val{A},::Val{B},a,b,*) where {A,B}
     return α*β
 end
 
-@pure isnull(::Expr) = false
-@pure isnull(::Symbol) = false
-isnull(n) = iszero(n)
+import DirectSum: loworder, isnull
 
-for M ∈ (:Signature,:DiagonalForm)
-    @eval @pure loworder(V::$M{N,M,S,D,O}) where {N,M,S,D,O} = O≠0 ? $M{N,M,S,D,O-1}() : V
-end
-@pure loworder(::SubManifold{N,M,S}) where {N,M,S} = SubManifold{N,loworder(M),S}()
-
-add_val(set,expr,val,OP) = Expr(OP∉(:-,:+) ? :.= : set,expr,OP∉(:-,:+) ? Expr(:.,OP,Expr(:tuple,expr,val)) : val)
+bcast(op,arg) = op ∈ (:(AbstractTensors.:∑),:(AbstractTensors.:-)) ? Expr(:.,op,arg) : Expr(:call,op,arg.args...)
 
 set_val(set,expr,val) = Expr(:(=),expr,set≠:(=) ? Expr(:call,:($Sym.:∑),expr,val) : val)
 
 pre_val(set,expr,val) = set≠:(=) ? :(isnull($expr) ? ($expr=Expr(:call,:($Sym.:∑),$val)) : push!($expr.args,$val)) : Expr(:(=),expr,val)
+
[…]
