+++ src/parity.jl
@@ -1,84 +1,25 @@
 
-#   This file is part of Grassmann.jl. It is licensed under the GPL license
+#   This file is part of Grassmann.jl. It is licensed under the AGPL license
 #   Grassmann Copyright (C) 2019 Michael Reed
+
+import DirectSum: parityreverse, parityinvolute, parityconj, parityright, parityleft, parityrighthodge, paritylefthodge, involute, grade_basis, odd, even
+import DirectSum: complementleft, complementright, ⋆, complementlefthodge, complementrighthodge, complement
+
+## complement
+
+export complementleft, complementright, ⋆, complementlefthodge, complementrighthodge
+
+## reverse
+
+import Base: reverse, conj, ~
+export involute
+
+## product parities
 
 @pure function parityjoin(N,S,a,b)
     B = DirectSum.digits_fast(b<<1,N)
     isodd(sum(DirectSum.digits_fast(a,N) .* cumsum(B))+count_ones((a & b) & S))
 end
-
-## adjoint parities
-
-@pure parityreverse(G) = isodd(Int((G-1)*G/2))
-@pure parityinvolute(G) = isodd(G)
-@pure parityconj(G) = parityreverse(G)⊻parityinvolute(G)
-
-## complement parity
-
-@pure parityrighthodge(V::Int,B,G,N=nothing) = isodd(V)⊻parityright(V,B,G,N)
-@pure paritylefthodge(V::Int,B,G,N) = (isodd(G) && iseven(N)) ⊻ parityrightgrade(V,B,G,N)
-@pure parityright(V::Int,B,G,N=nothing) = isodd(B+Int((G+1)*G/2))
-@pure parityleft(V::Int,B,G,N) = (isodd(G) && iseven(N)) ⊻ parityright(V,B,G,N)
-
-for side ∈ (:left,:right)
-    p = Symbol(:parity,side)
-    pg = Symbol(p,:hodge)
-    pn = Symbol(p,:null)
-    pnp = Symbol(pn,:pre)
-    @eval begin
-        @pure $p(V::Bits,B::Bits,N::Int) = $p(0,sum(indices(B,N)),count_ones(B),N)
-        @pure $pg(V::Bits,B::Bits,N::Int) = $pg(count_ones(V&B),sum(indices(B,N)),count_ones(B),N)
-        @inline $pn(V,B,v) = v
-        @inline function $pn(V::Signature,B,v)
-            hi,ho = hasinf(V),hasorigin(V)
-            if hi && ho && count_ones(B&UInt(3)) ==1
-                isodd(B) ? (2v) : (v/2)
-            else
-                v
-            end
-        end
-        @inline $pnp(V,B,v) = v
-        @inline function $pnp(V::Signature,B,v)
-            hi,ho = hasinf(V),hasorigin(V)
-            if hi && ho && count_ones(B&UInt(3)) ==1
-                isodd(B) ? Expr(:call,:*,2,v) : Expr(:call,:/,v,2)
-            else
-                v
-            end
-        end
-        @pure function $p(V::Signature,B,G=count_ones(B))
-            b = B&(UInt(1)<<(ndims(V)-diffvars(V))-1)
-            $p(0,sum(indices(b,ndims(V))),count_ones(b),ndims(V)-diffvars(V))
-        end
-        @pure function $pg(V::Signature,B,G=count_ones(B))
-            o = hasorigin(V) && hasinf(V) && (iszero(B&UInt(1))&(!iszero(B&UInt(2))))
-            b = B&(UInt(1)<<(ndims(V)-diffvars(V))-1)
-            $pg(count_ones(value(V)&b),sum(indices(b,ndims(V))),count_ones(b),ndims(V)-diffvars(V))⊻o
-        end
[…]
