### 2020-01-16 ab89650
finalized Grassmann v0.5 release with DirectSum v0.5.1
