+++ src/forms.jl
@@ -9,98 +9,124 @@
 #@pure supblade(N,S,B) = bladeindex(N,expandbits(N,S,B))
 #@pure supmulti(N,S,B) = basisindex(N,expandbits(N,S,B))
 
-function (W::Signature)(b::Chain{V,G,T}) where {V,G,T}
-    V==W && (return b)
-    !(V⊆W) && throw(error("cannot convert from $(V) to $(W)"))
-    WC,VC = mixedmode(W),mixedmode(V)
-    #if ((C1≠C2)&&(C1≥0)&&(C2≥0))
-    #    return V0
-    N,M = ndims(V),ndims(W)
-    out = zeros(choicevec(M,G,valuetype(b)))
-    ib = indexbasis(N,G)
-    for k ∈ 1:length(ib)
-        @inbounds if b[k] ≠ 0
-            @inbounds B = typeof(V)<:SubManifold ? expandbits(M,bits(V),ib[k]) : ib[k]
-            if WC<0 && VC≥0
-                @inbounds setblade!(out,b[k],mixed(V,B),Val{M}())
-            elseif WC≥0 && VC≥0
-                @inbounds setblade!(out,b[k],B,Val{M}())
+(W::Signature)(b::Chain{V,G,T}) where {V,G,T} = SubManifold(W)(b)
+function (W::SubManifold{Q,M})(b::Chain{V,G,T}) where {Q,M,V,G,T}
+    if isbasis(W)
+        if Q == V
+            if G == M == 1
+                x = bits(a)
+                X = isdyadic(V) ? x>>Int(ndims(V)/2) : x
+                Y = 0≠X ? X : x
+                @inbounds out = b.v[bladeindex(ndims(V),Y)]
+                return Simplex{V}((V[intlog(Y)+1] ? -(out) : out),SubManifold{V}())
+            elseif G == 1 && M == 2
+                (!isdyadic(V)) && throw(error("wrong basis"))
+                ib,(m1,m2) = indexbasis(N,1),DirectSum.eval_shift(W)
+                @inbounds ((V[m2] ? -(b.v[m2]) : b.v[m2])*getbasis(V,ib[m1]))
             else
-                throw(error("arbitrary Manifold intersection not yet implemented."))
+                throw(error("not yet possible"))
             end
+        else
+            interform(W,b)
         end
-    end
-    return Chain{W,G,T}(out)
-end
-function (W::SubManifold{V,M,S})(b::Chain{V,1,T}) where {M,V,S,T}
-    Chain{W,1,T}(b.v[indices(bits(W),ndims(V))])
-end
-function (W::SubManifold{V,M,S})(b::Chain{V,G,T}) where {M,V,S,T,G}
-    out,N = zeros(choicevec(M,G,valuetype(b))),ndims(V)
-    ib = indexbasis(N,G)
-    for k ∈ 1:length(ib)
-        @inbounds if b[k] ≠ 0
-            @inbounds if count_ones(ib[k]&S) == G
-                @inbounds setblade!(out,b[k],lowerbits(M,S,ib[k]),Val{M}())
+    elseif V==W
+        return b
+    elseif W⊆V
+        if G == 1
+            Chain{W,1,T}(b.v[indices(bits(W),ndims(V))])
+        else
+            out,N = zeros(choicevec(M,G,valuetype(b))),ndims(V)
+            ib,S = indexbasis(N,G),bits(W)
+            for k ∈ 1:length(ib)
+                @inbounds if b[k] ≠ 0
+                    @inbounds if count_ones(ib[k]&S) == G
[…]
