+++ Project.toml
@@ -19,7 +19,7 @@ SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 julia = "1"
 Reduce = "1.2"
 Leibniz = "0.0.3"
-DirectSum = "0.5"
+DirectSum = "0.5.1"
 AbstractTensors = "0.4.2"
 AbstractLattices = "0.1.2"
 ComputedFieldTypes = "0.1"

[…]
