### 2020-01-16 747edd1
updated README to Grassmann v0.5 docs
