+++ README.md
@@ -34,10 +34,10 @@ Please consider donating to show your thanks and appreciation to this project at
   * [TensorAlgebra design, Manifold code generation](#tensoralgebra-design-manifold-code-generation)
 	 * [Requirements](#requirements)
 	 * [Grassmann for enterprise](#grassmann-for-enterprise)
-	 * [Direct-sum yields VectorBundle parametric type polymorphism ⨁](#direct-sum-yields-vectorspace-parametric-type-polymorphism-)
+	 * [DirectSum yields VectorBundle parametric type polymorphism ⨁](#directsum-yields-vectorbundle-parametric-type-polymorphism-)
 	 * [Interoperability for TensorAlgebra{V}](#interoperability-for-tensoralgebrav)
   * [Grassmann elements and geometric algebra Λ(V)](#grassmann-elements-and-geometric-algebra-λv)
-	 * [Approaching ∞ dimensions with SparseAlgebra and ExtendedAlgebra](#approaching--dimensions-with-sparsealgebra-and-extendedalgebra)
+	 * [Approaching ∞ dimensions with SparseBasis and ExtendedBasis](#approaching--dimensions-with-sparsebasis-and-extendedbasis)
   * [References](#references)
 
 #### `TensorAlgebra` design, `Manifold` code generation
@@ -69,7 +69,7 @@ Mixed-symmetry algebra with *Leibniz.jl* and *Grassmann.jl*, having the geometri
 Most importantly, the Dirac-Clifford product yields generalized Hodge-Laplacian and the Betti numbers with Euler characteristic `χ`.
 
 The *Grassmann.jl* package and its accompanying support packages provide an extensible platform for high performance computing with geometric algebra at high dimensions.
-This enables the usage of many different types of `TensorAlgebra` along with various `VectorBundle` parameters and interoperability for a wide range of scientific and research applications.
+This enables the usage of many different types of `TensorAlgebra` along with various `Manifold` parameters and interoperability for a wide range of scientific and research applications.
 
 More information and tutorials are available at [https://grassmann.crucialflow.com/dev](https://grassmann.crucialflow.com/dev)
 
@@ -109,7 +109,7 @@ Sponsor this at [liberapay](https://liberapay.com/chakravala), [GitHub Sponsors]
 
 The maintainers of Grassmann and thousands of other packages are working with Tidelift to deliver commercial support and maintenance for the open source dependencies you use to build your applications. Save time, reduce risk, and improve code health, while paying the maintainers of the exact dependencies you use. [Learn more.](https://tidelift.com/subscription/pkg/julia-grassmann?utm_source=julia-grassmann&utm_medium=referral&utm_campaign=enterprise&utm_term=repo)
 
-## Direct-sum yields `VectorBundle` parametric type polymorphism ⨁
+## DirectSum yields `VectorBundle` parametric type polymorphism ⨁
 
 The *DirectSum.jl* package is a work in progress providing the necessary tools to work with an arbitrary `Manifold` specified by an encoding.
 Due to the parametric type system for the generating `VectorBundle`, the Julia compiler can fully preallocate and often cache values efficiently ahead of run-time.
@@ -121,10 +121,10 @@ The type `VectorBundle{n,ℙ,g,ν,μ}` uses *byte-encoded* data available at pre
 `g` is a bilinear form that specifies the metric of the space,
 and `μ` is an integer specifying the order of the tangent bundle (i.e. multiplicity limit of Leibniz-Taylor monomials). Lastly, `ν` is the number of tangent variables.
 
[…]
