+++ versions.js
@@ -1,3 +1,5 @@
 var DOC_VERSIONS = [
+  "stable",
+  "v0.5",
   "dev",
 ];
