### 2020-01-18 71cb113
fixed SubManifold docs issue
