### 2020-02-06 d019b8d
created ChainBundle and reorganized code
