### 2020-02-13 5869646
refined ChainBundle and meshing
