### 2020-02-16 8889dd4
improved ChainBundle for mesh boundary
