+++ v0.5.2/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"references/#References-1","page":"References","title":"References","text":"","category":"section"},{"location":"references/#","page":"References","title":"References","text":"(Image: YouTube) (Image: DropBox) (Image: Docs Stable) (Image: Docs Dev) (Image: Gitter) (Image: BiVector)","category":"page"},{"location":"references/#","page":"References","title":"References","text":"Michael Reed, Differential geometric algebra with Leibniz and Grassmann (2019)\nEmil Artin, Geometric Algebra (1957)\nJohn Browne, Grassmann Algebra, Volume 1: Foundations (2011)\nC. Doran, D. Hestenes, F. Sommen, and N. Van Acker, Lie groups as spin groups, J. Math Phys. (1993)\nDavid Hestenes, Universal Geometric Algebra, Pure and Applied (1988)\nDavid Hestenes, Renatus Ziegler, Projective Geometry with Clifford Algebra, Acta Appl. Math. (1991)\nDavid Hestenes, Tutorial on geometric calculus. Advances in Applied Clifford Algebra (2013)\nLachlan Gunn, Derek Abbott, James Chappell, Ashar Iqbal, Functions of multivector variables (2011)\nAaron D. Schutte, A nilpotent algebra approach to Lagrangian mechanics and constrained motion (2016)\nVladimir and Tijana Ivancevic, Undergraduate lecture notes in DeRahm-Hodge theory. arXiv (2011)\nPeter Woit, Clifford algebras and spin groups, Lecture Notes (2012)","category":"page"},{"location":"tutorials/dyadic-tensors/#Dyadic-tensor-product-1","page":"Dyadic tensor product ⊗","title":"Dyadic tensor product ⊗","text":"","category":"section"},{"location":"tutorials/dyadic-tensors/#","page":"Dyadic tensor product ⊗","title":"Dyadic tensor product ⊗","text":"Dyadic tensors are represented with the Grassmann’s exterior product algebra, generating a 2^2n-dimensional mother algebra with the direct sum of the n-dimensional vector space and its dual vector space. The product of the vector basis and covector basis elements form the n^2-dimensional bivector subspace of the full frac(2n)2(2n2)-dimensional bivector sub-algebra. The package Grassmann is working towards making the full extent of this number system available in Julia by using static compiled parametric type information to handle sparse sub-algebras, such as the dyadic (1,1)-tensor bivector algebra.","category":"page"},{"location":"tutorials/dyadic-tensors/#Constructing-linear-transformations-1","page":"Dyadic tensor product ⊗","title":"Constructing linear transformations","text":"","category":"section"},{"location":"tutorials/dyadic-tensors/#","page":"Dyadic tensor product ⊗","title":"Dyadic tensor product ⊗","text":"Note that Λ(3) gives the vector basis, and Λ(3)' gives the covector basis:","category":"page"},{"location":"tutorials/dyadic-tensors/#","page":"Dyadic tensor product ⊗","title":"Dyadic tensor product ⊗","text":"using Grassmann","category":"page"},{"location":"tutorials/dyadic-tensors/#","page":"Dyadic tensor product ⊗","title":"Dyadic tensor pr
[…]
