### 2020-02-21 fa7ffd3
improved performance of algebra element constructors
