+++ src/algebra.jl
@@ -9,6 +9,10 @@ export tangent
 
 ## mutating operations
 
+@pure tvec(N,G,t::Symbol=:Any) = :(SVector{$(binomial(N,G)),$t})
+@pure tvec(N,t::Symbol=:Any) = :(SVector{$(1<<N),$t})
+@pure tvec(N,μ::Bool) = tvec(N,μ ? :Any : :t)
+
 import DirectSum: g_one, g_zero, Field, ExprField
 const Sym = :AbstractTensors
 const SymField = Any
@@ -730,14 +734,14 @@ for (op,eop) ∈ ((:+,:(+=)),(:-,:(-=)))
                 out = zeros(mvec(N,A,t))
                 setblade!(out,value(a,t),bits(a),Val{N}())
                 setblade!(out,$op(value(b,t)),bits(b),Val{N}())
-                return Chain{V,A,t}(out)
+                return Chain{V,A}(out)
             else
                 #@warn("sparse MultiGrade{V} objects not properly handled yet")
                 #return MultiGrade{V}(a,b)
                 $(insert_expr((:N,:t,:out))...)
                 setmulti!(out,value(a,t),bits(a),Val{N}())
                 setmulti!(out,$op(value(b,t)),bits(b),Val{N}())
-                return MultiVector{V,t}(out)
+                return MultiVector{V}(out)
             end
         end
         function $op(a::SparseChain{V,G,T},b::SparseChain{V,G,S}) where {V,G,T,S}
@@ -828,10 +832,10 @@ function generate_sums(Field=Field,VEC=:mvec,MUL=:*,ADD=:+,SUB=:-,CONJ=:conj,PAR
         Base.adjoint(b::Simplex{V,G,B,T}) where {V,G,B,T<:$Field} = Simplex{dual(V),G,B',$TF}($CONJ(value(b)))
     end
     @eval begin
-        *(a::F,b::Chain{V,G,T}) where {F<:$Field,V,G,T<:$Field} = Chain{V,G,promote_type(T,F)}(broadcast($MUL,Ref(a),b.v))
-        *(a::Chain{V,G,T},b::F) where {F<:$Field,V,G,T<:$Field} = Chain{V,G,promote_type(T,F)}(broadcast($MUL,a.v,Ref(b)))
-        *(a::F,b::MultiVector{V,T}) where {F<:$Field,T,V} = MultiVector{V,promote_type(T,F)}(broadcast($Sym.∏,Ref(a),b.v))
-        *(a::MultiVector{V,T},b::F) where {F<:$Field,T,V} = MultiVector{V,promote_type(T,F)}(broadcast($Sym.∏,a.v,Ref(b)))
+        *(a::F,b::Chain{V,G,T}) where {F<:$Field,V,G,T<:$Field} = Chain{V,G}(broadcast($MUL,Ref(a),b.v))
+        *(a::Chain{V,G,T},b::F) where {F<:$Field,V,G,T<:$Field} = Chain{V,G}(broadcast($MUL,a.v,Ref(b)))
+        *(a::F,b::MultiVector{V,T}) where {F<:$Field,T,V} = MultiVector{V}(broadcast($Sym.∏,Ref(a),b.v))
+        *(a::MultiVector{V,T},b::F) where {F<:$Field,T,V} = MultiVector{V}(broadcast($Sym.∏,a.v,Ref(b)))
         *(a::F,b::MultiGrade{V,G}) where {F<:$EF,V,G} = MultiGrade{V,G}(broadcast($MUL,Ref(a),b.v))
         *(a::MultiGrade{V,G},b::F) where {F<:$EF,V,G} = MultiGrade{V,G}(broadcast($MUL,a.v,Ref(b)))
[…]
