### 2020-03-06 87d6e5d
improved performance of algebra operations
