+++ src/composite.jl
@@ -323,23 +323,25 @@ end
     return sum(terms[1:end-1])
 end=#
 
-import LinearAlgebra: det
-export detsimplex, mean, barycenter, means, barycenters, initmesh
+export detsimplex, initmesh
 
-det(m::Vector{Chain{V,G,T,X}} where {G,T,X}) where V = [∧(V[k]...) for k ∈ value.(m)]
-detsimplex(m::Vector{Chain{V,G,T,X}} where {G,T,X}) where V = det(m)/factorial(ndims(V)-1)
+detsimplex(m::Vector{Chain{V,G,T,X}} where {G,T,X}) where V = ∧(m)/factorial(ndims(V)-1)
 mean(m::Vector{Chain{V,G,T,X}} where {V,G,T,X}) = sum(m)/length(m)
-mean(m::SVector{N,Chain{V,G,T,X}} where {V,G,T,X}) where N = +(m...)/N
+mean(m::T) where T<:SVector = sum(m)/length(m)
+barycenter(m::SVector{N,Chain{V,G,T,X}} where {V,G,T,X}) where N = (s=sum(m);s/s[1])
 barycenter(m::Vector{Chain{V,G,T,X}} where {V,G,T,X}) = (s=sum(m);s/s[1])
-for op ∈ (:det,:detsimplex,:mean,:barycenter)
+curl(m::SVector{N,Chain{V,G,T,X}} where {N,G,T,X}) where V = curl(Chain{V,1}(m))
+curl(m::T) where T<:TensorAlgebra = Manifold(m)(∇)×m
+for op ∈ (:∧,:detsimplex)
     @eval @pure $op(m::ChainBundle) = ChainBundle($op(value(m)))
 end
-for op ∈ (:mean,:barycenter)
+for op ∈ (:mean,:barycenter,:curl)
     ops = Symbol(op,:s)
     @eval begin
+        export $op, $ops
         @pure $ops(m::ChainBundle{p}) where p = $ops(m,p)
         @pure $ops(m::ChainBundle,::SubManifold{p}) where p = $ops(m,p)
-        @pure $ops(m::ChainBundle,p::ChainBundle) = [$op(p[k]) for k ∈ value.(value(m))]
+        @pure $ops(m::ChainBundle,p) = [$op(p[k]) for k ∈ value.(value(m))]
     end
 end
 
