### 2020-03-07 113b30b
improved performance of composite methods
