### 2020-03-11 a9ce73c
fixed zero element dispatch issue
