### 2020-03-16 d35448c
fixed #41 issue with dimension mismatch
