### 2020-03-16 6f26633
add example for computing barycentric coords in G3
