### 2020-03-16 5420667
remove extra backpacks
