### 2020-03-21 04dcb9e
improved and extended various methods
