### 2020-03-21 b8219e1
changed documentation build script
