### 2020-03-25 e242eba
extended SparseArray multiplication syntax
