+++ .travis.yml
@@ -5,7 +5,7 @@ os:
   - osx
 julia:
   - 1.0
-  - 1.3
+  - 1.4
   - nightly
 jobs:
   allow_failures:
