+++ appveyor.yml
@@ -1,7 +1,7 @@
 environment:
   matrix:
   - julia_version: 1
-  - julia_version: 1.3
+  - julia_version: 1.4
   - julia_version: nightly
 
 platform:
