### 2020-03-26 1c859b4
extended AbstractPlotting.arrows interoperability
