### 2020-04-04 017336a
extended StaticArrays._diff for Chain values
