+++ .github/FUNDING.yml
@@ -3,7 +3,7 @@
 github: [chakravala]
 patreon: dreamscatter
 liberapay: chakravala
-open_collective: # Replace with a single Open Collective username
+open_collective: grassmannjl
 ko_fi: # Replace with a single Ko-fi username
 tidelift: "julia/Grassmann"
 otechie: # Replace with a single Otechie username
