+++ src/Grassmann.jl
@@ -314,6 +314,8 @@ function __init__()
         AbstractPlotting.arrows!(p::ChainBundle{V},v;args...) where V = AbstractPlotting.arrows!(value(p),v;args...)
         AbstractPlotting.arrows(p::Vector{Chain{V,G,T,X}} where {G,T,X},v;args...) where V = AbstractPlotting.arrows(GeometryTypes.Point.(V(2:ndims(V)...).(p)),GeometryTypes.Point.(value(v));args...)
         AbstractPlotting.arrows!(p::Vector{Chain{V,G,T,X}} where {G,T,X},v;args...) where V = AbstractPlotting.arrows!(GeometryTypes.Point.(V(2:ndims(V)...).(p)),GeometryTypes.Point.(value(v));args...)
+        AbstractPlotting.scatter(p::ChainBundle,x,;args...) = AbstractPlotting.scatter(submesh(p)[:,1],x;args...)
+        AbstractPlotting.scatter!(p::ChainBundle,x,;args...) = AbstractPlotting.scatter!(submesh(p)[:,1],x;args...)
         AbstractPlotting.mesh(t::ChainBundle;args...) = AbstractPlotting.mesh(points(t),t;args...)
         function AbstractPlotting.mesh(p::ChainBundle,t::ChainBundle;args...)
             if ndims(p) == 2
@@ -321,16 +323,6 @@ function __init__()
             else
                 AbstractPlotting.mesh(submesh(p),array(t);args...)
             end
-        end
-        const submesh_cache = (Array{T,2} where T)[]
-        export submesh
-        submesh(m) = [m[i][j] for i∈1:length(m),j∈2:ndims(Manifold(m))]
-        function submesh(m::ChainBundle{V,G,T,B} where {V,G,T}) where B
-            for k ∈ length(submesh_cache):B
-                push!(submesh_cache,Array{Any,2}(undef,0,0))
-            end
-            isempty(submesh_cache[B]) && (submesh_cache[B] = submesh(value(m)))
-            return submesh_cache[B]
         end
     end
     #@require Makie="ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a" nothing
@@ -350,25 +342,46 @@ function __init__()
                 return matlab_cache[B]
             end
         end
+        initmesh(g,args...) = initmeshall(g,args...)[1:3]
+        initmeshall(g::Matrix{Int},args...) = initmeshall(Matrix{Float64}(g),args...)
         function initmeshall(g,args...)
-            p,e,t = MATLAB.mxcall(:initmesh,3,Matrix{Float64}(g),args...)
-            s = size(p,1)+1; V = SubManifold(ℝ^s)
-            P = ChainBundle([Chain{V,1,Float64}(vcat(1,p[:,k])) for k ∈ 1:size(p,2)])
-            E = ChainBundle([Chain{P(2:s...),1,Int}(Int.(e[1:s-1,k])) for k ∈ 1:size(e,2)])
-            T = ChainBundle([Chain{P,1,Int}(Int.(t[1:s,k])) for k ∈ 1:size(t,2)])
-            return (P,E,T,t)
-        end
-        function initmesh(g,args...)
-            P,E,T = initmeshall(g,args...)
-            #matlab(p,bundle(P)); matlab(e,bundle(E)); matlab(t,bundle(T))
-            return (P,E,T)
+            P,E,T = MATLAB.mxcall(:initmesh,3,g,args...)
[…]
