### 2020-04-16 41d6870
Added financial contributors to the README
