### 2020-04-20 bd26975
resolved crashing bug with CGA exp
