### 2020-04-21 1bc6d67
revised AbstractPlotting.lines
