### 2020-04-21 4093188
updated AbstractPlotting.lines documentation
