### 2020-04-22 a03e7ab
Correct typo in README
