### 2020-04-22 01b2003
removed Reduce from testing
