+++ Project.toml
@@ -9,14 +9,12 @@ ComputedFieldTypes = "459fdd68-db75-56b8-8c15-d717a790f88e"
 DirectSum = "22fd7b30-a8c0-5bf2-aabe-97783860d07c"
 Leibniz = "edad4870-8a01-11e9-2d75-8f02e448fc59"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-Reduce = "93e0c654-6965-5f22-aba9-9c1ae6b3c259"
 Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 
 [compat]
 julia = "1"
-Reduce = "1.2"
 Leibniz = "0.0.5"
 DirectSum = "0.5.3"
 AbstractTensors = "0.4.3"
