+++ .travis.yml
@@ -3,6 +3,7 @@ language: julia
 os:
   - linux
   - osx
+  - windows
 julia:
   - 1.0
   - 1.4
