### 2020-04-22 c36924b
fixed symbolic display issue
