### 2020-04-23 d56151f
Implement `zero` for multivectors accepting a type as argument
