+++ src/multivectors.jl
@@ -54,6 +54,7 @@ setindex!(m::Chain{V,G,T} where {V,G},k::T,i::Int) where T = (m.v[i] = k)
 Base.firstindex(m::Chain) = 1
 @pure Base.lastindex(m::Chain{V,G}) where {V,G} = binomial(ndims(V),G)
 @pure Base.length(m::Chain{V,G}) where {V,G} = binomial(ndims(V),G)
+Base.zero(::Type{<:Chain{V,G,T}}) where {V,G,T} = Chain{V,G}(zeros(svec(ndims(V),G,T)))
 Base.zero(::Chain{V,G,T}) where {V,G,T} = Chain{V,G}(zeros(svec(ndims(V),G,T)))
 
 function (m::Chain{V,G,T})(i::Integer) where {V,G,T}
