### 2020-04-23 faac7d9
added some mappings for SymPy and SymEngine functions
