### 2020-04-23 f10c276
ensure symbolic functions work on Chain, MultiVector and Simplex objects
