### 2020-04-23 ed641ae
added numeric conversion functions for SymPy and SymEngine
