+++ src/Grassmann.jl
@@ -263,6 +263,9 @@ function __init__()
         SymPy.apart(x::T) = map(v -> symbolic(v) ? SymPy.apart(v) : v, x)
         SymPy.cancel(x::T) = map(v -> symbolic(v) ? SymPy.cancel(v) : v, x)
         # SymPy.expand_trig(x::T) = map(v -> symbolic(v) ? SymPy.expand_trig(v) : v, x)
+        SymPy.N(x::T, args...) = map(v -> symbolic(v) ? SymPy.N(v, args...) : v, x)
+        # SymPy.evalf(X::T, args...) = map(v -> symbolic(v) ? SymPy.evalf(v, args...) : v, x)
+        SymPy.convert(X::T, args...) = map(v -> symbolic(v) ? SymPy.convert(v, args...) : v, x)
         SymPy.subs(x::T, args...) = map(v -> symbolic(v) ? SymPy.subs(v, args...) : v, x)
         SymPy.collect(x::T, args...) = map(v -> symbolic(v) ? SymPy.collect(v, args...) : v, x)
 
@@ -273,6 +276,9 @@ function __init__()
         symbolic = v -> typeof(v) == SymEngine.Basic
         SymEngine.expand(x::T) = map(v -> symbolic(v) ? SymEngine.expand(v) : v, x)
         SymEngine.subs(x::T, args...) = map(v -> symbolic(v) ? SymEngine.subs(v, args...) : v, x)
+        SymEngine.N(x::T) = map(v -> symbolic(v) ? SymEngine.N(v) : v, x)
+        SymEngine.evalf(X::T, args...) = map(v -> symbolic(v) ? SymEngine.evalf(v, args...) : v, x)
+        SymEngine.convert(X::T, args...) = map(v -> symbolic(v) ? SymEngine.convert(v, args...) : v, x)
     end
     @require AbstractAlgebra="c3fe647b-3220-5bb0-a1ea-a7954cac585d" generate_algebra(:AbstractAlgebra,:SetElem)
     @require GaloisFields="8d0d7f98-d412-5cd4-8397-071c807280aa" generate_algebra(:GaloisFields,:AbstractGaloisField)
