+++ .github/pull_request_template.md
@@ -1,0 +1,23 @@
+**Type of pull request**
+Grammar fixes don't require any further explanation.
+Documentation extensions or examples require justification.
+Bug fixes require inclusion of or reference to bug report.
+New features require justification and desired goals.
+
+**Describe the feature**
+A clear and concise description of what the feature is.
+Describe the changes proposed in the pull request.
+Reference any related issue in repository.
+
+**To Reproduce**
+Steps to reproduce the desired behavior:
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**Environment:** (if necessary)
+ - OS: [e.g. GNU/Linux/OSX/Windows]
+ - Version [e.g. Julia v1.4, Grassmann v0.5.7]
+
+**Additional context**
+Add any other context about the problem here.
