### 2020-04-23 21c3a48
avoid Union types and type piracy in symbolic function definitions
