### 2020-04-24 3f38acf
Mappings to common SymPy and SymEngine functions (#54)
* added some mappings for SymPy and SymEngine functions

* ensure symbolic functions work on Chain, MultiVector and Simplex objects

* added numeric conversion functions for SymPy and SymEngine

* avoid Union types and type piracy in symbolic function definitions

* remove extra line
