### 2020-04-24 d701a6e
fixed ⊘ docstring #55
