### 2020-04-24 05f8b7e
fixed #56 function name for >>> docstring
