### 2020-04-25 8bd31e4
expand description of ⊖ definition #57
