### 2020-04-27 90db55d
extended ∧ and LinearAlgebra.det
