### 2020-04-27 ace55c1
extended ⋅ and : for nested dyadics
