### 2020-04-28 c2aa2e6
significantly improved *,∧,∨ product algebras
