+++ src/algebra.jl
@@ -178,7 +178,7 @@ function generate_mutators(M,F,set_val,SUB,MUL)
                 @inline function $(Symbol(:join,s))(V::W,m::$M,a::UInt,b::UInt,v::S) where W<:Manifold{N} where {N,T<:$F,S<:$F,M}
                     if v ≠ 0 && !diffcheck(V,a,b)
                         A,B,Q,Z = symmetricmask(V,a,b)
-                        val = (typeof(V)<:Signature || count_ones(A&B)==0) ? (parity(A,B,V) ? $SUB(v) : v) : $MUL(parityinner(A,B,V),v)
+                        val = (typeof(V)<:Signature || count_ones(A&B)==0) ? (parity(V,A,B) ? $SUB(v) : v) : $MUL(parityinner(V,A,B),v)
                         if diffvars(V)≠0
                             !iszero(Z) && (T≠Any ? (return true) : (val *= getbasis(loworder(V),Z)))
                             count_ones(Q)+order(val)>diffmode(V) && (return false)
@@ -190,7 +190,7 @@ function generate_mutators(M,F,set_val,SUB,MUL)
                 @inline function $(Symbol(:join,spre))(V::W,m::$M,a::UInt,b::UInt,v::S) where W<:Manifold{N} where {N,T<:$F,S<:$F,M}
                     if v ≠ 0 && !diffcheck(V,a,b)
                         A,B,Q,Z = symmetricmask(V,a,b)
-                        val = (typeof(V)<:Signature || count_ones(A&B)==0) ? (parity(A,B,V) ? :($$SUB($v)) : v) : :($$MUL($(parityinner(A,B,V)),$v))
+                        val = (typeof(V)<:Signature || count_ones(A&B)==0) ? (parity(V,A,B) ? :($$SUB($v)) : v) : :($$MUL($(parityinner(V,A,B)),$v))
                         if diffvars(V)≠0
                             !iszero(Z) && (val = Expr(:call,:*,val,getbasis(loworder(V),Z)))
                             val = :(h=$val;iszero(h)||$(count_ones(Q))+order(h)>$(diffmode(V)) ? 0 : h)
@@ -202,8 +202,8 @@ function generate_mutators(M,F,set_val,SUB,MUL)
                 @inline function $(Symbol(:geom,s))(V::W,m::$M,a::UInt,b::UInt,v::S) where W<:Manifold{N} where {N,T<:$F,S<:$F,M}
                     if v ≠ 0 && !diffcheck(V,a,b)
                         A,B,Q,Z = symmetricmask(V,a,b)
-                        pcc,bas,cc = (hasinf(V) && hasorigin(V)) ? conformal(A,B,V) : (false,A⊻B,false)
-                        val = (typeof(V)<:Signature || count_ones(A&B)==0) ? (parity(A,B,V)⊻pcc ? $SUB(v) : v) : $MUL(parityinner(A,B,V),pcc ? $SUB(v) : v)
+                        pcc,bas,cc = (hasinf(V) && hasorigin(V)) ? conformal(V,A,B) : (false,A⊻B,false)
+                        val = (typeof(V)<:Signature || count_ones(A&B)==0) ? (parity(V,A,B)⊻pcc ? $SUB(v) : v) : $MUL(parityinner(V,A,B),pcc ? $SUB(v) : v)
                         if istangent(V)
[…]
