+++ src/products.jl
@@ -55,15 +55,154 @@ function generate_products(Field=Field,VEC=:mvec,MUL=:*,ADD=:+,SUB=:-,CONJ=:conj
             v = derive_mul(V,bits(ba),bits(bb),a.v,b.v,$MUL)
             Simplex(v,mul(ba,bb,v))
         end
-        ∧(a::$Field,b::$Field) = $MUL(a,b)
-        ∧(a::F,b::B) where B<:TensorTerm{V,G} where {F<:$EF,V,G} = Simplex{V,G}(a,b)
-        ∧(a::A,b::F) where A<:TensorTerm{V,G} where {F<:$EF,V,G} = Simplex{V,G}(b,a)
-        #=∧(a::$Field,b::Chain{V,G,T}) where {V,G,T<:$Field} = Chain{V,G,T}(a.*b.v)
-        ∧(a::Chain{V,G,T},b::$Field) where {V,G,T<:$Field} = Chain{V,G,T}(a.v.*b)
-        ∧(a::$Field,b::MultiVector{V,T}) where {V,T<:$Field} = MultiVector{V,T}(a.*b.v)
-        ∧(a::MultiVector{V,T},b::$Field) where {V,T<:$Field} = MultiVector{V,T}(a.v.*b)
-        ∧(a::$Field,b::MultiGrade{V,G}) where V = MultiGrade{V,G}(a.*b.v)
-        ∧(a::MultiGrade{V,G},b::$Field) where V = MultiGrade{V,G}(a.v.*b)=#
+        @generated function *(a::Chain{V,G,T},b::SubManifold{V,L}) where {V,G,L,T<:$Field}
+            if G == 0
+                return :(a[1]*b)
+            elseif L == ndims(V) && !istangent(V) && !DirectSum.hasconformal(V)
+                return :(⋆(~a))
+            elseif G == ndims(V) && !istangent(V)
+                return :(a[1]*⋆(~involute(b)))
+            elseif binomial(ndims(V),G)<(1<<cache_limit)
+                $(insert_expr((:N,:t,:out,:ib,:μ),:svec)...)
+                for i ∈ 1:binomial(N,G)
+                    @inbounds geomaddmulti!_pre(V,out,ib[i],bits(b),derive_pre(V,ib[i],bits(b),:(a[$i]),true))
+                end
+                return insert_t(:(MultiVector{V}($(Expr(:call,tvec(N,μ),out...)))))
+            else return quote
+                $(insert_expr((:N,:t,:out,:ib,:μ),$(QuoteNode(VEC)))...)
+                for i ∈ 1:binomial(N,G)
+                    if @inbounds geomaddmulti!(V,out,ib[i],bits(b),derive_mul(V,ib[i],bits(b),a[i],true))&μ
+                        $(insert_expr((:out,);mv=:out)...)
+                        @inbounds geomaddmulti!(V,out,ib[i],bits(b),derive_mul(V,ib[i],bits(b),a[i],true))
+                    end
+                end
+                return MultiVector{V}(out)
+            end end
+        end
+        @generated function *(a::SubManifold{V,L},b::Chain{V,G,T}) where {V,G,L,T<:$Field}
+            if G == 0
+                return :(a*b[1])
+            elseif G == ndims(V) && !istangent(V)
+                return :(⋆(~a)*b[1])
+            #=elseif L == ndims(V) && !istangent(V)
+                return :(⋆(~involute(b)))=#
+            elseif binomial(ndims(V),G)<(1<<cache_limit)
+                $(insert_expr((:N,:t,:out,:ib,:μ),:svec)...)
+                for i ∈ 1:binomial(N,G)
[…]

[…]
