### 2020-04-28 7bbc28e
transitioned ∧ ∨ clifford conj and rand samplers #59
