+++ Project.toml
@@ -12,11 +12,12 @@ LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
+Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 
 [compat]
 julia = "1"
 Leibniz = "0.0.5"
-DirectSum = "0.5.3"
+DirectSum = "0.5.7"
 AbstractTensors = "0.4.3"
 ComputedFieldTypes = "0.1"
 StaticArrays = "0"
