### 2020-04-29 7bd38c3
improved complement parity for *,∧,∨,⋅ products
