### 2020-04-30 b324d1c
improved CGA contraction type stability, fixed #63
