### 2020-05-01 d8c7101
revised nested type dispatch and CGA contraction
