+++ src/products.jl
@@ -352,11 +352,7 @@ function generate_products(Field=Field,VEC=:mvec,MUL=:*,ADD=:+,SUB=:-,CONJ=:conj
                     end
                 end
                 return if μ
-                    if !istangent(V)
-                        insert_t(:(MultiVector{$V}($(Expr(:call,tvec(N,:t),out...)))))
-                    else
-                        insert_t(:(MultiVector{$V}($(Expr(:call,tvec(N),out...)))))
-                    end
+                    insert_t(:(MultiVector{$V}($(Expr(:call,istangent(V) ? tvec(N) : tvec(N,:t),out...)))))
                 else
                     insert_t(:(value_diff(Chain{$V,G-L}($(Expr(:call,tvec(N,G-L,:t),out...))))))
                 end
