### 2020-05-01 333a2c1
improved value dispatch and fixed bug
