### 2020-05-12 f21379d
Make up/downarrow operators type-stable
