### 2020-05-13 2300b87
revised JuliaCon paper terminology
