+++ paper/paper.tex
@@ -82,11 +82,11 @@ The \textit{DirectSum.jl} package is a work in progress providing the necessary 
 Although intended for use with the \textit{Grassmann.jl} package, \verb`DirectSum` can be used independently.
 
 \begin{definition}[Vector bundle of submanifolds]
-	Let $M = T^\mu V\in\text{Vect}_{\mb K}$ be a \verb`VectorBundle<:Manifold` of rank $n$,
+	Let $M = T^\mu V\in\text{Vect}_{\mb K}$ be a \verb`TensorBundle<:Manifold` of rank $n$,
 	\begin{align*}
 		T^\mu V &= (n,\mb P,g,\nu,\mu), & \mb P &\subseteq\gen{v_\infty,v_\emptyset}, & g &:V\times V\ra\mb K
 	\end{align*}
-	The type \verb+VectorBundle{n,+$\mb P$\verb+,g,+$\nu,\mu$\verb+}+ uses \textit{byte-encoded} data available at pre-compilation, where
+	The type \verb+TensorBundle{n,+$\mb P$\verb+,g,+$\nu,\mu$\verb+}+ uses \textit{byte-encoded} data available at pre-compilation, where
 	$\mb P$ specifies the basis for up and down projection,
 	$g$ is a bilinear form that specifies the metric of the space,
 	and $\mu$ is an integer specifying the order of the tangent bundle (i.e. multiplicity limit of Leibniz-Taylor monomials). Lastly, $\nu$ is the number of tangent variables.
@@ -113,8 +113,8 @@ Given \verb`M(s::Int...)` one can encode \verb+SubManifold{+$|s|,M,s$\verb+}+ wi
 $$T^eV \subset T^\mu W \iff \exists Z\in\text{Vect}_{\mb K}(T^e(V\oplus Z) = T^{e\leq \mu}W,\,V\perp Z),$$
 such that computing unions of submanifolds is done by inspecting the parameter $s\in V\subseteq W$ and $s\notin Z$. Operations on \verb`Manifold` types is automatically handled at compile time.
 
-The metric signature of the \verb+Basis{V,1}+ elements of a vector space \verb+V+ can be specified with the \verb+V"..."+ constructor by using \verb-+- and \verb+-+ to specify whether the \verb+Basis{V,1}+ element of the corresponding index squares to \verb`+1` or \verb`-1`.
-For example, \verb`S"+++"` constructs a positive definite 3-dimensional \verb`VectorBundle`.
+The metric signature of the \verb+SubManifold{V,1}+ elements of a vector space \verb+V+ can be specified with the \verb+V"..."+ constructor by using \verb-+- and \verb+-+ to specify whether the element of the corresponding index squares to \verb`+1` or \verb`-1`.
+For example, \verb`S"+++"` constructs a positive definite 3-dimensional \verb`TensorBundle`.
 It is also possible to specify an arbitrary \verb`DiagonalForm` having numerical values for the basis with degenracy \verb`D"1,1,1,0"`, although the $\pm$ format has a more compact representation.
 Further development will result in more metric types.
 
@@ -127,18 +127,18 @@ The \verb`tangent(V,`$\mu$\verb`,`$\nu$\verb`)` map can be used to specify $\mu$
 \section{Tensor basis equivalence classes}% of Leibniz and Grassmann}
 
 The \verb`AbstractTensors` package is intended for universal interoperability of the abstract \verb`TensorAlgebra` type system.
-All \verb`TensorAlgebra{V}` subtypes have type parameter \verb`V`, used to store a \verb`VectorBundle` value obtained from \textit{DirectSum.jl}.
[…]
