### 2020-05-27 4b45a9f
fixed #51 fixed #68 fixed #69
