+++ paper/paper.tex
@@ -210,7 +210,7 @@ Both symmetry classes can be characterized by the same geometric product, which 
 	Let $v_\pm^2 = \pm1$ be a basis with $v_\infty = v_++v_-$ and $v_\emptyset = (v_--v_+)/2$
 An embedding space $\mb R^{p+1,q+1}$ carrying the action from the group $O(p+1,q+1)$ then has
 $v_\infty^2 =0$, $v_\emptyset^2 =0$,
-$v_\infty \cdot v_\emptyset = 1$,  and $v_{\infty\emptyset}^2 = 1$ with
+$v_\infty \cdot v_\emptyset = -1$,  and $v_{\infty\emptyset}^2 = 1$ with
 Minkowski plane $v_{\infty\emptyset}$ having the Hestenes-Dirac-Clifford product properties
 \begin{align*}
 	v_{\infty\emptyset}\ominus v_\infty &= -v_\infty, &  v_{\infty\emptyset}\ominus v_\emptyset &= v_\emptyset, \\
