### 2020-06-15 6efc000
improved support for nested dyadic tensors
