+++ src/multivectors.jl
@@ -30,7 +30,12 @@ end
 Chain type with pseudoscalar `V::Manifold`, grade/rank `G::Int`, scalar field `𝕂::Type`.
 """
 Chain{V,G}(val::S) where {V,G,S<:AbstractVector{𝕂}} where 𝕂 = Chain{V,G,𝕂}(val)
-Chain{V,G}(args::𝕂...) where {V,G,𝕂} = Chain{V,G}(SVector{binomial(ndims(V),G),𝕂}(args...))
+#Chain{V,G}(args::𝕂...) where {V,G,𝕂} = Chain{V,G}(SVector{binomial(ndims(V),G)}(args...))
+@generated function Chain{V,G}(args::𝕂...) where {V,G,𝕂}
+    bg = binomial(ndims(V),G)
+    ref = SVector{bg}([:(args[$i]) for i ∈ 1:bg])
+    :(Chain{V,G}($(Expr(:call,:(SVector{$bg,𝕂}),ref...))))
+end
 function Chain(val::𝕂,v::SubManifold{V,G}) where {V,G,𝕂}
     N = ndims(V)
     Chain{V,G}(setblade!(zeros(mvec(N,G,𝕂)),val,bits(v),Val{N}()))
@@ -66,7 +71,46 @@ end
 Chain{V,1}(m::SMatrix{N,N}) where {V,N} = Chain{V,1}(Chain{V,1}.(getindex.(Ref(m),:,SVector{N}(1:N))))
 Chain{V,1,Chain{W,1}}(m::SMatrix{M,N}) where {V,W,M,N} = Chain{V,1}(Chain{W,1}.(getindex.(Ref(m),:,SVector{N}(1:N))))
 
-Base.inv(m::Chain{V,1,<:Chain{W,1}}) where {V,W} = Chain{V,1,Chain{W,1}}(inv(SMatrix(m)))
+transpose_row(t::SVector{N,<:Chain{V}},i) where {N,V} = Chain{V,1}(getindex.(t,i))
+transpose_row(t::Chain{V,1,<:Chain},i) where V = transpose_row(value(t),i)
+@generated _transpose(t::SVector{N,<:Chain{V,1}}) where {N,V} = :(Chain{V,1}(transpose_row.(Ref(t),$(SVector{ndims(V)}(indices(V))))))
+Base.transpose(t::Chain{V,1,<:Chain{V,1}}) where V = _transpose(value(t))
+
+@pure function Cramer(N::Int)
+    x,y = SVector{N}([Symbol(:x,i) for i ∈ 1:N]),SVector{N}([Symbol(:y,i) for i ∈ 1:N])
+    xy = [:(($(x[1+i]),$(y[1+i])) = ($(x[i])∧t[$(1+i)],t[end-$i]∧$(y[i]))) for i ∈ 1:N-1]
+    return x,y,xy
+end
+
+@generated function Base.:\(t::Chain{V,1,<:Chain{V,1}},v::Chain{V,1}) where V
+    N = ndims(V)-1 # pase this into the REPL for faster eval
+    x,y,xy = Grassmann.Cramer(N)
+    mid = [:($(x[i])∧v∧$(y[end-i])) for i ∈ 1:N-1]
+    out = Expr(:call,:SVector,:(v∧$(y[end])),mid...,:($(x[end])∧v))
+    return Expr(:block,:((x1,y1)=(t[1],t[end])),xy...,
+        :(Chain{V,1}(getindex.($(Expr(:call,:./,out,:(t[1]∧$(y[end])))),1))))
+end
+
+@generated function Base.in(v::Chain{V,1},t::Chain{V,1,<:Chain{V,1}}) where V
+    N = ndims(V)-1
+    x,y,xy = Grassmann.Cramer(N)
[…]
