### 2020-06-16 0c94616
Fixed typo in `README.md`
