### 2020-06-18 215f26a
revised Base.:\ method dispatch
