+++ src/composite.jl
@@ -325,6 +325,44 @@ end
     return sum(terms[1:end-1])
 end=#
 
+function Cramer(N::Int)
+    x,y = SVector{N}([Symbol(:x,i) for i ∈ 1:N]),SVector{N}([Symbol(:y,i) for i ∈ 1:N])
+    xy = [:(($(x[1+i]),$(y[1+i])) = ($(x[i])∧t[$(1+i)],t[end-$i]∧$(y[i]))) for i ∈ 1:N-1]
+    return x,y,xy
+end
+
+@generated function Base.:\(t::SVector{N,<:Chain{V,1}} where N,v::Chain{V,1}) where V
+    N = ndims(V)-1 # paste this into the REPL for faster eval
+    x,y,xy = Grassmann.Cramer(N)
+    mid = [:($(x[i])∧v∧$(y[end-i])) for i ∈ 1:N-1]
+    out = Expr(:call,:SVector,:(v∧$(y[end])),mid...,:($(x[end])∧v))
+    return Expr(:block,:((x1,y1)=(t[1],t[end])),xy...,
+        :(Chain{V,1}(getindex.($(Expr(:call,:./,out,:(t[1]∧$(y[end])))),1))))
+end
+
+Base.:\(t::Chain{V,1,<:Chain{V,1}},v::Chain{V,1}) where V = value(t)\v
+
+@generated function Base.in(v::Chain{V,1},t::Chain{V,1,<:Chain{V,1}}) where V
+    N = ndims(V)-1
+    x,y,xy = Grassmann.Cramer(N)
+    out = Expr(:call,:SVector,:(s==signbit((v∧$(y[end]))[1])),[:(s==signbit(($(x[i])∧v∧$(y[end-i]))[1])) for i ∈ 1:N-1]...,:(s==signbit(($(x[end])∧v)[1])))
+    return Expr(:block,:((x1,y1)=(t[1],t[end])),xy...,:(s=signbit((t[1]∧$(y[end]))[1])),
+        Expr(:call,:prod,out))
+end
+
+@generated function Base.inv(t::Chain{V,1,<:Chain{V,1}}) where V
+    N = ndims(V)-1
+    x,y,xy = Grassmann.Cramer(N)
+    out = if iseven(N)
+        Expr(:call,:SVector,y[end],[:($(y[end-i])∧$(x[i])) for i ∈ 1:N-1]...,x[end])
+    else
+        Expr(:call,:SVector,:(-$(y[end])),[:($(isodd(i) ? :+ : :-)($(y[end-i])∧$(x[i]))) for i ∈ 1:N-1]...,x[end])
+    end
+    return Expr(:block,:((x1,y1)=(t[1],t[end])),xy...,:(_transpose(.⋆($(Expr(:call,:./,out,:((t[1]∧$(y[end]))[1])))))))
+end
+
+INV(m::Chain{V,1,<:Chain{V,1}}) where V = Chain{V,1,Chain{V,1}}(inv(SMatrix(m)))
+
 export detsimplex, initmesh, refinemesh, refinemesh!, select, submesh
 
 detsimplex(m::Vector{<:Chain{V}}) where V = ∧(m)/factorial(ndims(V)-1)
