### 2020-06-20 d1b800b
implemented compatibility for MiniQhull delaunay
