### 2020-06-24 059d993
generalized Base.in method for embedded affine simplices
