+++ src/Grassmann.jl
@@ -333,6 +333,15 @@ function __init__()
         Base.convert(::Type{GeometryBasics.Point},t::T) where T<:TensorAlgebra = GeometryBasics.Point(value(vector(t)))
         Base.convert(::Type{GeometryBasics.Point},t::Chain{V,G,T}) where {V,G,T} = G == 1 ? GeometryBasics.Point(value(vector(t))) : GeometryBasics.Point(zeros(T,ndims(V))...)
         GeometryBasics.Point(t::T) where T<:TensorAlgebra = convert(GeometryBasics.Point,t)
+        function initmesh(m::GeometryBasics.Mesh)
+            c,f = coordinates(m),faces(m)
+            s = size(eltype(c))[1]+1; V = SubManifold(ℝ^s)
+            n = size(eltype(f))[1]
+            p = ChainBundle([Chain{V,1}(SVector{s,Float64}(1.0,k...)) for k ∈ c])
+            M = p(s-n+1:s...)
+            t = ChainBundle([Chain{M,1}(SVector{n,Int}(k)) for k ∈ f])
+            return (p,t)
+        end
         @pure ptype(::GeometryBasics.Point{N,T} where N) where T = T
         export vectorfield, chainfield
         vectorfield(t,V=Manifold(t),W=V) = p->GeometryBasics.Point(V(vector(↓(↑((V∪Manifold(t))(Chain{W,1,ptype(p)}(p.data)))⊘t))))
@@ -395,7 +404,7 @@ function __init__()
         function initmeshall(g,args...)
             P,E,T = MATLAB.mxcall(:initmesh,3,g,args...)
             s = size(P,1)+1; V = SubManifold(ℝ^s)
-            p = ChainBundle([Chain{V,1,Float64}(vcat(1,P[:,k])) for k ∈ 1:size(P,2)])
+            p = ChainBundle([Chain{V,1,Float64}(vcat(1.0,P[:,k])) for k ∈ 1:size(P,2)])
             e = ChainBundle([Chain{p(2:s...),1,Int}(Int.(E[1:s-1,k])) for k ∈ 1:size(E,2)])
             t = ChainBundle([Chain{p,1,Int}(Int.(T[1:s,k])) for k ∈ 1:size(T,2)])
             return (p,e,t,T,E,P)
