+++ src/composite.jl
@@ -325,9 +325,10 @@ end
     return sum(terms[1:end-1])
 end=#
 
-function Cramer(N::Int)
+function Cramer(N::Int,j=0)
+    t = j ≠ 0 ? :T : :t
     x,y = SVector{N}([Symbol(:x,i) for i ∈ 1:N]),SVector{N}([Symbol(:y,i) for i ∈ 1:N])
-    xy = [:(($(x[1+i]),$(y[1+i])) = ($(x[i])∧t[$(1+i)],t[end-$i]∧$(y[i]))) for i ∈ 1:N-1]
+    xy = [:(($(x[1+i]),$(y[1+i])) = ($(x[i])∧$t[$(1+i-j)],$t[end-$i]∧$(y[i]))) for i ∈ 1:N-1-j]
     return x,y,xy
 end
 
@@ -340,12 +341,19 @@ end
         :(Chain{V,1}(getindex.($(Expr(:call,:./,out,:(t[1]∧$(y[end])))),1))))
 end
 
-@generated function Base.in(v::Chain{V,1},t::SVector{N,<:Chain{V,1}} where N) where V
-    N = ndims(V)-1
-    x,y,xy = Grassmann.Cramer(N)
-    out = Expr(:call,:SVector,:(s==signbit((v∧$(y[end]))[1])),[:(s==signbit(($(x[i])∧v∧$(y[end-i]))[1])) for i ∈ 1:N-1]...,:(s==signbit(($(x[end])∧v)[1])))
-    return Expr(:block,:((x1,y1)=(t[1],t[end])),xy...,:(s=signbit((t[1]∧$(y[end]))[1])),
-        Expr(:call,:prod,out))
+@generated function Base.in(v::Chain{V,1},t::SVector{N,<:Chain{V,1}}) where {V,N}
+    if N == ndims(V)
+        x,y,xy = Grassmann.Cramer(N-1)
+        mid = [:(s==signbit(($(x[i])∧v∧$(y[end-i]))[1])) for i ∈ 1:N-2]
+        out = SVector(:(s==signbit((v∧$(y[end]))[1])),mid...,:(s==signbit(($(x[end])∧v)[1])))
+        return Expr(:block,:((x1,y1)=(t[1],t[end])),xy...,:(s=signbit((t[1]∧$(y[end]))[1])),ands(out))
+    else
+        x,y,xy = Grassmann.Cramer(N-1,1)
+        mid = [:(signscalar(($(x[i])∧(v-x1)∧$(y[end-i]))/d)) for i ∈ 1:N-2]
+        out = SVector(:(signscalar((v∧∧(vectors(t,v)))/d)),mid...,:(signscalar(($(x[end])∧(v-x1))/d)))
+        return Expr(:block,:(T=vectors(t)),:((x1,y1)=(t[1],T[end])),xy...,
+            :($(x[end])=$(x[end-1])∧T[end-1];d=$(x[end])∧T[end]),ands(out))
+    end
 end
 
 @generated function Base.inv(t::SVector{N,<:Chain{V,1}} where N) where V
@@ -360,9 +368,49 @@ end
 end
 
 Base.:\(t::Chain{V,1,<:Chain{V,1}},v::Chain{V,1}) where V = value(t)\v
-Base.in(v::Chain{V,1},t::Chain{V,1,<:Chain{V,1}}) where V = v ∈ value(t)
+Base.in(v::Chain{V,1},t::Chain{W,1,<:Chain{V,1}}) where {V,W} = v ∈ value(t)
 Base.inv(t::Chain{V,1,<:Chain{V,1}}) where V = inv(value(t))
[…]
