### 2020-06-28 f53719c
genralized simplicial mesh support for embedded manifolds
