+++ Project.toml
@@ -18,7 +18,7 @@ Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 julia = "1"
 Leibniz = "0.0.5"
 DirectSum = "0.5.8"
-AbstractTensors = "0.4.8"
+AbstractTensors = "0.4.9"
 ComputedFieldTypes = "0.1"
 StaticArrays = "0"
 Requires = "1"
