### 2020-06-28 5833b0c
fixed embedded ChainBundle getindex bug
