### 2020-06-28 ca4193b
created generalized SubManifold indexing
