### 2020-07-01 6d52e45
improved performance for mesh & plotting
