+++ src/Grassmann.jl
@@ -121,6 +121,7 @@ end
 
 @generated function ↑(ω::T) where T<:TensorAlgebra
     V = Manifold(ω)
+    T<:SubManifold && !isbasis(ω) && (return DirectSum.supermanifold(V))
     !(hasinf(V)||hasorigin(V)) && (return :ω)
     G = Λ(V)
     return if hasinf(V) && hasorigin(V)
@@ -133,6 +134,7 @@ end
         end
     end
 end
+↑(ω::ChainBundle) = ω
 function ↑(ω,b)
     ω2 = ω^2
     iω2 = inv(ω2+1)
@@ -145,18 +147,21 @@ function ↑(ω,p,m)
 end
 
 @generated function ↓(ω::T) where T<:TensorAlgebra
-    V = Manifold(ω)
-    !(hasinf(V)||hasorigin(V)) && (return :ω)
+    V,M = Manifold(ω),T<:SubManifold && !isbasis(ω)
+    !(hasinf(V)||hasorigin(V)) && (return M ? V(2:ndims(V)) : :ω)
     G = Λ(V)
     return if hasinf(V) && hasorigin(V)
+        M && (return ω(3:ndims(V)))
         :(inv(one(valuetype(ω))*$G.v∞∅)*($G.v∞∅∧ω)/(-ω⋅$G.v∞))
     else
+        M && (return V(2:ndims(V)))
         quote
             b = hasinf($V) ? $G.v∞ : $G.v∅
             ((ω∧b)*b)/(1-b⋅ω)
         end
     end
 end
+↓(ω::ChainBundle) = ω(list(2,ndims(ω)))
 ↓(ω,b) = ((b∧ω)*b)/(1-ω⋅b)
 ↓(ω,∞,∅) = (m=∞∧∅;inv(m)*(m∧ω)/(-ω⋅∞))
 
@@ -217,6 +222,31 @@ function skeleton(x::MultiVector{V},v::Val{T}=Val{true}()) where {V,T}
         end
     end
     return g
+end
+
+function pointset(e)
+    ndims(e) == 1 && (return column(e))
+    out = Int[]
+    for i ∈ value(e)
+        for k ∈ value(i)
+            k ∉ out && push!(out,k)
+        end
+    end
+    return out
+end
+
+column(t,i=1) = getindex.(value(t),i)
+columns(t,i=1,j=ndims(t)) = column.(Ref(value(t)),list(i,j))
+
+function edges(t,cols=columns(t))
+    ndims(t) == 2 && (return t)
+    np,N = length(points(t)),ndims(Manifold(t))
+    A = spzeros(np,np),points(t)(list(N-1,N)...)
+    for c ∈ combo(N,2)
+        A += sparse(cols[c[1]],cols[c[2]],1,np,np)
+    end
+    f = findall(x->x>0,triu(A+transpose(A)))
+    [Chain{M,1}(SVector{2,Int}(f[n].I)) for n ∈ 1:length(f)]
 end
 
 generate_products()
@@ -338,12 +368,13 @@ function __init__()
         Base.convert(::Type{GeometryBasics.Point},t::T) where T<:TensorAlgebra = GeometryBasics.Point(value(vector(t)))
         Base.convert(::Type{GeometryBasics.Point},t::Chain{V,G,T}) where {V,G,T} = G == 1 ? GeometryBasics.Point(value(vector(t))) : GeometryBasics.Point(zeros(T,ndims(V))...)
         GeometryBasics.Point(t::T) where T<:TensorAlgebra = convert(GeometryBasics.Point,t)
+        pointpair(p,V) = Pair(GeometryBasics.Point.(V.(value(p)))...)
         function initmesh(m::GeometryBasics.Mesh)
[…]
