+++ Project.toml
@@ -1,7 +1,7 @@
 name = "Grassmann"
 uuid = "4df31cd9-4c27-5bea-88d0-e6a7146666d8"
 authors = ["Michael Reed"]
-version = "0.5.15"
+version = "0.5.16"
 
 [deps]
 AbstractTensors = "a8e43f4a-99b7-5565-8bf1-0165161caaea"
@@ -17,7 +17,7 @@ Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 [compat]
 julia = "1"
 Leibniz = "0.0.5"
-DirectSum = "0.5.8"
+DirectSum = "0.5.11"
 AbstractTensors = "0.4.9"
 ComputedFieldTypes = "0.1"
 StaticArrays = "0"
