### 2020-07-02 56a7b12
fixed bug with columns & edges
