### 2020-07-03 60f4444
created nullary operators for ∧, ∨
