+++ src/algebra.jl
@@ -356,6 +356,7 @@ Exterior product as defined by the anti-symmetric quotient Λ≡⊗/~
 export ∧, ∨, ⊗
 
 @inline ∧(t) = t
+@inline ∧() = 1
 
 @pure function ∧(a::SubManifold{V},b::SubManifold{V}) where V
     ba,bb = bits(a),bits(b)
@@ -416,7 +417,18 @@ Regressive product as defined by the DeMorgan's law: ∨(ω...) = ⋆⁻¹(∧(�
 @inline ∨(a::X,b::Y) where {X<:TensorAlgebra,Y<:TensorAlgebra} = interop(∨,a,b)
 @inline ∨(a::TensorAlgebra{V},b::UniformScaling{T}) where {V,T<:Field} = a∨V(b)
 @inline ∨(a::UniformScaling{T},b::TensorAlgebra{V}) where {V,T<:Field} = V(a)∨b
+@generated ∨(t::T) where T<:SVector = Expr(:call,:∨,[:(t[$k]) for k ∈ 1:length(t)]...)
+@generated ∨(t::T) where T<:SizedVector = Expr(:call,:∨,[:(t[$k]) for k ∈ 1:length(t)]...)
+∨(::SVector{0,<:Chain{V}}) where V = SubManifold(V) # ∨() = I
+∨(::SizedVector{0,<:Chain{V}}) where V = SubManifold(V)
+∨(t::Chain{V,1,<:Chain} where V) = ∧(value(t))
 ∨(a::X,b::Y,c::Z...) where {X<:TensorAlgebra,Y<:TensorAlgebra,Z<:TensorAlgebra} = ∨(a∨b,c...)
+
+export ∧, ∨, ⊗
+
+@inline ∨(t) = t
+@inline ∨() = I
+
 
 for X ∈ TAG, Y ∈ TAG
     @eval Base.:&(a::$X{V},b::$Y{V}) where V = a∨b
