+++ Project.toml
@@ -17,8 +17,8 @@ Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 [compat]
 julia = "1"
 Leibniz = "0.0.5"
-DirectSum = "0.5.11"
-AbstractTensors = "0.4.9"
+DirectSum = "0.6"
+AbstractTensors = "0.5"
 ComputedFieldTypes = "0.1"
 StaticArrays = "0"
 Requires = "1"
