+++ src/Grassmann.jl
@@ -85,20 +85,21 @@ end
 d(ω::T) where T<:TensorAlgebra = Manifold(ω)(∇)∧ω
 δ(ω::T) where T<:TensorAlgebra = -∂(ω)
 
-function boundary_rank(t::T,d=gdims(t)) where T<:TensorAlgebra
+function boundary_rank(t,d=gdims(t))
     out = gdims(∂(t))
     out[1] = 0
-    for k ∈ 2:ndims(t)
+    for k ∈ 2:length(out)-1
         @inbounds out[k] = min(out[k],d[k+1])
     end
     return SVector(out)
 end
 
-function boundary_null(t::T) where T<:TensorAlgebra
+function boundary_null(t)
     d = gdims(t)
     r = boundary_rank(t,d)
-    out = zeros(MVector{ndims(t)+1,Int})
-    for k ∈ 1:ndims(V)
+    l = length(d)
+    out = zeros(MVector{l,Int})
+    for k ∈ 1:l-1
         @inbounds out[k] = d[k+1] - r[k]
     end
     return SVector(out)
@@ -112,8 +113,9 @@ Compute the Betti numbers.
 function betti(t::T) where T<:TensorAlgebra
     d = gdims(t)
     r = boundary_rank(t,d)
-    out = zeros(MVector{ndims(t),Int})
-    for k ∈ 1:ndims(V)
+    l = length(d)-1
+    out = zeros(MVector{l,Int})
+    for k ∈ 1:l
         @inbounds out[k] = d[k+1] - r[k] - r[k+1]
     end
     return SVector(out)
@@ -226,10 +228,10 @@ end
 
 # mesh
 
-export pointset, edges, incidence, adjacency, degrees, column, columns
+export pointset, edges, facets, incidence, adjacency, degrees, column, columns
 
 function pointset(e)
-    ndims(e) == 1 && (return column(e))
+    ndims(Manifold(e)) == 1 && (return column(e))
     out = Int[]
     for i ∈ value(e)
         for k ∈ value(i)
@@ -243,30 +245,104 @@ column(t,i=1) = getindex.(value(t),i)
 columns(t,i=1,j=ndims(Manifold(t))) = column.(Ref(value(t)),list(i,j))
 
 degrees(t,A=incidence(t)) = sum(A,dims=2)[:]
-function incidence(t)
+function incidence(t,cols=columns(t))
     np,nt = length(points(t)),length(t)
     A = spzeros(Int,np,nt)
     for i ∈ Grassmann.list(1,ndims(Manifold(t)))
-        A += sparse(column(t,i),1:nt,1,np,nt)
+        A += sparse(cols[i],1:nt,1,np,nt)
     end
     return A
 end # node-element incidence, A[i,j]=1 -> i∈t[j]
 
-function adjacency(t,cols=columns(t))
+antiadjacency(t::ChainBundle,cols=columns(t)) = (A = sparse(t,cols); A-transpose(A))
+adjacency(t,cols=columns(t)) = (A = sparse(t,cols); A+transpose(A))
+function SparseArrays.sparse(t,cols=columns(t))
     np,N = length(points(t)),ndims(Manifold(t))
     A = spzeros(Int,np,np)
     for c ∈ combo(N,2)
         A += sparse(cols[c[1]],cols[c[2]],1,np,np)
     end
-    return A+transpose(A)
+    return A
 end
 
-function edges(t,cols=columns(t))
+edges(t,cols::SVector) = edges(t,adjacency(t,cols))
+function edges(t,adj=adjacency(t))
     ndims(t) == 2 && (return t)
     N = ndims(Manifold(t)); M = points(t)(list(N-1,N)...)
[…]
