### 2020-07-19 b0aa17e
transferred degrees to Adapode and revised API
