+++ src/Grassmann.jl
@@ -228,6 +228,23 @@ end
 
 # mesh
 
+function initpoints(P)
+    V = SubManifold(ℝ^(size(P,1)+1))
+    p = [Chain{V,1,Float64}(vcat(1.0,P[:,k])) for k ∈ 1:size(P,2)]
+end
+
+function initpointsdata(P,E)
+    p = ChainBundle(initpoints(P)); el = list(1,size(P,1))
+    e = [Chain{↓(p),1,Int}(Int.(E[el,k])) for k ∈ 1:size(E,2)]
+    return p,e
+end
+
+function initmeshdata(P,E,T)
+    p,e = initpointsdata(P,E); tl = list(1,size(P,1)+1)
+    t = [Chain{p,1,Int}(Int.(T[tl,k])) for k ∈ 1:size(T,2)]
+    return p,ChainBundle(e),ChainBundle(t)
+end
+
 export pointset, edges, facets, adjacency, column, columns
 
 column(t,i=1) = getindex.(value(t),i)
@@ -290,7 +307,7 @@ function faces(t::Vector{<:Chain{V}},::Val{N}) where {V,N}
     N == 0 && (return Chain{W,1}(list(2,1)))
     out = Chain{W,1,Int,N}[]
     for i ∈ value(t)
-        for w ∈ Chain{W,1}.(combinations(sort(value(i)),N))
+        for w ∈ Chain{W,1}.(DirectSum.combinations(sort(value(i)),N))
             w ∉ out && push!(out,w)
         end
     end
@@ -306,7 +323,7 @@ function faces(t::Vector{<:Chain{V}},h,::Val{N},g=identity) where {V,N}
     for i ∈ 1:length(t)
         vec[:] = value(t[i])
         par = indexparity!(vec)
-        w = Chain{W,1}.(combinations(par[2],N))
+        w = Chain{W,1}.(DirectSum.combinations(par[2],N))
         for k ∈ 1:binomial(ndims(V),N)
             j = findfirst(isequal(w[k]),out)
             v = h[i]*(par[1] ? -val[k] : val[k])
@@ -527,11 +544,78 @@ function __init__()
     end
     #@require Makie="ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a" nothing
     @require MiniQhull="978d7f02-9e05-4691-894f-ae31a51d76ca" begin
-        function MiniQhull.delaunay(p::ChainBundle,n=1:length(p))
+        MiniQhull.delaunay(p::Vector{<:Chain},n=1:length(p)) = MiniQhull.delaunay(ChainBundle(p),n)
+        function MiniQhull.delaunay(p::ChainBundle,n=1:length(p)); l = list(1,ndims(p))
             T = MiniQhull.delaunay(Matrix(submesh(length(n)==length(p) ? p : p[n])'))
-            l = list(1,ndims(p)); [Chain{p,1,Int}(Int.(T[l,k])) for k ∈ 1:size(T,2)]
+            [Chain{p,1,Int}(getindex.(Ref(n),Int.(T[l,k]))) for k ∈ 1:size(T,2)]
         end
-        initmesh(p::ChainBundle) = (t=delaunay(p); (p,ChainBundle(∂(t)),ChainBundle(t)))
+        #initmesh(p::ChainBundle) = (t=delaunay(p); (p,ChainBundle(∂(t)),ChainBundle(t)))
+    end
+    @require Triangulate = "f7e6ffb2-c36d-4f8f-a77e-16e897189344" begin
+        const triangle_cache = (Array{T,2} where T)[]
+        function triangle(p::Array{T,2} where T,B)
+            for k ∈ length(triangle_cache):B
+                push!(triangle_cache,Array{Any,2}(undef,0,0))
+            end
[…]
