+++ README.md
@@ -103,6 +103,7 @@ The package is compatible via [Requires.jl](https://github.com/MikeInnes/Require
 [Compose.jl](https://github.com/GiovineItalia/Compose.jl),
 [AbstractPlotting.jl](https://github.com/JuliaPlots/AbstractPlotting.jl),
 [GeometryBasics.jl](https://github.com/JuliaGeometry/GeometryBasics.jl),
+[Delaunay.jl](https://github.com/eschnett/Delaunay.jl),
 [MATLAB.jl](https://github.com/JuliaInterop/MATLAB.jl),
 [MiniQhull.jl](https://github.com/gridap/MiniQhull.jl),
 [Triangulate.jl](https://github.com/JuliaGeometry/Triangulate.jl),
