+++ README.md
@@ -101,14 +101,15 @@ The package is compatible via [Requires.jl](https://github.com/MikeInnes/Require
 [GaloisFields.jl](https://github.com/tkluck/GaloisFields.jl),
 [LightGraphs.jl](https://github.com/JuliaGraphs/LightGraphs.jl),
 [Compose.jl](https://github.com/GiovineItalia/Compose.jl),
+[Makie.jl](https://github.com/JuliaPlots/Makie.jl),
 [AbstractPlotting.jl](https://github.com/JuliaPlots/AbstractPlotting.jl),
 [GeometryBasics.jl](https://github.com/JuliaGeometry/GeometryBasics.jl),
-[Delaunay.jl](https://github.com/eschnett/Delaunay.jl),
-[MATLAB.jl](https://github.com/JuliaInterop/MATLAB.jl),
 [MiniQhull.jl](https://github.com/gridap/MiniQhull.jl),
+[QHull.jl](https://github.com/JuliaPolyhedra/QHull.jl),
+[Delaunay.jl](https://github.com/eschnett/Delaunay.jl),
 [Triangulate.jl](https://github.com/JuliaGeometry/Triangulate.jl),
 [TetGen.jl](https://github.com/JuliaGeometry/TetGen.jl),
-[Makie.jl](https://github.com/JuliaPlots/Makie.jl).
+[MATLAB.jl](https://github.com/JuliaInterop/MATLAB.jl).
 
 ## Grassmann for enterprise
 
