### 2020-07-27 #18 @vchuravy
I withdrew your paper since you did not reply to https://github.com/JuliaCon/proceedings-review/issues/18#issuecomment-550494959 without the paper in the right format it is impossible for us to proceed.
