### 2020-08-06 #18 @chakravala
Well, I prepared everything, except that I needed to modify the `cls` file a bit.

What am I supposed to do to address the issue? I submitted everything, the paper compiles fine.

I would like to have this published.
