### 2020-08-17 #18 @matbesancon
to come back to the issue of publishing it, we can't let you modify the template, because we only provide it to authors in order to be able to build locally and not rely on pushing changes to a github repo. This means modifying the template creates an out-of-sync with the actual template we used, and we need to keep papers standardized.

The way forward is to take the TeX and CLS template from https://github.com/JuliaCon/JuliaConSubmission.jl for the paper. Again, we don't have the time or manpower to restructure the content.

Given the comments of the reviewers and the current content as shared on Dropbox, I would argue that the paper is not adapted to JuliaCon as a publishing venue and out of scope. It can be of interest, but needs to make less assumption on knowledge of differential geometry.

A large part of the paper is setting definitions, definitions should be referred to in other papers, bringing a whole lot of new symbols and notation does not help the reader understand what the package is doing. Same thing for theorems.
