### 2020-08-17 #18 @chakravala
Thank you for the response.

> A large part of the paper is setting definitions, definitions should be referred to in other papers, bringing a whole lot of new symbols and notation does not help the reader understand what the package is doing. Same thing for theorems.

It is completely necessary to go over all the definitions in this paper, since some of the definitions are entirely new and have never been published before. So, I literally can't reference other papers for some of these definitions.

I am willing to add 2 more pages to the paper to expand more on what the package is doing. It is now 1 year later, and I have much more insight into how this package can be used and what it can be used for.
