### 2020-08-17 #18 @snoeyink
While I realize that one can't allow authors to edit latex templates that are used uniformly for all articles, I'll be sad if the resolution is that this does not get published at all.
I've used tex since 1985 and latex since 1990, so I'd be happy to work with you to make something that would compile with the standard template. (I was just contacted as a technical reviewer, and have no pull within JuliaCon for whether something can still be accepted after it was dropped)
Last semester was completely crazy, and this semester promises to be only a little better, but I'd still like to connect with you at some point and learn from your experiences with geometric algebra in Julia.
