### 2020-08-17 #18 @matbesancon
> It is completely necessary to go over all the definitions in this paper, since some of the definitions are entirely new and have never been published before.

Ok I see, in that case, as highlighted before, the novel contributions in differential algebraic geometry are probably more adapted to another paper submitted to a mathematics venue (journal or conference), with this proceeding highlighting more the package.
