### 2020-08-17 #18 @matbesancon
> I am willing to add 2 more pages to the paper to expand more on what the package is doing. It is now 1 year later, and I have much more insight into how this package can be used and what it can be used for.

I don't think the solution is to just add more pages, but restructure the current content and focus on what you propose here: possible applications of Grassmann.jl and how to use it with some examples
