### 2020-08-17 #18 @eschnett
Since this paper has been public for more than a year, I would suggest posting a preprint to arxiv.org to make it more official. Many physics, math, and computer science papers are effectively published there before they are submitted to a journal (all my papers are).

As for a description for a Julia package, another venue is JOSS <https://joss.theoj.org>. The publications there are essentially an extended documentation for a software package.
