### 2020-08-17 #18 @matbesancon
I agree for the preprint aspect.
Thing to note for JOSS, there will be the same scope issue since JOSS is not meant to publish new contribution in the field of the software library, but review the library itself
