### 2020-08-17 #18 @chakravala
> probably more adapted to another paper submitted to a mathematics venue (journal or conference), with this proceeding highlighting more the package.

Indeed, I am currently writing a paper to a mathematics journal (Advances in Applied Clifford Algebras), but at the original time of writing of the JuliaCon paper that wasn't the case yet, since I only submitted that abstract recently.

I am also currently submitting a paper to American Institute of Aeronautics and Astronautics on applications of this package.

> I would suggest posting a preprint to arxiv.org to make it more official.

That's not possible, since I do not have permission to post on arxiv.

> JOSS is not meant to publish new contribution in the field of the software library, but review the library itself

It may be possible for me to modify the JuliaCon paper to meet this criteria, since I am also now submitting 2 other papers, which I could reference in the JuliaCon paper. At the original time of writing, that wasn't the case yet.

However, if it requires a complete rewrite like this, then I may not be able to fix the paper immediately in a short period of time.
