### 2020-08-17 #18 @matbesancon
As general advice, if you are writing and submitting multiple papers at the same time, having a preprint available (the version of the paper you submitted) is generally a good way to reference some of them in the others. If not arxiv, then a personal website is an okay choice, but an established preprint server will be preferred by most venues.

If you are re-structuring some parts in another article and citing it, then yes this will likely need some rewrite here as to not repeat what you already cite
