### 2020-08-17 #18 @eschnett
> That's not possible, since I do not have permission to post on arxiv.
Can you explain? arxiv.org is quite general in accepting authors and submissions.
