### 2020-08-17 #18 @chakravala
Have you tried submitting a paper on arxiv before? I tried it, and found that I can't just make an account and submit things.
