### 2020-08-17 #18 @eschnett
Yes, I have; see <http://arxiv.org/a/schnetter_e_1>. I find that arxiv.org works quite well. If there's a specific thing holding you up, then you should email their help desk.
