### 2020-08-18 #18 @chakravala
Honestly, i dont even care about posting my articles on arxiv or some journal, I'm just doing it for fun and couldn't care less how official anyone thinks it is. I have written hundreds of emails criticising univerisities to the point that I will never be able to get a job. I'll think about rewriting this paper, but actually I dont care what anyone out there thinks about my work anymore, you're all planning to leech off my work for free while I worked my ass off for years learning this, so I am not going to bend over backwards to please you or your publication, but I will casually work on rewriting some more if i feel like it, but honestly I dont know why I should bother with making my hard work available on someone else's preprint server, especially if those academic institutions sanctioned me and want me to suffer the fate of not having economic opportunities.
