### 2020-08-20 #18 @chakravala
Apologies if I offended anyone with my last (now deleted) post, it's not directed at anyone specifically, it was just me being abstractly frustrated.
